Network.AWS.AutoScaling.AttachInstances

Description

Attaches one or more EC2 instances to the specified Auto Scaling group.

When you attach instances, Auto Scaling increases the desired capacity of the group by the number of instances being attached. If the number of instances being attached plus the desired capacity of the group exceeds the maximum size of the group, the operation fails.

If there is a Classic Load Balancer attached to your Auto Scaling group, the instances are also registered with the load balancer. If there are target groups attached to your Auto Scaling group, the instances are also registered with the target groups.

For more information, see Attach EC2 Instances to Your Auto Scaling Group in the Auto Scaling User Guide .

Creating a Request

attachInstances

Arguments

:: Text
aiAutoScalingGroupName

-> AttachInstances

Creates a value of AttachInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiInstanceIds - One or more instance IDs.

aiAutoScalingGroupName - The name of the group.

data AttachInstances

Contains the parameters for AttachInstances.

See: attachInstances smart constructor.

Request Lenses

aiInstanceIds :: Lens' AttachInstances [Text]

One or more instance IDs.

aiAutoScalingGroupName :: Lens' AttachInstances Text

The name of the group.

Destructuring the Response

attachInstancesResponse :: AttachInstancesResponse

Creates a value of AttachInstancesResponse with the minimum fields required to make a request.

data AttachInstancesResponse

See: attachInstancesResponse smart constructor.