Network.AWS.AutoScaling.AttachLoadBalancerTargetGroups

Description

Attaches one or more target groups to the specified Auto Scaling group.

To describe the target groups for an Auto Scaling group, use DescribeLoadBalancerTargetGroups . To detach the target group from the Auto Scaling group, use DetachLoadBalancerTargetGroups .

For more information, see Attach a Load Balancer to Your Auto Scaling Group in the Auto Scaling User Guide .

Creating a Request

attachLoadBalancerTargetGroups

Arguments

:: Text
albtgAutoScalingGroupName

-> AttachLoadBalancerTargetGroups

Creates a value of AttachLoadBalancerTargetGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albtgAutoScalingGroupName - The name of the Auto Scaling group.

albtgTargetGroupARNs - The Amazon Resource Names (ARN) of the target groups.

data AttachLoadBalancerTargetGroups

Contains the parameters for AttachLoadBalancerTargetGroups.

See: attachLoadBalancerTargetGroups smart constructor.

Request Lenses

albtgAutoScalingGroupName :: Lens' AttachLoadBalancerTargetGroups Text

The name of the Auto Scaling group.

albtgTargetGroupARNs :: Lens' AttachLoadBalancerTargetGroups [Text]

The Amazon Resource Names (ARN) of the target groups.

Destructuring the Response

attachLoadBalancerTargetGroupsResponse

Arguments

:: Int
albtgrsResponseStatus

-> AttachLoadBalancerTargetGroupsResponse

Creates a value of AttachLoadBalancerTargetGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albtgrsResponseStatus - -- | The response status code.

data AttachLoadBalancerTargetGroupsResponse

See: attachLoadBalancerTargetGroupsResponse smart constructor.

Response Lenses

albtgrsResponseStatus :: Lens' AttachLoadBalancerTargetGroupsResponse Int

- | The response status code.