Network.AWS.AutoScaling.AttachLoadBalancers

Description

Attaches one or more Classic Load Balancers to the specified Auto Scaling group.

To attach an Application Load Balancer instead, see AttachLoadBalancerTargetGroups .

To describe the load balancers for an Auto Scaling group, use DescribeLoadBalancers . To detach the load balancer from the Auto Scaling group, use DetachLoadBalancers .

For more information, see Attach a Load Balancer to Your Auto Scaling Group in the Auto Scaling User Guide .

Creating a Request

attachLoadBalancers

Arguments

:: Text
albAutoScalingGroupName

-> AttachLoadBalancers

Creates a value of AttachLoadBalancers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albAutoScalingGroupName - The name of the group.

albLoadBalancerNames - One or more load balancer names.

data AttachLoadBalancers

Contains the parameters for AttachLoadBalancers.

See: attachLoadBalancers smart constructor.

Request Lenses

albAutoScalingGroupName :: Lens' AttachLoadBalancers Text

The name of the group.

albLoadBalancerNames :: Lens' AttachLoadBalancers [Text]

One or more load balancer names.

Destructuring the Response

attachLoadBalancersResponse

Arguments

:: Int
albrsResponseStatus

-> AttachLoadBalancersResponse

Creates a value of AttachLoadBalancersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

albrsResponseStatus - -- | The response status code.

data AttachLoadBalancersResponse

Contains the output of AttachLoadBalancers.

See: attachLoadBalancersResponse smart constructor.

Response Lenses

albrsResponseStatus :: Lens' AttachLoadBalancersResponse Int

- | The response status code.