Network.AWS.AutoScaling.CompleteLifecycleAction

Description

Completes the lifecycle action for the associated token initiated under the given lifecycle hook with the specified result.

This operation is a part of the basic sequence for adding a lifecycle hook to an Auto Scaling group:

Create a notification target. A target can be either an Amazon SQS queue or an Amazon SNS topic. Create an IAM role. This role allows Auto Scaling to publish lifecycle notifications to the designated SQS queue or SNS topic. Create the lifecycle hook. You can create a hook that acts when instances launch or when instances terminate. If necessary, record the lifecycle action heartbeat to keep the instance in a pending state. Complete the lifecycle action. For more information, see Auto Scaling Pending State and Auto Scaling Terminating State in the Auto Scaling Developer Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CompleteLifecycleAction.html

Request

data CompleteLifecycleAction

Request constructor

completeLifecycleAction

Arguments

:: Text
claLifecycleHookName

-> Text
claAutoScalingGroupName

-> Text
claLifecycleActionToken

-> Text
claLifecycleActionResult

-> CompleteLifecycleAction

CompleteLifecycleAction constructor.

The fields accessible through corresponding lenses are:

claAutoScalingGroupName :: Text

claLifecycleActionResult :: Text

claLifecycleActionToken :: Text

claLifecycleHookName :: Text

Request lenses

claAutoScalingGroupName :: Lens' CompleteLifecycleAction Text

The name of the group for the lifecycle hook.

claLifecycleActionResult :: Lens' CompleteLifecycleAction Text

The action for the group to take. This parameter can be either CONTINUE or ABANDON.

claLifecycleActionToken :: Lens' CompleteLifecycleAction Text

A universally unique identifier (UUID) that identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

claLifecycleHookName :: Lens' CompleteLifecycleAction Text

The name of the lifecycle hook.

Response

data CompleteLifecycleActionResponse

Response constructor

completeLifecycleActionResponse :: CompleteLifecycleActionResponse

CompleteLifecycleActionResponse constructor.