Network.AWS.AutoScaling.CreateAutoScalingGroup

Description

Creates a new Auto Scaling group with the specified name and other attributes. When the creation request is completed, the Auto Scaling group is ready to be used in other calls.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateAutoScalingGroup.html

Request

data CreateAutoScalingGroup

Request constructor

createAutoScalingGroup

Arguments

:: Text
casgAutoScalingGroupName

-> Int
casgMinSize

-> Int
casgMaxSize

-> NonEmpty Text
casgAvailabilityZones

-> CreateAutoScalingGroup

CreateAutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

casgAutoScalingGroupName :: Text

casgAvailabilityZones :: NonEmpty Text

casgDefaultCooldown :: Maybe Int

casgDesiredCapacity :: Maybe Int

casgHealthCheckGracePeriod :: Maybe Int

casgHealthCheckType :: Maybe Text

casgInstanceId :: Maybe Text

casgLaunchConfigurationName :: Maybe Text

casgLoadBalancerNames :: [Text]

casgMaxSize :: Int

casgMinSize :: Int

casgPlacementGroup :: Maybe Text

casgTags :: [Tag]

casgTerminationPolicies :: [Text]

casgVPCZoneIdentifier :: Maybe Text

Request lenses

casgAutoScalingGroupName :: Lens' CreateAutoScalingGroup Text

The name of the Auto Scaling group.

casgAvailabilityZones :: Lens' CreateAutoScalingGroup (NonEmpty Text)

A list of Availability Zones for the Auto Scaling group. This is required unless you have specified subnets.

casgDefaultCooldown :: Lens' CreateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, between a successful scaling activity and the succeeding scaling activity. If a DefaultCooldown period is not specified, Auto Scaling uses the default value of 300 as the default cool down period for the Auto Scaling group. For more information, see Cooldown Period.

casgDesiredCapacity :: Lens' CreateAutoScalingGroup (Maybe Int)

The number of Amazon EC2 instances that should be running in the group. The desired capacity must be greater than or equal to the minimum size and less than or equal to the maximum size specified for the Auto Scaling group.

casgHealthCheckGracePeriod :: Lens' CreateAutoScalingGroup (Maybe Int)

Length of time in seconds after a new Amazon EC2 instance comes into service that Auto Scaling starts checking its health. During this time any health check failure for the that instance is ignored. This is required if you are adding ELB health check. Frequently, new instances need to warm up, briefly, before they can pass a health check. To provide ample warm-up time, set the health check grace period of the group to match the expected startup period of your application. For more information, see Add an Elastic Load Balancing Health Check.

casgHealthCheckType :: Lens' CreateAutoScalingGroup (Maybe Text)

The service you want the health checks from, Amazon EC2 or Elastic Load Balancer. Valid values are EC2 or ELB. By default, the Auto Scaling health check uses the results of Amazon EC2 instance status checks to determine the health of an instance. For more information, see Health Check.

casgInstanceId :: Lens' CreateAutoScalingGroup (Maybe Text)

The ID of the Amazon EC2 instance you want to use to create the Auto Scaling group. Use this attribute if you want to create an Auto Scaling group using an EC2 instance instead of a launch configuration. When you use an instance to create an Auto Scaling group, a new launch configuration is first created and then associated with the Auto Scaling group. The new launch configuration derives all its attributes from the instance that is used to create the Auto Scaling group, with the exception of BlockDeviceMapping. For more information, see Create an Auto Scaling Group Using EC2 Instance in the Auto Scaling Developer Guide.

casgLaunchConfigurationName :: Lens' CreateAutoScalingGroup (Maybe Text)

The name of an existing launch configuration to use to launch new instances. Use this attribute if you want to create an Auto Scaling group using an existing launch configuration instead of an EC2 instance.

casgLoadBalancerNames :: Lens' CreateAutoScalingGroup [Text]

A list of existing Elastic Load Balancing load balancers to use. The load balancers must be associated with the AWS account. For information on using load balancers, see Load Balance Your Auto Scaling Group in the Auto Scaling Developer Guide.

casgMaxSize :: Lens' CreateAutoScalingGroup Int

The maximum size of the Auto Scaling group.

casgMinSize :: Lens' CreateAutoScalingGroup Int

The minimum size of the Auto Scaling group.

casgPlacementGroup :: Lens' CreateAutoScalingGroup (Maybe Text)

Physical location of an existing cluster placement group into which you want to launch your instances. For information about cluster placement group, see Using Cluster Instances.

casgTags :: Lens' CreateAutoScalingGroup [Tag]

The tag to be created or updated. Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values: key=value, value=value, propagate=true or false. Value and propagate are optional parameters. For information about using tags, see Tag Your Auto Scaling Groups and Amazon EC2 Instances in the Auto Scaling Developer Guide.

casgTerminationPolicies :: Lens' CreateAutoScalingGroup [Text]

A standalone termination policy or a list of termination policies used to select the instance to terminate. The policies are executed in the order that they are listed. For more information on configuring a termination policy for your Auto Scaling group, see Instance Termination Policy for Your Auto Scaling Group in the Auto Scaling Developer Guide.

casgVPCZoneIdentifier :: Lens' CreateAutoScalingGroup (Maybe Text)

A comma-separated list of subnet identifiers of Amazon Virtual Private Clouds (Amazon VPCs). If you specify subnets and Availability Zones with this call, ensure that the subnets' Availability Zones match the Availability Zones specified. For information on launching your Auto Scaling group into Amazon VPC subnets, see Auto Scaling in Amazon Virtual Private Cloud in the Auto Scaling Developer Guide .

Response

data CreateAutoScalingGroupResponse

Response constructor

createAutoScalingGroupResponse :: CreateAutoScalingGroupResponse

CreateAutoScalingGroupResponse constructor.