Network.AWS.AutoScaling.CreateLaunchConfiguration

Description

Creates a launch configuration. If you exceed your maximum limit of launch configurations, which by default is 100 per region, the call fails. For information about viewing and updating these limits, see DescribeAccountLimits.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateLaunchConfiguration.html

Request

data CreateLaunchConfiguration

Request constructor

createLaunchConfiguration

Arguments

:: Text
clcLaunchConfigurationName

-> CreateLaunchConfiguration

CreateLaunchConfiguration constructor.

The fields accessible through corresponding lenses are:

clcAssociatePublicIpAddress :: Maybe Bool

clcBlockDeviceMappings :: [BlockDeviceMapping]

clcEbsOptimized :: Maybe Bool

clcIamInstanceProfile :: Maybe Text

clcImageId :: Maybe Text

clcInstanceId :: Maybe Text

clcInstanceMonitoring :: Maybe InstanceMonitoring

clcInstanceType :: Maybe Text

clcKernelId :: Maybe Text

clcKeyName :: Maybe Text

clcLaunchConfigurationName :: Text

clcPlacementTenancy :: Maybe Text

clcRamdiskId :: Maybe Text

clcSecurityGroups :: [Text]

clcSpotPrice :: Maybe Text

clcUserData :: Maybe Text

Request lenses

clcAssociatePublicIpAddress :: Lens' CreateLaunchConfiguration (Maybe Bool)

Used for groups that launch instances into a virtual private cloud (VPC). Specifies whether to assign a public IP address to each instance. For more information, see Auto Scaling and Amazon VPC in the Auto Scaling Developer Guide. If you specify a value for this parameter, be sure to specify at least one subnet using the VPCZoneIdentifier parameter when you create your group. Default: If the instance is launched into a default subnet, the default is true. If the instance is launched into a nondefault subnet, the default is false. For more information, see Supported Platforms in the Amazon Elastic Compute Cloud User Guide.

clcBlockDeviceMappings :: Lens' CreateLaunchConfiguration [BlockDeviceMapping]

One or more mappings that specify how block devices are exposed to the instance. For more information, see Block Device Mapping in the Amazon Elastic Compute Cloud User Guide.

clcEbsOptimized :: Lens' CreateLaunchConfiguration (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS I/O. By default, the instance is not optimized for EBS I/O. The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal I/O performance. This optimization is not available with all instance types. Additional usage charges apply. For more information, see Amazon EBS-Optimized Instances in the Amazon Elastic Compute Cloud User Guide.

clcIamInstanceProfile :: Lens' CreateLaunchConfiguration (Maybe Text)

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. Amazon EC2 instances launched with an IAM role will automatically have AWS security credentials available. You can use IAM roles with Auto Scaling to automatically enable applications running on your Amazon EC2 instances to securely access other AWS resources. For more information, see Launch Auto Scaling Instances with an IAM Role in the Auto Scaling Developer Guide.

clcImageId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon Elastic Compute Cloud User Guide.

clcInstanceId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the EC2 instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, with the exception of the block device mapping. To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request. For more information, see Create a Launch Configuration Using an EC2 Instance in the Auto Scaling Developer Guide.

clcInstanceMonitoring :: Lens' CreateLaunchConfiguration (Maybe InstanceMonitoring)

Enables detailed monitoring if it is disabled. Detailed monitoring is enabled by default. When detailed monitoring is enabled, Amazon Cloudwatch generates metrics every minute and your account is charged a fee. When you disable detailed monitoring, by specifying False, Cloudwatch generates metrics every 5 minutes. For more information, see Monitor Your Auto Scaling Instances in the Auto Scaling Developer Guide.

clcInstanceType :: Lens' CreateLaunchConfiguration (Maybe Text)

The instance type of the Amazon EC2 instance. For information about available Amazon EC2 instance types, see Available Instance Types in the Amazon Elastic Cloud Compute User Guide.

clcKernelId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the kernel associated with the Amazon EC2 AMI.

clcKeyName :: Lens' CreateLaunchConfiguration (Maybe Text)

The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon Elastic Compute Cloud User Guide.

clcLaunchConfigurationName :: Lens' CreateLaunchConfiguration Text

The name of the launch configuration. This name must be unique within the scope of your AWS account.

clcPlacementTenancy :: Lens' CreateLaunchConfiguration (Maybe Text)

The tenancy of the instance. An instance with a tenancy of dedicated runs on single-tenant hardware and can only be launched in a VPC. You must set the value of this parameter to dedicated if want to launch Dedicated Instances in a shared tenancy VPC (VPC with instance placement tenancy attribute set to default). If you specify a value for this parameter, be sure to specify at least one VPC subnet using the VPCZoneIdentifier parameter when you create your group. For more information, see Auto Scaling and Amazon VPC in the Auto Scaling Developer Guide. Valid values: default | dedicated.

clcRamdiskId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the RAM disk associated with the Amazon EC2 AMI.

clcSecurityGroups :: Lens' CreateLaunchConfiguration [Text]

One or more security groups with which to associate the instances. If your instances are launched in EC2-Classic, you can either specify security group names or the security group IDs. For more information about security groups for EC2-Classic, see Amazon EC2 Security Groups in the Amazon Elastic Compute Cloud User Guide. If your instances are launched in a VPC, specify security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

clcSpotPrice :: Lens' CreateLaunchConfiguration (Maybe Text)

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot market price. For more information, see Launch Spot Instances in Your Auto Scaling Group in the Auto Scaling Developer Guide.

clcUserData :: Lens' CreateLaunchConfiguration (Maybe Text)

The user data to make available to the launched EC2 instances. For more information, see Instance Metadata and User Data in the Amazon Elastic Compute Cloud User Guide. At this time, launch configurations don't support compressed (zipped) user data files.

Response

data CreateLaunchConfigurationResponse

Response constructor

createLaunchConfigurationResponse :: CreateLaunchConfigurationResponse

CreateLaunchConfigurationResponse constructor.