Network.AWS.AutoScaling.CreateOrUpdateTags

Description

Creates or updates tags for the specified Auto Scaling group.

A tag's definition is composed of a resource ID, resource type, key and value, and the propagate flag. Value and the propagate flag are optional parameters. See the Request Parameters for more information. For more information, see Add, Modify, or Remove Auto Scaling Group Tags in the AutoScaling Developer Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateOrUpdateTags.html

Request

data CreateOrUpdateTags

Request constructor

createOrUpdateTags :: CreateOrUpdateTags

CreateOrUpdateTags constructor.

The fields accessible through corresponding lenses are:

coutTags :: [Tag]

Request lenses

coutTags :: Lens' CreateOrUpdateTags [Tag]

The tag to be created or updated. Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. The resource type and resource ID identify the type and name of resource for which the tag is created. Currently, 'auto-scaling-group' is the only supported resource type. The valid value for the resource ID is groupname.

The PropagateAtLaunch flag defines whether the new tag will be applied to instances launched by the group. Valid values are true or false. However, instances that are already running will not get the new or updated tag. Likewise, when you modify a tag, the updated version will be applied only to new instances launched by the group after the change. Running instances that had the previous version of the tag will continue to have the older tag.

When you create a tag and a tag of the same name already exists, the operation overwrites the previous tag definition, but you will not get an error message.

Response

data CreateOrUpdateTagsResponse

Response constructor

createOrUpdateTagsResponse :: CreateOrUpdateTagsResponse

CreateOrUpdateTagsResponse constructor.