Network.AWS.AutoScaling.DeleteAutoScalingGroup

Description

Deletes the specified Auto Scaling group if the group has no instances and no scaling activities in progress. To remove all instances before calling DeleteAutoScalingGroup, you can call UpdateAutoScalingGroup to set the minimum and maximum size of the AutoScalingGroup to zero. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ForceDelete=true &Version=2011-01-01 &Action=DeleteAutoScalingGroup &AUTHPARAMS 70a76d42-9665-11e2-9fdf-211deEXAMPLE.

Request

data DeleteAutoScalingGroup

Request constructor

deleteAutoScalingGroup

Arguments

:: Text
dasgAutoScalingGroupName

-> DeleteAutoScalingGroup

Smart constructor for the minimum required parameters to construct a valid DeleteAutoScalingGroup request.

The fields accessible through corresponding lenses are:

AutoScalingGroupName :: Text

ForceDelete :: Maybe Bool

Request lenses

dasgAutoScalingGroupName :: Lens' DeleteAutoScalingGroup Text

The name of the Auto Scaling group to delete.

dasgForceDelete :: Lens' DeleteAutoScalingGroup (Maybe Bool)

Starting with API version 2011-01-01, specifies that the Auto Scaling group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any lifecycle actions associated with the group.

Response

data DeleteAutoScalingGroupResponse

Response constructor

deleteAutoScalingGroupResponse :: DeleteAutoScalingGroupResponse

Smart constructor for the minimum required parameters to construct a valid DeleteAutoScalingGroupResponse response.

This constructor is provided for convenience and testing purposes.