Network.AWS.AutoScaling.DeleteLaunchConfiguration

Description

Deletes the specified launch configuration.

The launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteLaunchConfiguration.html

Request

data DeleteLaunchConfiguration

Request constructor

deleteLaunchConfiguration

Arguments

:: Text
dlcLaunchConfigurationName

-> DeleteLaunchConfiguration

DeleteLaunchConfiguration constructor.

The fields accessible through corresponding lenses are:

dlcLaunchConfigurationName :: Text

Request lenses

dlcLaunchConfigurationName :: Lens' DeleteLaunchConfiguration Text

The name of the launch configuration.

Response

data DeleteLaunchConfigurationResponse

Response constructor

deleteLaunchConfigurationResponse :: DeleteLaunchConfigurationResponse

DeleteLaunchConfigurationResponse constructor.