Network.AWS.AutoScaling.DeleteLifecycleHook

Description

Deletes the specified lifecycle hook.

If there are any outstanding lifecycle actions, they are completed first (ABANDON for launching instances, CONTINUE for terminating instances).

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteLifecycleHook.html

Request

data DeleteLifecycleHook

Request constructor

deleteLifecycleHook

Arguments

:: Text
dlh1LifecycleHookName

-> Text
dlh1AutoScalingGroupName

-> DeleteLifecycleHook

DeleteLifecycleHook constructor.

The fields accessible through corresponding lenses are:

dlh1AutoScalingGroupName :: Text

dlh1LifecycleHookName :: Text

Request lenses

dlh1AutoScalingGroupName :: Lens' DeleteLifecycleHook Text

The name of the Auto Scaling group for the lifecycle hook.

dlh1LifecycleHookName :: Lens' DeleteLifecycleHook Text

The name of the lifecycle hook.

Response

data DeleteLifecycleHookResponse

Response constructor

deleteLifecycleHookResponse :: DeleteLifecycleHookResponse

DeleteLifecycleHookResponse constructor.