Network.AWS.AutoScaling.DeletePolicy

Description

Deletes the specified Auto Scaling policy.

Deleting a policy deletes the underlying alarm action, but does not delete the alarm, even if it no longer has an associated action.

Creating a Request

deletePolicy

Arguments

:: Text
dpPolicyName

-> DeletePolicy

Creates a value of DeletePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpAutoScalingGroupName - The name of the Auto Scaling group.

dpPolicyName - The name or Amazon Resource Name (ARN) of the policy.

data DeletePolicy

Contains the parameters for DeletePolicy.

See: deletePolicy smart constructor.

Request Lenses

dpAutoScalingGroupName :: Lens' DeletePolicy (Maybe Text)

The name of the Auto Scaling group.

dpPolicyName :: Lens' DeletePolicy Text

The name or Amazon Resource Name (ARN) of the policy.

Destructuring the Response

deletePolicyResponse :: DeletePolicyResponse

Creates a value of DeletePolicyResponse with the minimum fields required to make a request.

data DeletePolicyResponse

See: deletePolicyResponse smart constructor.