Network.AWS.AutoScaling.DescribeAccountLimits

Description

Describes the current Auto Scaling resource limits for your AWS account.

For information about requesting an increase in these limits, see AWSService Limits.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeAccountLimits.html

Request

data DescribeAccountLimits

Request constructor

describeAccountLimits :: DescribeAccountLimits

DescribeAccountLimits constructor.

Response

data DescribeAccountLimitsResponse

Response constructor

describeAccountLimitsResponse :: DescribeAccountLimitsResponse

DescribeAccountLimitsResponse constructor.

The fields accessible through corresponding lenses are:

dalrMaxNumberOfAutoScalingGroups :: Maybe Int

dalrMaxNumberOfLaunchConfigurations :: Maybe Int

Response lenses

dalrMaxNumberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of groups allowed for your AWS account. The default limit is 20 per region.

dalrMaxNumberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of launch configurations allowed for your AWS account. The default limit is 100 per region.