Network.AWS.AutoScaling.DescribeAdjustmentTypes

Description

Describes the policy adjustment types for use with PutScalingPolicy .

Creating a Request

describeAdjustmentTypes :: DescribeAdjustmentTypes

Creates a value of DescribeAdjustmentTypes with the minimum fields required to make a request.

data DescribeAdjustmentTypes

See: describeAdjustmentTypes smart constructor.

Destructuring the Response

describeAdjustmentTypesResponse

Arguments

:: Int
datrsResponseStatus

-> DescribeAdjustmentTypesResponse

Creates a value of DescribeAdjustmentTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datrsAdjustmentTypes - The policy adjustment types.

datrsResponseStatus - -- | The response status code.

data DescribeAdjustmentTypesResponse

Contains the parameters for DescribeAdjustmentTypes.

See: describeAdjustmentTypesResponse smart constructor.

Response Lenses

datrsAdjustmentTypes :: Lens' DescribeAdjustmentTypesResponse [AdjustmentType]

The policy adjustment types.

datrsResponseStatus :: Lens' DescribeAdjustmentTypesResponse Int

- | The response status code.