Network.AWS.AutoScaling.DescribeAutoScalingGroups

Description

Returns a full description of each Auto Scaling group in the given list. This includes all Amazon EC2 instances that are members of the group. If a list of names is not provided, the service returns the full details of all Auto Scaling groups. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter. https:/autoscaling.amazonaws.com?AutoScalingGroupNames.member.1=my-test-asg-lbs &MaxRecords=20 &Version=2011-01-01 &Action=DescribeAutoScalingGroups &AUTHPARAMS my-test-asg-lbs ELB 2013-05-06T17:47:15.107Z my-test-lc 2 us-east-1b us-east-1a my-test-asg-loadbalancer 2 120 300 arn:aws:autoscaling:us-east-1:803981987763:autoScalingGroup:ca861182-c8f9-4ca7-b1eb-cd35505f5ebb :autoScalingGroupName/my-test-asg-lbs Default 10 0f02a07d-b677-11e2-9eb0-dd50EXAMPLE.

Request

data DescribeAutoScalingGroups

The AutoScalingGroupNamesType data type.

Request constructor

mkDescribeAutoScalingGroups :: DescribeAutoScalingGroups

Smart constructor for the minimum required parameters to construct a valid DescribeAutoScalingGroups request.

The fields accessible through corresponding lenses are:

AutoScalingGroupNames :: [Text]

NextToken :: Maybe Text

MaxRecords :: Maybe Integer

Request lenses

dasg1AutoScalingGroupNames :: Lens' DescribeAutoScalingGroups [Text]

A list of Auto Scaling group names.

dasg1NextToken :: Lens' DescribeAutoScalingGroups (Maybe Text)

A string that marks the start of the next batch of returned results.

dasg1MaxRecords :: Lens' DescribeAutoScalingGroups (Maybe Integer)

The maximum number of records to return.

Response

data DescribeAutoScalingGroupsResponse

The AutoScalingGroupsType data type.

Response constructor

mkDescribeAutoScalingGroupsResponse

Arguments

:: [AutoScalingGroup]
dasgrAutoScalingGroups

-> DescribeAutoScalingGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAutoScalingGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AutoScalingGroups :: [AutoScalingGroup]

NextToken :: Maybe Text

Response lenses

dasgrAutoScalingGroups :: Lens' DescribeAutoScalingGroupsResponse [AutoScalingGroup]

A list of Auto Scaling groups.

dasgrNextToken :: Lens' DescribeAutoScalingGroupsResponse (Maybe Text)

A string that marks the start of the next batch of returned results.