Network.AWS.AutoScaling.DescribeAutoScalingInstances

Description

Returns a description of each Auto Scaling instance in the InstanceIds list. If a list is not provided, the service returns the full details of all instances up to a maximum of 50. By default, the service returns a list of 20 items. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter. https:/autoscaling.amazonaws.com?MaxRecords=20 &InstanceIds.member.1=i-78e0d40b &Version=2011-01-01 &Action=DescribeAutoScalingInstances &AUTHPARAMS Healthy my-test-asg us-east-1e i-78e0d40b my-test-lc InService df992dc3-b72f-11e2-81e1-750aa6EXAMPLE.

Request

data DescribeAutoScalingInstances

Request constructor

describeAutoScalingInstances :: DescribeAutoScalingInstances

Smart constructor for the minimum required parameters to construct a valid DescribeAutoScalingInstances request.

The fields accessible through corresponding lenses are:

InstanceIds :: [Text]

MaxRecords :: Maybe Integer

NextToken :: Maybe Text

Request lenses

dasiInstanceIds :: Lens' DescribeAutoScalingInstances [Text]

The list of Auto Scaling instances to describe. If this list is omitted, all auto scaling instances are described. The list of requested instances cannot contain more than 50 items. If unknown instances are requested, they are ignored with no error.

dasiMaxRecords :: Lens' DescribeAutoScalingInstances (Maybe Integer)

The maximum number of Auto Scaling instances to be described with each call.

dasiNextToken :: Lens' DescribeAutoScalingInstances (Maybe Text)

The token returned by a previous call to indicate that there is more data available.

Response

data DescribeAutoScalingInstancesResponse

The AutoScalingInstancesType data type.

Response constructor

describeAutoScalingInstancesResponse :: DescribeAutoScalingInstancesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAutoScalingInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AutoScalingInstances :: [AutoScalingInstanceDetails]

NextToken :: Maybe Text

Response lenses

dasirAutoScalingInstances :: Lens' DescribeAutoScalingInstancesResponse [AutoScalingInstanceDetails]

A list of Auto Scaling instances.

dasirNextToken :: Lens' DescribeAutoScalingInstancesResponse (Maybe Text)

A string that marks the start of the next batch of returned results.