Network.AWS.AutoScaling.DescribeAutoScalingNotificationTypes

Description

Describes the notification types that are supported by Auto Scaling.

Creating a Request

describeAutoScalingNotificationTypes :: DescribeAutoScalingNotificationTypes

Creates a value of DescribeAutoScalingNotificationTypes with the minimum fields required to make a request.

data DescribeAutoScalingNotificationTypes

See: describeAutoScalingNotificationTypes smart constructor.

Destructuring the Response

describeAutoScalingNotificationTypesResponse

Arguments

:: Int
dasntrsResponseStatus

-> DescribeAutoScalingNotificationTypesResponse

Creates a value of DescribeAutoScalingNotificationTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasntrsAutoScalingNotificationTypes - The notification types.

dasntrsResponseStatus - -- | The response status code.

data DescribeAutoScalingNotificationTypesResponse

Contains the output of DescribeAutoScalingNotificationTypes.

See: describeAutoScalingNotificationTypesResponse smart constructor.

Response Lenses

dasntrsAutoScalingNotificationTypes :: Lens' DescribeAutoScalingNotificationTypesResponse [Text]

The notification types.

dasntrsResponseStatus :: Lens' DescribeAutoScalingNotificationTypesResponse Int

- | The response status code.