Network.AWS.AutoScaling.DescribeLifecycleHooks

Description

Describes the lifecycle hooks that currently belong to the specified Auto Scaling group.

Request

data DescribeLifecycleHooks

Request constructor

describeLifecycleHooks

Arguments

:: Text
dlh1AutoScalingGroupName

-> DescribeLifecycleHooks

Smart constructor for the minimum required parameters to construct a valid DescribeLifecycleHooks request.

The fields accessible through corresponding lenses are:

AutoScalingGroupName :: Text

LifecycleHookNames :: [Text]

Request lenses

dlh1AutoScalingGroupName :: Lens' DescribeLifecycleHooks Text

The name of one or more Auto Scaling groups.

dlh1LifecycleHookNames :: Lens' DescribeLifecycleHooks [Text]

The name of one or more lifecycle hooks.

Response

data DescribeLifecycleHooksResponse

The output of the DescribeLifecycleHooks action.

Response constructor

describeLifecycleHooksResponse :: DescribeLifecycleHooksResponse

Smart constructor for the minimum required parameters to construct a valid DescribeLifecycleHooksResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LifecycleHooks :: [LifecycleHook]

Response lenses

dlhrrLifecycleHooks :: Lens' DescribeLifecycleHooksResponse [LifecycleHook]

A list describing the lifecycle hooks that belong to the specified Auto Scaling group.