Network.AWS.AutoScaling.DescribeNotificationConfigurations

Description

Describes the notification actions associated with the specified Auto Scaling group.

This operation returns paginated results.

Creating a Request

describeNotificationConfigurations :: DescribeNotificationConfigurations

Creates a value of DescribeNotificationConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dncAutoScalingGroupNames - The name of the group.

dncNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dncMaxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

data DescribeNotificationConfigurations

Contains the parameters for DescribeNotificationConfigurations.

See: describeNotificationConfigurations smart constructor.

Request Lenses

dncAutoScalingGroupNames :: Lens' DescribeNotificationConfigurations [Text]

The name of the group.

dncNextToken :: Lens' DescribeNotificationConfigurations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dncMaxRecords :: Lens' DescribeNotificationConfigurations (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Destructuring the Response

describeNotificationConfigurationsResponse

Arguments

:: Int
dncrsResponseStatus

-> DescribeNotificationConfigurationsResponse

Creates a value of DescribeNotificationConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dncrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dncrsResponseStatus - -- | The response status code.

dncrsNotificationConfigurations - The notification configurations.

data DescribeNotificationConfigurationsResponse

Contains the output from DescribeNotificationConfigurations.

See: describeNotificationConfigurationsResponse smart constructor.

Response Lenses

dncrsNextToken :: Lens' DescribeNotificationConfigurationsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dncrsResponseStatus :: Lens' DescribeNotificationConfigurationsResponse Int

- | The response status code.

dncrsNotificationConfigurations :: Lens' DescribeNotificationConfigurationsResponse [NotificationConfiguration]

The notification configurations.