Network.AWS.AutoScaling.DescribePolicies

Description

Returns descriptions of what each policy does. This action supports pagination. If the response includes a token, there are more records available. To get the additional records, repeat the request with the response token as the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribePolicies.html

Request

data DescribePolicies

Request constructor

describePolicies :: DescribePolicies

DescribePolicies constructor.

The fields accessible through corresponding lenses are:

dp1AutoScalingGroupName :: Maybe Text

dp1MaxRecords :: Maybe Int

dp1NextToken :: Maybe Text

dp1PolicyNames :: [Text]

Request lenses

dp1AutoScalingGroupName :: Lens' DescribePolicies (Maybe Text)

The name of the Auto Scaling group.

dp1MaxRecords :: Lens' DescribePolicies (Maybe Int)

The maximum number of policies that will be described with each call.

dp1NextToken :: Lens' DescribePolicies (Maybe Text)

A string that is used to mark the start of the next batch of returned results for pagination.

dp1PolicyNames :: Lens' DescribePolicies [Text]

A list of policy names or policy ARNs to be described. If this list is omitted, all policy names are described. If an auto scaling group name is provided, the results are limited to that group. The list of requested policy names cannot contain more than 50 items. If unknown policy names are requested, they are ignored with no error.

Response

data DescribePoliciesResponse

Response constructor

describePoliciesResponse :: DescribePoliciesResponse

DescribePoliciesResponse constructor.

The fields accessible through corresponding lenses are:

dprNextToken :: Maybe Text

dprScalingPolicies :: [ScalingPolicy]

Response lenses

dprNextToken :: Lens' DescribePoliciesResponse (Maybe Text)

A string that marks the start of the next batch of returned results.

dprScalingPolicies :: Lens' DescribePoliciesResponse [ScalingPolicy]

A list of scaling policies.