Network.AWS.AutoScaling.DescribeScalingProcessTypes

Description

Returns scaling process types for use in the ResumeProcesses and SuspendProcesses actions. https:/autoscaling.amazonaws.com?Version=2011-01-01 &Action=DescribeScalingProcessTypes &AUTHPARAMS AZRebalance AddToLoadBalancer AlarmNotification HealthCheck Launch ReplaceUnhealthy ScheduledActions Terminate 27f2eacc-b73f-11e2-ad99-c7aba3a9c963.

Request

data DescribeScalingProcessTypes

Request constructor

mkDescribeScalingProcessTypes :: DescribeScalingProcessTypes

Smart constructor for the minimum required parameters to construct a valid DescribeScalingProcessTypes request.

Response

data DescribeScalingProcessTypesResponse

The output of the DescribeScalingProcessTypes action.

Response constructor

mkDescribeScalingProcessTypesResponse :: DescribeScalingProcessTypesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeScalingProcessTypesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Processes :: [ProcessType]

Response lenses

dsptrProcesses :: Lens' DescribeScalingProcessTypesResponse [ProcessType]

A list of ProcessType names.