Network.AWS.AutoScaling.DescribeScheduledActions

Description

Describes the actions scheduled for your Auto Scaling group that haven't run. To describe the actions that have already run, use DescribeScalingActivities .

This operation returns paginated results.

Creating a Request

describeScheduledActions :: DescribeScheduledActions

Creates a value of DescribeScheduledActions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsasStartTime - The earliest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

dsasNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dsasAutoScalingGroupName - The name of the group.

dsasMaxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

dsasEndTime - The latest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

dsasScheduledActionNames - Describes one or more scheduled actions. If you omit this parameter, all scheduled actions are described. If you specify an unknown scheduled action, it is ignored with no error. You can describe up to a maximum of 50 instances with a single call. If there are more items to return, the call returns a token. To get the next set of items, repeat the call with the returned token.

data DescribeScheduledActions

Contains the parameters for DescribeScheduledActions.

See: describeScheduledActions smart constructor.

Request Lenses

dsasStartTime :: Lens' DescribeScheduledActions (Maybe UTCTime)

The earliest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

dsasNextToken :: Lens' DescribeScheduledActions (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dsasAutoScalingGroupName :: Lens' DescribeScheduledActions (Maybe Text)

The name of the group.

dsasMaxRecords :: Lens' DescribeScheduledActions (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

dsasEndTime :: Lens' DescribeScheduledActions (Maybe UTCTime)

The latest scheduled start time to return. If scheduled action names are provided, this parameter is ignored.

dsasScheduledActionNames :: Lens' DescribeScheduledActions [Text]

Describes one or more scheduled actions. If you omit this parameter, all scheduled actions are described. If you specify an unknown scheduled action, it is ignored with no error. You can describe up to a maximum of 50 instances with a single call. If there are more items to return, the call returns a token. To get the next set of items, repeat the call with the returned token.

Destructuring the Response

describeScheduledActionsResponse

Arguments

:: Int
dsarsResponseStatus

-> DescribeScheduledActionsResponse

Creates a value of DescribeScheduledActionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsarsScheduledUpdateGroupActions - The scheduled actions.

dsarsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dsarsResponseStatus - -- | The response status code.

data DescribeScheduledActionsResponse

Contains the output of DescribeScheduledActions.

See: describeScheduledActionsResponse smart constructor.

Response Lenses

dsarsScheduledUpdateGroupActions :: Lens' DescribeScheduledActionsResponse [ScheduledUpdateGroupAction]

The scheduled actions.

dsarsNextToken :: Lens' DescribeScheduledActionsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dsarsResponseStatus :: Lens' DescribeScheduledActionsResponse Int

- | The response status code.