Network.AWS.AutoScaling.DescribeTags

Description

Describes the specified tags.

You can use filters to limit the results. For example, you can query for the tags for a specific Auto Scaling group. You can specify multiple values for a filter. A tag must match at least one of the specified values for it to be included in the results.

You can also specify multiple filters. The result includes information for a particular tag only if it matches all the filters. If there's no match, no special message is returned.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeTags.html

Request

data DescribeTags

Request constructor

describeTags :: DescribeTags

DescribeTags constructor.

The fields accessible through corresponding lenses are:

dtFilters :: [Filter]

dtMaxRecords :: Maybe Int

dtNextToken :: Maybe Text

Request lenses

dtFilters :: Lens' DescribeTags [Filter]

The value of the filter type used to identify the tags to be returned. For example, you can filter so that tags are returned according to Auto Scaling group, the key and value, or whether the new tag will be applied to instances launched after the tag is created (PropagateAtLaunch).

dtMaxRecords :: Lens' DescribeTags (Maybe Int)

The maximum number of items to return with this call.

dtNextToken :: Lens' DescribeTags (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

Response

data DescribeTagsResponse

Response constructor

describeTagsResponse :: DescribeTagsResponse

DescribeTagsResponse constructor.

The fields accessible through corresponding lenses are:

dtrNextToken :: Maybe Text

dtrTags :: [TagDescription]

Response lenses

dtrNextToken :: Lens' DescribeTagsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dtrTags :: Lens' DescribeTagsResponse [TagDescription]

The tags.