Network.AWS.AutoScaling.DescribeTerminationPolicyTypes

Description

Describes the termination policies supported by Auto Scaling.

Creating a Request

describeTerminationPolicyTypes :: DescribeTerminationPolicyTypes

Creates a value of DescribeTerminationPolicyTypes with the minimum fields required to make a request.

data DescribeTerminationPolicyTypes

See: describeTerminationPolicyTypes smart constructor.

Destructuring the Response

describeTerminationPolicyTypesResponse

Arguments

:: Int
dtptrsResponseStatus

-> DescribeTerminationPolicyTypesResponse

Creates a value of DescribeTerminationPolicyTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtptrsTerminationPolicyTypes - The termination policies supported by Auto Scaling (OldestInstance , OldestLaunchConfiguration , NewestInstance , ClosestToNextInstanceHour , and Default ).

dtptrsResponseStatus - -- | The response status code.

data DescribeTerminationPolicyTypesResponse

Contains the output of DescribeTerminationPolicyTypes.

See: describeTerminationPolicyTypesResponse smart constructor.

Response Lenses

dtptrsTerminationPolicyTypes :: Lens' DescribeTerminationPolicyTypesResponse [Text]

The termination policies supported by Auto Scaling (OldestInstance , OldestLaunchConfiguration , NewestInstance , ClosestToNextInstanceHour , and Default ).

dtptrsResponseStatus :: Lens' DescribeTerminationPolicyTypesResponse Int

- | The response status code.