Network.AWS.AutoScaling.DetachInstances

Description

Removes one or more instances from the specified Auto Scaling group.

After the instances are detached, you can manage them independently from the rest of the Auto Scaling group.

If you do not specify the option to decrement the desired capacity, Auto Scaling launches instances to replace the ones that are detached.

If there is a Classic Load Balancer attached to the Auto Scaling group, the instances are deregistered from the load balancer. If there are target groups attached to the Auto Scaling group, the instances are deregistered from the target groups.

For more information, see Detach EC2 Instances from Your Auto Scaling Group in the Auto Scaling User Guide .

Creating a Request

detachInstances

Arguments

:: Text
diAutoScalingGroupName

-> Bool
diShouldDecrementDesiredCapacity

-> DetachInstances

Creates a value of DetachInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diInstanceIds - One or more instance IDs.

diAutoScalingGroupName - The name of the group.

diShouldDecrementDesiredCapacity - If True , the Auto Scaling group decrements the desired capacity value by the number of instances detached.

data DetachInstances

Contains the parameters for DetachInstances.

See: detachInstances smart constructor.

Request Lenses

diInstanceIds :: Lens' DetachInstances [Text]

One or more instance IDs.

diAutoScalingGroupName :: Lens' DetachInstances Text

The name of the group.

diShouldDecrementDesiredCapacity :: Lens' DetachInstances Bool

If True , the Auto Scaling group decrements the desired capacity value by the number of instances detached.

Destructuring the Response

detachInstancesResponse

Arguments

:: Int
dirsResponseStatus

-> DetachInstancesResponse

Creates a value of DetachInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dirsActivities - The activities related to detaching the instances from the Auto Scaling group.

dirsResponseStatus - -- | The response status code.

data DetachInstancesResponse

Contains the output of DetachInstances.

See: detachInstancesResponse smart constructor.

Response Lenses

dirsActivities :: Lens' DetachInstancesResponse [Activity]

The activities related to detaching the instances from the Auto Scaling group.

dirsResponseStatus :: Lens' DetachInstancesResponse Int

- | The response status code.