Network.AWS.AutoScaling.DetachLoadBalancers

Description

Detaches one or more Classic Load Balancers from the specified Auto Scaling group.

Note that this operation detaches only Classic Load Balancers. If you have Application Load Balancers, use DetachLoadBalancerTargetGroups instead.

When you detach a load balancer, it enters the Removing state while deregistering the instances in the group. When all instances are deregistered, then you can no longer describe the load balancer using DescribeLoadBalancers . Note that the instances remain running.

Creating a Request

detachLoadBalancers

Arguments

:: Text
dAutoScalingGroupName

-> DetachLoadBalancers

Creates a value of DetachLoadBalancers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dAutoScalingGroupName - The name of the Auto Scaling group.

dLoadBalancerNames - One or more load balancer names.

data DetachLoadBalancers

Contains the parameters for DetachLoadBalancers.

See: detachLoadBalancers smart constructor.

Request Lenses

dAutoScalingGroupName :: Lens' DetachLoadBalancers Text

The name of the Auto Scaling group.

dLoadBalancerNames :: Lens' DetachLoadBalancers [Text]

One or more load balancer names.

Destructuring the Response

detachLoadBalancersResponse

Arguments

:: Int
dlbsrsResponseStatus

-> DetachLoadBalancersResponse

Creates a value of DetachLoadBalancersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbsrsResponseStatus - -- | The response status code.

data DetachLoadBalancersResponse

Contains the output for DetachLoadBalancers.

See: detachLoadBalancersResponse smart constructor.

Response Lenses

dlbsrsResponseStatus :: Lens' DetachLoadBalancersResponse Int

- | The response status code.