Network.AWS.AutoScaling.DisableMetricsCollection

Description

Disables monitoring of the specified metrics for the specified Auto Scaling group.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DisableMetricsCollection.html

Request

data DisableMetricsCollection

Request constructor

disableMetricsCollection

Arguments

:: Text
dmcAutoScalingGroupName

-> DisableMetricsCollection

DisableMetricsCollection constructor.

The fields accessible through corresponding lenses are:

dmcAutoScalingGroupName :: Text

dmcMetrics :: [Text]

Request lenses

dmcAutoScalingGroupName :: Lens' DisableMetricsCollection Text

The name or Amazon Resource Name (ARN) of the group.

dmcMetrics :: Lens' DisableMetricsCollection [Text]

One or more of the following metrics:

GroupMinSize

GroupMaxSize

GroupDesiredCapacity

GroupInServiceInstances

GroupPendingInstances

GroupStandbyInstances

GroupTerminatingInstances

GroupTotalInstances

If you omit this parameter, all metrics are disabled.

Response

data DisableMetricsCollectionResponse

Response constructor

disableMetricsCollectionResponse :: DisableMetricsCollectionResponse

DisableMetricsCollectionResponse constructor.