Network.AWS.AutoScaling.EnterStandby

Description

Moves the specified instances into Standby mode.

For more information, see Auto Scaling InService State in the Auto ScalingDeveloper Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_EnterStandby.html

Request

data EnterStandby

Request constructor

enterStandby

Arguments

:: Text
esAutoScalingGroupName

-> Bool
esShouldDecrementDesiredCapacity

-> EnterStandby

EnterStandby constructor.

The fields accessible through corresponding lenses are:

esAutoScalingGroupName :: Text

esInstanceIds :: [Text]

esShouldDecrementDesiredCapacity :: Bool

Request lenses

esAutoScalingGroupName :: Lens' EnterStandby Text

The name of the Auto Scaling group.

esInstanceIds :: Lens' EnterStandby [Text]

One or more instances to move into Standby mode. You must specify at least one instance ID.

esShouldDecrementDesiredCapacity :: Lens' EnterStandby Bool

Specifies whether the instances moved to Standby mode count as part of the Auto Scaling group's desired capacity. If set, the desired capacity for the Auto Scaling group decrements by the number of instances moved to Standby mode.

Response

data EnterStandbyResponse

Response constructor

enterStandbyResponse :: EnterStandbyResponse

EnterStandbyResponse constructor.

The fields accessible through corresponding lenses are:

esr1Activities :: [Activity]

Response lenses

esr1Activities :: Lens' EnterStandbyResponse [Activity]

The activities related to moving instances into Standby mode.