Network.AWS.AutoScaling.ExecutePolicy

Description

Executes the specified policy.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ExecutePolicy.html

Request

data ExecutePolicy

Request constructor

executePolicy

Arguments

:: Text
epPolicyName

-> ExecutePolicy

ExecutePolicy constructor.

The fields accessible through corresponding lenses are:

epAutoScalingGroupName :: Maybe Text

epHonorCooldown :: Maybe Bool

epPolicyName :: Text

Request lenses

epAutoScalingGroupName :: Lens' ExecutePolicy (Maybe Text)

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

epHonorCooldown :: Lens' ExecutePolicy (Maybe Bool)

Set to True if you want Auto Scaling to wait for the cooldown period associated with the Auto Scaling group to complete before executing the policy.

Set to False if you want Auto Scaling to circumvent the cooldown period associated with the Auto Scaling group and execute the policy before the cooldown period ends.

For more information, see Understanding Auto Scaling Cooldowns in the AutoScaling Developer Guide.

epPolicyName :: Lens' ExecutePolicy Text

The name or ARN of the policy.

Response

data ExecutePolicyResponse

Response constructor

executePolicyResponse :: ExecutePolicyResponse

ExecutePolicyResponse constructor.