Network.AWS.AutoScaling.ExitStandby

Description

Moves the specified instances out of Standby mode.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

Creating a Request

exitStandby

Arguments

:: Text
eAutoScalingGroupName

-> ExitStandby

Creates a value of ExitStandby with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eInstanceIds - One or more instance IDs. You must specify at least one instance ID.

eAutoScalingGroupName - The name of the Auto Scaling group.

data ExitStandby

Contains the parameters for ExitStandby.

See: exitStandby smart constructor.

Request Lenses

eInstanceIds :: Lens' ExitStandby [Text]

One or more instance IDs. You must specify at least one instance ID.

eAutoScalingGroupName :: Lens' ExitStandby Text

The name of the Auto Scaling group.

Destructuring the Response

exitStandbyResponse

Arguments

:: Int
esrsResponseStatus

-> ExitStandbyResponse

Creates a value of ExitStandbyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esrsActivities - The activities related to moving instances out of Standby mode.

esrsResponseStatus - -- | The response status code.

data ExitStandbyResponse

Contains the parameters for ExitStandby.

See: exitStandbyResponse smart constructor.

Response Lenses

esrsActivities :: Lens' ExitStandbyResponse [Activity]

The activities related to moving instances out of Standby mode.

esrsResponseStatus :: Lens' ExitStandbyResponse Int

- | The response status code.