Network.AWS.AutoScaling.PutNotificationConfiguration

Description

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to this topic can have messages for events delivered to an endpoint such as a web server or email address.

For more information see Getting Notifications When Your Auto Scaling GroupChanges in the Auto Scaling Developer Guide.

This configuration overwrites an existing configuration.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutNotificationConfiguration.html

Request

data PutNotificationConfiguration

Request constructor

putNotificationConfiguration

Arguments

:: Text
pncAutoScalingGroupName

-> Text
pncTopicARN

-> PutNotificationConfiguration

PutNotificationConfiguration constructor.

The fields accessible through corresponding lenses are:

pncAutoScalingGroupName :: Text

pncNotificationTypes :: [Text]

pncTopicARN :: Text

Request lenses

pncAutoScalingGroupName :: Lens' PutNotificationConfiguration Text

The name of the Auto Scaling group.

pncNotificationTypes :: Lens' PutNotificationConfiguration [Text]

The type of event that will cause the notification to be sent. For details about notification types supported by Auto Scaling, see DescribeAutoScalingNotificationTypes.

pncTopicARN :: Lens' PutNotificationConfiguration Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

Response

data PutNotificationConfigurationResponse

Response constructor

putNotificationConfigurationResponse :: PutNotificationConfigurationResponse

PutNotificationConfigurationResponse constructor.