Network.AWS.AutoScaling.PutScheduledUpdateGroupAction

Description

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged in the affected Auto Scaling group. For information on creating or updating a scheduled action for your Auto Scaling group, see Scale Based on a Schedule. Auto Scaling supports the date and time expressed in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only. Schedule based on a specific date and time https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ScheduledActionName=ScaleUp &StartTime=2013-05-25T08:00:00Z &DesiredCapacity=3 &Version=2011-01-01 &Action=PutScheduledUpdateGroupAction &AUTHPARAMS 3bc8c9bc-6a62-11e2-8a51-4b8a1EXAMPLE Recurring Schedule https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ScheduledActionName=scaleup-schedule-year &Recurrence="30 0 1 1,6,12 *" &DesiredCapacity=3 &Version=2011-01-01 &Action=PutScheduledUpdateGroupAction &AUTHPARAMS 3bc8c9bc-6a62-11e2-8a51-4b8a1EXAMPLE.

Request

data PutScheduledUpdateGroupAction

Request constructor

putScheduledUpdateGroupAction

Arguments

:: Text
psugaAutoScalingGroupName

-> Text
psugaScheduledActionName

-> PutScheduledUpdateGroupAction

Smart constructor for the minimum required parameters to construct a valid PutScheduledUpdateGroupAction request.

The fields accessible through corresponding lenses are:

AutoScalingGroupName :: Text

ScheduledActionName :: Text

Time :: Maybe ISO8601

StartTime :: Maybe ISO8601

EndTime :: Maybe ISO8601

Recurrence :: Maybe Text

MinSize :: Maybe Integer

MaxSize :: Maybe Integer

DesiredCapacity :: Maybe Integer

Request lenses

psugaAutoScalingGroupName :: Lens' PutScheduledUpdateGroupAction Text

The name or ARN of the Auto Scaling group.

psugaScheduledActionName :: Lens' PutScheduledUpdateGroupAction Text

The name of this scaling action.

psugaTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

Time is deprecated. The time for this action to start. Time is an alias for StartTime and can be specified instead of StartTime, or vice versa. If both Time and StartTime are specified, their values should be identical. Otherwise, PutScheduledUpdateGroupAction will return an error.

psugaStartTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

The time for this action to start, as in --start-time 2010-06-01T00:00:00Z. If you try to schedule your action in the past, Auto Scaling returns an error message. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

psugaEndTime :: Lens' PutScheduledUpdateGroupAction (Maybe ISO8601)

The time for this action to end.

psugaRecurrence :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

The time when recurring future actions will start. Start time is specified by the user following the Unix cron syntax format. For information about cron syntax, go to Wikipedia, The Free Encyclopedia. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

psugaMinSize :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

The minimum size for the new Auto Scaling group.

psugaMaxSize :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

The maximum size for the Auto Scaling group.

psugaDesiredCapacity :: Lens' PutScheduledUpdateGroupAction (Maybe Integer)

The number of Amazon EC2 instances that should be running in the group.

Response

data PutScheduledUpdateGroupActionResponse

Response constructor

putScheduledUpdateGroupActionResponse :: PutScheduledUpdateGroupActionResponse

Smart constructor for the minimum required parameters to construct a valid PutScheduledUpdateGroupActionResponse response.

This constructor is provided for convenience and testing purposes.