Network.AWS.AutoScaling.RecordLifecycleActionHeartbeat

Description

Records a heartbeat for the lifecycle action associated with a specific token. This extends the timeout by the length of time defined by the HeartbeatTimeout parameter of PutLifecycleHook.

This operation is a part of the basic sequence for adding a lifecycle hook to an Auto Scaling group:

Create a notification target. A target can be either an Amazon SQS queue or an Amazon SNS topic. Create an IAM role. This role allows Auto Scaling to publish lifecycle notifications to the designated SQS queue or SNS topic. Create the lifecycle hook. You can create a hook that acts when instances launch or when instances terminate. If necessary, record the lifecycle action heartbeat to keep the instance in a pending state. Complete the lifecycle action. For more information, see Auto Scaling Pending State and Auto Scaling Terminating State in the Auto Scaling Developer Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_RecordLifecycleActionHeartbeat.html

Request

data RecordLifecycleActionHeartbeat

Request constructor

recordLifecycleActionHeartbeat

Arguments

:: Text
rlahLifecycleHookName

-> Text
rlahAutoScalingGroupName

-> Text
rlahLifecycleActionToken

-> RecordLifecycleActionHeartbeat

RecordLifecycleActionHeartbeat constructor.

The fields accessible through corresponding lenses are:

rlahAutoScalingGroupName :: Text

rlahLifecycleActionToken :: Text

rlahLifecycleHookName :: Text

Request lenses

rlahAutoScalingGroupName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the Auto Scaling group for the hook.

rlahLifecycleActionToken :: Lens' RecordLifecycleActionHeartbeat Text

A token that uniquely identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

rlahLifecycleHookName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the lifecycle hook.

Response

data RecordLifecycleActionHeartbeatResponse

Response constructor

recordLifecycleActionHeartbeatResponse :: RecordLifecycleActionHeartbeatResponse

RecordLifecycleActionHeartbeatResponse constructor.