Network.AWS.AutoScaling.ResumeProcesses

Description

Resumes the specified suspended Auto Scaling processes for the specified Auto Scaling group. To resume specific processes, use the ScalingProcesses parameter. To resume all processes, omit the ScalingProcesses parameter. For more information, see Suspend and Resume Auto Scaling Processes in the AutoScaling Developer Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_ResumeProcesses.html

Request

data ResumeProcesses

Request constructor

resumeProcesses

Arguments

:: Text
rpAutoScalingGroupName

-> ResumeProcesses

ResumeProcesses constructor.

The fields accessible through corresponding lenses are:

rpAutoScalingGroupName :: Text

rpScalingProcesses :: [Text]

Request lenses

rpAutoScalingGroupName :: Lens' ResumeProcesses Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

rpScalingProcesses :: Lens' ResumeProcesses [Text]

One or more of the following processes:

Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer

Response

data ResumeProcessesResponse

Response constructor

resumeProcessesResponse :: ResumeProcessesResponse

ResumeProcessesResponse constructor.