Network.AWS.AutoScaling.SetDesiredCapacity

Description

Sets the size of the specified Auto Scaling group.

For more information about desired capacity, see What Is Auto Scaling? in the Auto Scaling User Guide .

Creating a Request

setDesiredCapacity

Arguments

:: Text
sdcAutoScalingGroupName

-> Int
sdcDesiredCapacity

-> SetDesiredCapacity

Creates a value of SetDesiredCapacity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sdcHonorCooldown - By default, SetDesiredCapacity overrides any cooldown period associated with the Auto Scaling group. Specify True to make Auto Scaling to wait for the cool-down period associated with the Auto Scaling group to complete before initiating a scaling activity to set your Auto Scaling group to its new capacity.

sdcAutoScalingGroupName - The name of the Auto Scaling group.

sdcDesiredCapacity - The number of EC2 instances that should be running in the Auto Scaling group.

data SetDesiredCapacity

Contains the parameters for SetDesiredCapacity.

See: setDesiredCapacity smart constructor.

Request Lenses

sdcHonorCooldown :: Lens' SetDesiredCapacity (Maybe Bool)

By default, SetDesiredCapacity overrides any cooldown period associated with the Auto Scaling group. Specify True to make Auto Scaling to wait for the cool-down period associated with the Auto Scaling group to complete before initiating a scaling activity to set your Auto Scaling group to its new capacity.

sdcAutoScalingGroupName :: Lens' SetDesiredCapacity Text

The name of the Auto Scaling group.

sdcDesiredCapacity :: Lens' SetDesiredCapacity Int

The number of EC2 instances that should be running in the Auto Scaling group.

Destructuring the Response

setDesiredCapacityResponse :: SetDesiredCapacityResponse

Creates a value of SetDesiredCapacityResponse with the minimum fields required to make a request.

data SetDesiredCapacityResponse

See: setDesiredCapacityResponse smart constructor.