Network.AWS.AutoScaling.SetInstanceHealth

Description

Sets the health status of the specified instance. For more information, see Health Checks in the Auto Scaling Developer Guide.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_SetInstanceHealth.html

Request

data SetInstanceHealth

Request constructor

setInstanceHealth

Arguments

:: Text
sihInstanceId

-> Text
sihHealthStatus

-> SetInstanceHealth

SetInstanceHealth constructor.

The fields accessible through corresponding lenses are:

sihHealthStatus :: Text

sihInstanceId :: Text

sihShouldRespectGracePeriod :: Maybe Bool

Request lenses

sihHealthStatus :: Lens' SetInstanceHealth Text

The health status of the instance. Set to Healthy if you want the instance to remain in service. Set to Unhealthy if you want the instance to be out of service. Auto Scaling will terminate and replace the unhealthy instance.

sihInstanceId :: Lens' SetInstanceHealth Text

The ID of the EC2 instance.

sihShouldRespectGracePeriod :: Lens' SetInstanceHealth (Maybe Bool)

If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call will respect the grace period. Set this to False, if you do not want the call to respect the grace period associated with the group. For more information, see the HealthCheckGracePeriod parameter description for CreateAutoScalingGroup.

Response

data SetInstanceHealthResponse

Response constructor

setInstanceHealthResponse :: SetInstanceHealthResponse

SetInstanceHealthResponse constructor.