Network.AWS.AutoScaling.SuspendProcesses

Description

Suspends the specified Auto Scaling processes, or all processes, for the specified Auto Scaling group.

Note that if you suspend either the Launch or Terminate process types, it can prevent other process types from functioning properly.

To resume processes that have been suspended, use ResumeProcesses .

For more information, see Suspending and Resuming Auto Scaling Processes in the Auto Scaling User Guide .

Creating a Request

suspendProcesses

Arguments

:: Text
spAutoScalingGroupName

-> SuspendProcesses

Creates a value of SuspendProcesses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spScalingProcesses - One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

spAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the Auto Scaling group.

data SuspendProcesses

Contains the parameters for SuspendProcesses and ResumeProcesses.

See: suspendProcesses smart constructor.

Request Lenses

spScalingProcesses :: Lens' SuspendProcesses [Text]

One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

spAutoScalingGroupName :: Lens' SuspendProcesses Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

Destructuring the Response

suspendProcessesResponse :: SuspendProcessesResponse

Creates a value of SuspendProcessesResponse with the minimum fields required to make a request.

data SuspendProcessesResponse

See: suspendProcessesResponse smart constructor.