Network.AWS.AutoScaling.TerminateInstanceInAutoScalingGroup

Description

Terminates the specified instance and optionally adjusts the desired group size.

This call simply makes a termination request. The instances is not terminated immediately.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_TerminateInstanceInAutoScalingGroup.html

Request

data TerminateInstanceInAutoScalingGroup

Request constructor

terminateInstanceInAutoScalingGroup

Arguments

:: Text
tiiasgInstanceId

-> Bool
tiiasgShouldDecrementDesiredCapacity

-> TerminateInstanceInAutoScalingGroup

TerminateInstanceInAutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

tiiasgInstanceId :: Text

tiiasgShouldDecrementDesiredCapacity :: Bool

Request lenses

tiiasgInstanceId :: Lens' TerminateInstanceInAutoScalingGroup Text

The ID of the EC2 instance.

tiiasgShouldDecrementDesiredCapacity :: Lens' TerminateInstanceInAutoScalingGroup Bool

If true, terminating this instance also decrements the size of the Auto Scaling group.

Response

data TerminateInstanceInAutoScalingGroupResponse

Response constructor

terminateInstanceInAutoScalingGroupResponse :: TerminateInstanceInAutoScalingGroupResponse

TerminateInstanceInAutoScalingGroupResponse constructor.

The fields accessible through corresponding lenses are:

tiiasgrActivity :: Maybe Activity

Response lenses

tiiasgrActivity :: Lens' TerminateInstanceInAutoScalingGroupResponse (Maybe Activity)

A scaling activity.