Network.AWS.AutoScaling.Types

Service

data AutoScaling

Version 2011-01-01 of the Amazon Auto Scaling service.

Error

data RESTError :: *

TagDescription

data TagDescription

tagDescription

Arguments

:: Text
tdResourceId

-> Text
tdResourceType

-> Text
tdKey

-> Text
tdValue

-> Bool
tdPropagateAtLaunch

-> TagDescription

TagDescription constructor.

The fields accessible through corresponding lenses are:

tdKey :: Text

tdPropagateAtLaunch :: Bool

tdResourceId :: Text

tdResourceType :: Text

tdValue :: Text

tdKey :: Lens' TagDescription Text

The key of the tag.

tdPropagateAtLaunch :: Lens' TagDescription Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

tdResourceId :: Lens' TagDescription Text

The name of the Auto Scaling group.

tdResourceType :: Lens' TagDescription Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

tdValue :: Lens' TagDescription Text

The value of the tag.

Tag

data Tag

tag

Arguments

:: Text
tagResourceId

-> Text
tagResourceType

-> Text
tagKey

-> Text
tagValue

-> Bool
tagPropagateAtLaunch

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagPropagateAtLaunch :: Bool

tagResourceId :: Text

tagResourceType :: Text

tagValue :: Text

tagKey :: Lens' Tag Text

The key of the tag.

tagPropagateAtLaunch :: Lens' Tag Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

tagResourceId :: Lens' Tag Text

The name of the Auto Scaling group.

tagResourceType :: Lens' Tag Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

tagValue :: Lens' Tag Text

The value of the tag.

NotificationConfiguration

data NotificationConfiguration

notificationConfiguration :: NotificationConfiguration

NotificationConfiguration constructor.

The fields accessible through corresponding lenses are:

ncAutoScalingGroupName :: Maybe Text

ncNotificationType :: Maybe Text

ncTopicARN :: Maybe Text

ncAutoScalingGroupName :: Lens' NotificationConfiguration (Maybe Text)

Specifies the Auto Scaling group name.

ncNotificationType :: Lens' NotificationConfiguration (Maybe Text)

The types of events for an action to start.

ncTopicARN :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

BlockDeviceMapping

data BlockDeviceMapping

blockDeviceMapping

Arguments

:: Text
bdmDeviceName

-> BlockDeviceMapping

BlockDeviceMapping constructor.

The fields accessible through corresponding lenses are:

bdmDeviceName :: Text

bdmEbs :: Maybe Ebs

bdmNoDevice :: Maybe Bool

bdmVirtualName :: Maybe Text

bdmDeviceName :: Lens' BlockDeviceMapping Text

The name of the device within Amazon EC2 (for example, devsdh or xvdh).

bdmEbs :: Lens' BlockDeviceMapping (Maybe Ebs)

The Elastic Block Storage volume information.

bdmNoDevice :: Lens' BlockDeviceMapping (Maybe Bool)

Suppresses the device mapping.

bdmVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

The virtual name associated with the device.

LaunchConfiguration

data LaunchConfiguration

launchConfiguration

Arguments

:: Text
lcLaunchConfigurationName

-> Text
lcImageId

-> Text
lcInstanceType

-> UTCTime
lcCreatedTime

-> LaunchConfiguration

LaunchConfiguration constructor.

The fields accessible through corresponding lenses are:

lcAssociatePublicIpAddress :: Maybe Bool

lcBlockDeviceMappings :: [BlockDeviceMapping]

lcCreatedTime :: UTCTime

lcEbsOptimized :: Maybe Bool

lcIamInstanceProfile :: Maybe Text

lcImageId :: Text

lcInstanceMonitoring :: Maybe InstanceMonitoring

lcInstanceType :: Text

lcKernelId :: Maybe Text

lcKeyName :: Maybe Text

lcLaunchConfigurationARN :: Maybe Text

lcLaunchConfigurationName :: Text

lcPlacementTenancy :: Maybe Text

lcRamdiskId :: Maybe Text

lcSecurityGroups :: [Text]

lcSpotPrice :: Maybe Text

lcUserData :: Maybe Text

lcAssociatePublicIpAddress :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the instance is associated with a public IP address (true) or not (false).

lcBlockDeviceMappings :: Lens' LaunchConfiguration [BlockDeviceMapping]

Specifies how block devices are exposed to the instance. Each mapping is made up of a virtualName and a deviceName.

lcCreatedTime :: Lens' LaunchConfiguration UTCTime

Provides the creation date and time for this launch configuration.

lcEbsOptimized :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the instance is optimized for EBS I/O (true) or not (false).

lcIamInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

Provides the name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

lcImageId :: Lens' LaunchConfiguration Text

Provides the unique ID of the Amazon Machine Image (AMI) that was assigned during registration.

lcInstanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed monitoring or not.

lcInstanceType :: Lens' LaunchConfiguration Text

Specifies the instance type of the Amazon EC2 instance.

lcKernelId :: Lens' LaunchConfiguration (Maybe Text)

Provides the ID of the kernel associated with the Amazon EC2 AMI.

lcKeyName :: Lens' LaunchConfiguration (Maybe Text)

Provides the name of the Amazon EC2 key pair.

lcLaunchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

The launch configuration's Amazon Resource Name (ARN).

lcLaunchConfigurationName :: Lens' LaunchConfiguration Text

Specifies the name of the launch configuration.

lcPlacementTenancy :: Lens' LaunchConfiguration (Maybe Text)

Specifies the tenancy of the instance. It can be either default or dedicated. An instance with dedicated tenancy runs in an isolated, single-tenant hardware and it can only be launched in a VPC.

lcRamdiskId :: Lens' LaunchConfiguration (Maybe Text)

Provides ID of the RAM disk associated with the Amazon EC2 AMI.

lcSecurityGroups :: Lens' LaunchConfiguration [Text]

A description of the security groups to associate with the Amazon EC2 instances.

lcSpotPrice :: Lens' LaunchConfiguration (Maybe Text)

Specifies the price to bid when launching Spot Instances.

lcUserData :: Lens' LaunchConfiguration (Maybe Text)

The user data available to the launched Amazon EC2 instances.

AutoScalingGroup

data AutoScalingGroup

autoScalingGroup

Arguments

:: Text
asgAutoScalingGroupName

-> Text
asgLaunchConfigurationName

-> Int
asgMinSize

-> Int
asgMaxSize

-> Int
asgDesiredCapacity

-> Int
asgDefaultCooldown

-> NonEmpty Text
asgAvailabilityZones

-> Text
asgHealthCheckType

-> UTCTime
asgCreatedTime

-> AutoScalingGroup

AutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

asgAutoScalingGroupARN :: Maybe Text

asgAutoScalingGroupName :: Text

asgAvailabilityZones :: NonEmpty Text

asgCreatedTime :: UTCTime

asgDefaultCooldown :: Int

asgDesiredCapacity :: Int

asgEnabledMetrics :: [EnabledMetric]

asgHealthCheckGracePeriod :: Maybe Int

asgHealthCheckType :: Text

asgInstances :: [Instance]

asgLaunchConfigurationName :: Text

asgLoadBalancerNames :: [Text]

asgMaxSize :: Int

asgMinSize :: Int

asgPlacementGroup :: Maybe Text

asgStatus :: Maybe Text

asgSuspendedProcesses :: [SuspendedProcess]

asgTags :: [TagDescription]

asgTerminationPolicies :: [Text]

asgVPCZoneIdentifier :: Maybe Text

asgAutoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Auto Scaling group.

asgAutoScalingGroupName :: Lens' AutoScalingGroup Text

Specifies the name of the group.

asgAvailabilityZones :: Lens' AutoScalingGroup (NonEmpty Text)

Contains a list of Availability Zones for the group.

asgCreatedTime :: Lens' AutoScalingGroup UTCTime

Specifies the date and time the Auto Scaling group was created.

asgDefaultCooldown :: Lens' AutoScalingGroup Int

The number of seconds after a scaling activity completes before any further scaling activities can start.

asgDesiredCapacity :: Lens' AutoScalingGroup Int

Specifies the desired capacity for the Auto Scaling group.

asgEnabledMetrics :: Lens' AutoScalingGroup [EnabledMetric]

A list of metrics enabled for this Auto Scaling group.

asgHealthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Int)

The length of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

asgHealthCheckType :: Lens' AutoScalingGroup Text

The service of interest for the health status check, either EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

asgInstances :: Lens' AutoScalingGroup [Instance]

Provides a summary list of Amazon EC2 instances.

asgLaunchConfigurationName :: Lens' AutoScalingGroup Text

Specifies the name of the associated LaunchConfiguration.

asgLoadBalancerNames :: Lens' AutoScalingGroup [Text]

A list of load balancers associated with this Auto Scaling group.

asgMaxSize :: Lens' AutoScalingGroup Int

Contains the maximum size of the Auto Scaling group.

asgMinSize :: Lens' AutoScalingGroup Int

Contains the minimum size of the Auto Scaling group.

asgPlacementGroup :: Lens' AutoScalingGroup (Maybe Text)

The name of the cluster placement group, if applicable. For more information, go to Using Cluster Instances in the Amazon EC2 User Guide.

asgStatus :: Lens' AutoScalingGroup (Maybe Text)

The current state of the Auto Scaling group when a DeleteAutoScalingGroup action is in progress.

asgSuspendedProcesses :: Lens' AutoScalingGroup [SuspendedProcess]

Suspended processes associated with this Auto Scaling group.

asgTags :: Lens' AutoScalingGroup [TagDescription]

A list of tags for the Auto Scaling group.

asgTerminationPolicies :: Lens' AutoScalingGroup [Text]

A standalone termination policy or a list of termination policies for this Auto Scaling group.

asgVPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list. When you specify VPCZoneIdentifier with AvailabilityZones, ensure that the subnets' Availability Zones match the values you specify for AvailabilityZones.

ScalingPolicy

data ScalingPolicy

scalingPolicy :: ScalingPolicy

ScalingPolicy constructor.

The fields accessible through corresponding lenses are:

sp1AdjustmentType :: Maybe Text

sp1Alarms :: [Alarm]

sp1AutoScalingGroupName :: Maybe Text

sp1Cooldown :: Maybe Int

sp1MinAdjustmentStep :: Maybe Int

sp1PolicyARN :: Maybe Text

sp1PolicyName :: Maybe Text

sp1ScalingAdjustment :: Maybe Int

sp1AdjustmentType :: Lens' ScalingPolicy (Maybe Text)

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

sp1Alarms :: Lens' ScalingPolicy [Alarm]

A list of CloudWatch Alarms related to the policy.

sp1AutoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

The name of the Auto Scaling group associated with this scaling policy.

sp1Cooldown :: Lens' ScalingPolicy (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sp1MinAdjustmentStep :: Lens' ScalingPolicy (Maybe Int)

Changes the DesiredCapacity of the Auto Scaling group by at least the specified number of instances.

sp1PolicyARN :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

sp1PolicyName :: Lens' ScalingPolicy (Maybe Text)

The name of the scaling policy.

sp1ScalingAdjustment :: Lens' ScalingPolicy (Maybe Int)

The number associated with the specified adjustment type. A positive value adds to the current capacity and a negative value removes from the current capacity.

InstanceMonitoring

data InstanceMonitoring

instanceMonitoring :: InstanceMonitoring

InstanceMonitoring constructor.

The fields accessible through corresponding lenses are:

imEnabled :: Maybe Bool

imEnabled :: Lens' InstanceMonitoring (Maybe Bool)

If True, instance monitoring is enabled.

ScheduledUpdateGroupAction

data ScheduledUpdateGroupAction

scheduledUpdateGroupAction :: ScheduledUpdateGroupAction

ScheduledUpdateGroupAction constructor.

The fields accessible through corresponding lenses are:

sugaAutoScalingGroupName :: Maybe Text

sugaDesiredCapacity :: Maybe Int

sugaEndTime :: Maybe UTCTime

sugaMaxSize :: Maybe Int

sugaMinSize :: Maybe Int

sugaRecurrence :: Maybe Text

sugaScheduledActionARN :: Maybe Text

sugaScheduledActionName :: Maybe Text

sugaStartTime :: Maybe UTCTime

sugaTime :: Maybe UTCTime

sugaAutoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the Auto Scaling group to be updated.

sugaDesiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The number of instances you prefer to maintain in your Auto Scaling group.

sugaEndTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The time that the action is scheduled to end. This value can be up to one month in the future.

sugaMaxSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The maximum size of the Auto Scaling group.

sugaMinSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The minimum size of the Auto Scaling group.

sugaRecurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The regular schedule that an action occurs.

sugaScheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The Amazon Resource Name (ARN) of this scheduled action.

sugaScheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of this scheduled action.

sugaStartTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The time that the action is scheduled to begin. This value can be up to one month in the future. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

sugaTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

Time is deprecated. The time that the action is scheduled to begin. Time is an alias for StartTime.

ScalingProcessQuery

data ScalingProcessQuery

scalingProcessQuery

Arguments

:: Text
spqAutoScalingGroupName

-> ScalingProcessQuery

ScalingProcessQuery constructor.

The fields accessible through corresponding lenses are:

spqAutoScalingGroupName :: Text

spqScalingProcesses :: [Text]

spqAutoScalingGroupName :: Lens' ScalingProcessQuery Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

spqScalingProcesses :: Lens' ScalingProcessQuery [Text]

The processes that you want to suspend or resume, which can include one or more of the following: Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer To suspend all process types, omit this parameter.

Ebs

data Ebs

ebs :: Ebs

Ebs constructor.

The fields accessible through corresponding lenses are:

ebsDeleteOnTermination :: Maybe Bool

ebsIops :: Maybe Natural

ebsSnapshotId :: Maybe Text

ebsVolumeSize :: Maybe Natural

ebsVolumeType :: Maybe Text

ebsDeleteOnTermination :: Lens' Ebs (Maybe Bool)

Indicates whether to delete the volume on instance termination. Default: true.

ebsIops :: Lens' Ebs (Maybe Natural)

The number of I/O operations per second (IOPS) that the volume supports. The maximum ratio of IOPS to volume size is 30.0 Valid Values: Range is 100 to 4000. Default: None.

ebsSnapshotId :: Lens' Ebs (Maybe Text)

The snapshot ID.

ebsVolumeSize :: Lens' Ebs (Maybe Natural)

The volume size, in gigabytes. Valid values: If the volume type is io1, the minimum size of the volume is 10. Default: If you're creating the volume from a snapshot, and you don't specify a volume size, the default is the snapshot size. Required: Required when the volume type is io1.

ebsVolumeType :: Lens' Ebs (Maybe Text)

The volume type. Valid values: standard | io1 Default: standard.

AdjustmentType

data AdjustmentType

adjustmentType :: AdjustmentType

AdjustmentType constructor.

The fields accessible through corresponding lenses are:

atAdjustmentType :: Maybe Text

atAdjustmentType :: Lens' AdjustmentType (Maybe Text)

A policy adjustment type. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

MetricCollectionType

data MetricCollectionType

metricCollectionType :: MetricCollectionType

MetricCollectionType constructor.

The fields accessible through corresponding lenses are:

mctMetric :: Maybe Text

mctMetric :: Lens' MetricCollectionType (Maybe Text)

LifecycleHook

data LifecycleHook

lifecycleHook :: LifecycleHook

LifecycleHook constructor.

The fields accessible through corresponding lenses are:

lhAutoScalingGroupName :: Maybe Text

lhDefaultResult :: Maybe Text

lhGlobalTimeout :: Maybe Int

lhHeartbeatTimeout :: Maybe Int

lhLifecycleHookName :: Maybe Text

lhLifecycleTransition :: Maybe Text

lhNotificationMetadata :: Maybe Text

lhNotificationTargetARN :: Maybe Text

lhRoleARN :: Maybe Text

lhAutoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

The name of the Auto Scaling group to which the lifecycle action belongs.

lhDefaultResult :: Lens' LifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The value for this parameter can be either CONTINUE or ABANDON. The default value for this parameter is CONTINUE.

lhGlobalTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum length of time an instance can remain in a Pending:Wait or Terminating:Wait state. Currently, this value is set at 48 hours.

lhHeartbeatTimeout :: Lens' LifecycleHook (Maybe Int)

Defines the amount of time that can elapse before the lifecycle hook times out. When the lifecycle hook times out, Auto Scaling performs the action defined in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

lhLifecycleHookName :: Lens' LifecycleHook (Maybe Text)

The name of the lifecycle action hook.

lhLifecycleTransition :: Lens' LifecycleHook (Maybe Text)

The Amazon EC2 instance state to which you want to attach the lifecycle hook. See DescribeLifecycleHooks for a list of available lifecycle hook types.

lhNotificationMetadata :: Lens' LifecycleHook (Maybe Text)

Contains additional information that you want to include any time Auto Scaling sends a message to the notification target.

lhNotificationTargetARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the notification target that Auto Scaling will use to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target will include: Lifecycle action token User account ID Name of the Auto Scaling group Lifecycle hook name EC2 instance ID Lifecycle transition Notification metadata.

lhRoleARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the Amazon IAM role that allows the Auto Scaling group to publish to the specified notification target.

Activity

data Activity

activity

Arguments

:: Text
aActivityId

-> Text
aAutoScalingGroupName

-> Text
aCause

-> UTCTime
aStartTime

-> Text
aStatusCode

-> Activity

Activity constructor.

The fields accessible through corresponding lenses are:

aActivityId :: Text

aAutoScalingGroupName :: Text

aCause :: Text

aDescription :: Maybe Text

aDetails :: Maybe Text

aEndTime :: Maybe UTCTime

aProgress :: Maybe Int

aStartTime :: UTCTime

aStatusCode :: Text

aStatusMessage :: Maybe Text

aActivityId :: Lens' Activity Text

Specifies the ID of the activity.

aAutoScalingGroupName :: Lens' Activity Text

The name of the Auto Scaling group.

aCause :: Lens' Activity Text

Contains the reason the activity was begun.

aDescription :: Lens' Activity (Maybe Text)

Contains a friendly, more verbose description of the scaling activity.

aDetails :: Lens' Activity (Maybe Text)

Contains details of the scaling activity.

aEndTime :: Lens' Activity (Maybe UTCTime)

Provides the end time of this activity.

aProgress :: Lens' Activity (Maybe Int)

Specifies a value between 0 and 100 that indicates the progress of the activity.

aStartTime :: Lens' Activity UTCTime

Provides the start time of this activity.

aStatusCode :: Lens' Activity Text

Contains the current status of the activity.

aStatusMessage :: Lens' Activity (Maybe Text)

Contains a friendly, more verbose description of the activity status.

SuspendedProcess

data SuspendedProcess

suspendedProcess :: SuspendedProcess

SuspendedProcess constructor.

The fields accessible through corresponding lenses are:

spProcessName :: Maybe Text

spSuspensionReason :: Maybe Text

spProcessName :: Lens' SuspendedProcess (Maybe Text)

The name of the suspended process.

spSuspensionReason :: Lens' SuspendedProcess (Maybe Text)

The reason that the process was suspended.

MetricGranularityType

data MetricGranularityType

metricGranularityType :: MetricGranularityType

MetricGranularityType constructor.

The fields accessible through corresponding lenses are:

mgtGranularity :: Maybe Text

mgtGranularity :: Lens' MetricGranularityType (Maybe Text)

The granularity of a Metric.

Filter

data Filter

filter'

Arguments

:: Text
fName

-> Filter

Filter constructor.

The fields accessible through corresponding lenses are:

fName :: Text

fValues :: [Text]

fName :: Lens' Filter Text

The name of the filter. Valid Name values are: "auto-scaling-group", "key", "value", and "propagate-at-launch".

fValues :: Lens' Filter [Text]

The value of the filter.

ProcessType

data ProcessType

processType

Arguments

:: Text
ptProcessName

-> ProcessType

ProcessType constructor.

The fields accessible through corresponding lenses are:

ptProcessName :: Text

ptProcessName :: Lens' ProcessType Text

The name of a process.

Alarm

data Alarm

alarm :: Alarm

Alarm constructor.

The fields accessible through corresponding lenses are:

aAlarmARN :: Maybe Text

aAlarmName :: Maybe Text

aAlarmARN :: Lens' Alarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

aAlarmName :: Lens' Alarm (Maybe Text)

The name of the alarm.

EnabledMetric

data EnabledMetric

enabledMetric :: EnabledMetric

EnabledMetric constructor.

The fields accessible through corresponding lenses are:

emGranularity :: Maybe Text

emMetric :: Maybe Text

emGranularity :: Lens' EnabledMetric (Maybe Text)

The granularity of the enabled metric.

emMetric :: Lens' EnabledMetric (Maybe Text)

The name of the enabled metric.

Instance

data Instance

instance'

Arguments

:: Text
iInstanceId

-> Text
iAvailabilityZone

-> Text
iLifecycleState

-> Text
iHealthStatus

-> Text
iLaunchConfigurationName

-> Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iAvailabilityZone :: Text

iHealthStatus :: Text

iInstanceId :: Text

iLaunchConfigurationName :: Text

iLifecycleState :: Text

iAvailabilityZone :: Lens' Instance Text

Availability Zones associated with this instance.

iHealthStatus :: Lens' Instance Text

The instance's health status.

iInstanceId :: Lens' Instance Text

Specifies the ID of the Amazon EC2 instance.

iLaunchConfigurationName :: Lens' Instance Text

The launch configuration associated with this instance.

iLifecycleState :: Lens' Instance Text

Contains a description of the current lifecycle state.

LifecycleState

data LifecycleState

Constructors

Detached
Detached

Detaching
Detaching

EnteringStandby
EnteringStandby

InService
InService

Pending
Pending

PendingProceed
Pending:Proceed

PendingWait
Pending:Wait

Quarantined
Quarantined

Standby
Standby

Terminated
Terminated

Terminating
Terminating

TerminatingProceed
Terminating:Proceed

TerminatingWait
Terminating:Wait

AutoScalingInstanceDetails

data AutoScalingInstanceDetails

autoScalingInstanceDetails

Arguments

:: Text
asidInstanceId

-> Text
asidAutoScalingGroupName

-> Text
asidAvailabilityZone

-> Text
asidLifecycleState

-> Text
asidHealthStatus

-> Text
asidLaunchConfigurationName

-> AutoScalingInstanceDetails

AutoScalingInstanceDetails constructor.

The fields accessible through corresponding lenses are:

asidAutoScalingGroupName :: Text

asidAvailabilityZone :: Text

asidHealthStatus :: Text

asidInstanceId :: Text

asidLaunchConfigurationName :: Text

asidLifecycleState :: Text

asidAutoScalingGroupName :: Lens' AutoScalingInstanceDetails Text

The name of the Auto Scaling group associated with this instance.

asidAvailabilityZone :: Lens' AutoScalingInstanceDetails Text

The Availability Zone in which this instance resides.

asidHealthStatus :: Lens' AutoScalingInstanceDetails Text

The health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy. Auto Scaling should terminate and replace it.

asidInstanceId :: Lens' AutoScalingInstanceDetails Text

The instance ID of the Amazon EC2 instance.

asidLaunchConfigurationName :: Lens' AutoScalingInstanceDetails Text

The launch configuration associated with this instance.

asidLifecycleState :: Lens' AutoScalingInstanceDetails Text

The life cycle state of this instance. for more information, see Instance Lifecycle State in the Auto Scaling Developer Guide.

ScalingActivityStatusCode

data ScalingActivityStatusCode

Constructors

Cancelled
Cancelled

Failed
Failed

InProgress
InProgress

MidLifecycleAction
MidLifecycleAction

PreInService
PreInService

Successful
Successful

WaitingForELBConnectionDraining
WaitingForELBConnectionDraining

WaitingForInstanceId
WaitingForInstanceId

WaitingForSpotInstanceId
WaitingForSpotInstanceId

WaitingForSpotInstanceRequestId
WaitingForSpotInstanceRequestId