Network.AWS.AutoScaling.UpdateAutoScalingGroup

Description

Updates the configuration for the specified AutoScalingGroup.

To update an Auto Scaling group with a launch configuration that has the InstanceMonitoring flag set to False, you must first ensure that collection of group metrics is disabled. Otherwise, calls to UpdateAutoScalingGroup will fail. If you have previously enabled group metrics collection, you can disable collection of all group metrics by calling DisableMetricsCollection.

The new settings are registered upon the completion of this call. Any launch configuration settings take effect on any triggers after this call returns. Scaling activities that are currently in progress aren't affected.

If a new value is specified for MinSize without specifying the value for DesiredCapacity, and if the new MinSize is larger than the current size of the Auto Scaling group, there will be an implicit call to SetDesiredCapacity to set the group to the new MinSize.

If a new value is specified for MaxSize without specifying the value for DesiredCapacity, and the new MaxSize is smaller than the current size of the Auto Scaling group, there will be an implicit call to SetDesiredCapacity to set the group to the new MaxSize.

All other optional parameters are left unchanged if not passed in the request.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_UpdateAutoScalingGroup.html

Request

data UpdateAutoScalingGroup

Request constructor

updateAutoScalingGroup

Arguments

:: Text
uasgAutoScalingGroupName

-> NonEmpty Text
uasgAvailabilityZones

-> UpdateAutoScalingGroup

UpdateAutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

uasgAutoScalingGroupName :: Text

uasgAvailabilityZones :: NonEmpty Text

uasgDefaultCooldown :: Maybe Int

uasgDesiredCapacity :: Maybe Int

uasgHealthCheckGracePeriod :: Maybe Int

uasgHealthCheckType :: Maybe Text

uasgLaunchConfigurationName :: Maybe Text

uasgMaxSize :: Maybe Int

uasgMinSize :: Maybe Int

uasgPlacementGroup :: Maybe Text

uasgTerminationPolicies :: [Text]

uasgVPCZoneIdentifier :: Maybe Text

Request lenses

uasgAutoScalingGroupName :: Lens' UpdateAutoScalingGroup Text

The name of the Auto Scaling group.

uasgAvailabilityZones :: Lens' UpdateAutoScalingGroup (NonEmpty Text)

One or more Availability Zones for the group.

uasgDefaultCooldown :: Lens' UpdateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. For more information, see UnderstandingAuto Scaling Cooldowns.

uasgDesiredCapacity :: Lens' UpdateAutoScalingGroup (Maybe Int)

The number of EC2 instances that should be running in the Auto Scaling group. This value must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group.

uasgHealthCheckGracePeriod :: Lens' UpdateAutoScalingGroup (Maybe Int)

The amount of time, in second, that Auto Scaling waits before checking the health status of an instance. The grace period begins when the instance passes System Status and the Instance Status checks from Amazon EC2. For more information, see DescribeInstanceStatus.

uasgHealthCheckType :: Lens' UpdateAutoScalingGroup (Maybe Text)

The type of health check for the instances in the Auto Scaling group. The health check type can either be EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

uasgLaunchConfigurationName :: Lens' UpdateAutoScalingGroup (Maybe Text)

The name of the launch configuration.

uasgMaxSize :: Lens' UpdateAutoScalingGroup (Maybe Int)

The maximum size of the Auto Scaling group.

uasgMinSize :: Lens' UpdateAutoScalingGroup (Maybe Int)

The minimum size of the Auto Scaling group.

uasgPlacementGroup :: Lens' UpdateAutoScalingGroup (Maybe Text)

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups.

uasgTerminationPolicies :: Lens' UpdateAutoScalingGroup [Text]

A standalone termination policy or a list of termination policies used to select the instance to terminate. The policies are executed in the order that they are listed.

For more information, see Choosing a Termination Policy for Your AutoScaling Group in the Auto Scaling Developer Guide.

uasgVPCZoneIdentifier :: Lens' UpdateAutoScalingGroup (Maybe Text)

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list.

When you specify VPCZoneIdentifier with AvailabilityZones, ensure that the subnets' Availability Zones match the values you specify for AvailabilityZones .

For more information, see Auto Scaling and Amazon VPC in the Auto ScalingDeveloper Guide.

Response

data UpdateAutoScalingGroupResponse

Response constructor

updateAutoScalingGroupResponse :: UpdateAutoScalingGroupResponse

UpdateAutoScalingGroupResponse constructor.