Network.AWS.AutoScaling.V2011_01_01.DeleteAutoScalingGroup

Description

Deletes the specified Auto Scaling group if the group has no instances and no scaling activities in progress. To remove all instances before calling DeleteAutoScalingGroup, you can call UpdateAutoScalingGroup to set the minimum and maximum size of the AutoScalingGroup to zero. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ForceDelete=true &Version=2011-01-01 &Action=DeleteAutoScalingGroup &AUTHPARAMS 70a76d42-9665-11e2-9fdf-211deEXAMPLE.

Documentation

deleteAutoScalingGroup

Arguments

:: Text
_dasgtAutoScalingGroupName

-> DeleteAutoScalingGroup

Minimum specification for a DeleteAutoScalingGroup request.

data DeleteAutoScalingGroup

Constructors

DeleteAutoScalingGroup

Fields

_dasgtAutoScalingGroupName :: Text

The name of the Auto Scaling group to delete.

_dasgtForceDelete :: Maybe Bool

Starting with API version 2011-01-01, specifies that the Auto Scaling group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any lifecycle actions associated with the group.

dasgtForceDelete :: Lens' DeleteAutoScalingGroup (Maybe Bool)

dasgtAutoScalingGroupName :: Lens' DeleteAutoScalingGroup Text

data DeleteAutoScalingGroupResponse

Constructors

DeleteAutoScalingGroupResponse