Network.AWS.AutoScaling.V2011_01_01.DeleteLifecycleHook

Description

Deletes the specified lifecycle hook. If there are any outstanding lifecycle actions, they are completed first (ABANDON for launching instances, CONTINUE for terminating instances).

Documentation

data DeleteLifecycleHook

Constructors

DeleteLifecycleHook

Fields

_dlhuLifecycleHookName :: Text

The name of the lifecycle hook.

_dlhuAutoScalingGroupName :: Text

The name of the Auto Scaling group to which the lifecycle hook belongs.

dlhuLifecycleHookName :: Lens' DeleteLifecycleHook Text

dlhuAutoScalingGroupName :: Lens' DeleteLifecycleHook Text

data DeleteLifecycleHookResponse

Constructors

DeleteLifecycleHookResponse