Network.AWS.AutoScaling.V2011_01_01.DeleteTags

Description

Removes the specified tags or a set of tags from a set of resources.

Request

data DeleteTags

Request constructor

mkDeleteTags

Arguments

:: [Tag]
dtTags

-> DeleteTags

Smart constructor for the minimum required parameters to construct a valid DeleteTags request.

Request lenses

dtTags :: Lens' DeleteTags [Tag]

Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values are: Resource type = auto-scaling-group, Resource ID = AutoScalingGroupName, key=value, value=value, propagate=true or false.

Response

data DeleteTagsResponse

Response constructor

mkDeleteTagsResponse :: DeleteTagsResponse

Smart constructor for the minimum required parameters to construct a valid DeleteTagsResponse response.

This constructor is provided for convenience and testing purposes.