Network.AWS.AutoScaling.V2011_01_01.DescribeAccountLimits

Description

Returns the limits for the Auto Scaling resources currently allowed for your AWS account. Your AWS account comes with default limits on resources for Auto Scaling. There is a default limit of 20 Auto Scaling groups and 100 launch configurations per region. If you reach the limits for the number of Auto Scaling groups or the launch configurations, you can go to the Support Center and place a request to raise the limits. https:/autoscaling.amazonaws.com?Version=2011-01-01 &Action=DescribeAccountLimits &AUTHPARAMS 100 20 a32bd184-519d-11e3-a8a4-c1c467cbcc3b.

Documentation

data DescribeAccountLimits

Constructors

DescribeAccountLimits

data DescribeAccountLimitsResponse

Constructors

DescribeAccountLimitsResponse

Fields

_dalaMaxNumberOfAutoScalingGroups :: Maybe Integer

The maximum number of Auto Scaling groups allowed for your AWS account.

_dalaMaxNumberOfLaunchConfigurations :: Maybe Integer

The maximum number of launch configurations allowed for your AWS account.

dalaMaxNumberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Integer)

dalaMaxNumberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Integer)