Network.AWS.AutoScaling.V2011_01_01.DescribeAutoScalingGroups

Description

Returns a full description of each Auto Scaling group in the given list. This includes all Amazon EC2 instances that are members of the group. If a list of names is not provided, the service returns the full details of all Auto Scaling groups. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter. https:/autoscaling.amazonaws.com?AutoScalingGroupNames.member.1=my-test-asg-lbs &MaxRecords=20 &Version=2011-01-01 &Action=DescribeAutoScalingGroups &AUTHPARAMS my-test-asg-lbs ELB 2013-05-06T17:47:15.107Z my-test-lc 2 us-east-1b us-east-1a my-test-asg-loadbalancer 2 120 300 arn:aws:autoscaling:us-east-1:803981987763:autoScalingGroup:ca861182-c8f9-4ca7-b1eb-cd35505f5ebb :autoScalingGroupName/my-test-asg-lbs Default 10 0f02a07d-b677-11e2-9eb0-dd50EXAMPLE.

Documentation

describeAutoScalingGroups :: DescribeAutoScalingGroups

Minimum specification for a DescribeAutoScalingGroups request.

data DescribeAutoScalingGroups

Constructors

DescribeAutoScalingGroups

Fields

_asgntAutoScalingGroupNames :: [Text]

A list of Auto Scaling group names.

_asgntMaxRecords :: Maybe Integer

The maximum number of records to return.

_asgntNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

asgntNextToken :: Lens' DescribeAutoScalingGroups (Maybe Text)

asgntMaxRecords :: Lens' DescribeAutoScalingGroups (Maybe Integer)

asgntAutoScalingGroupNames :: Lens' DescribeAutoScalingGroups [Text]

data DescribeAutoScalingGroupsResponse

Constructors

DescribeAutoScalingGroupsResponse

Fields

_asgtAutoScalingGroups :: [AutoScalingGroup]

A list of Auto Scaling groups.

_asgtNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

asgtNextToken :: Lens' DescribeAutoScalingGroupsResponse (Maybe Text)

asgtAutoScalingGroups :: Lens' DescribeAutoScalingGroupsResponse [AutoScalingGroup]