Network.AWS.AutoScaling.V2011_01_01.DescribeAutoScalingInstances

Description

Returns a description of each Auto Scaling instance in the InstanceIds list. If a list is not provided, the service returns the full details of all instances up to a maximum of 50. By default, the service returns a list of 20 items. This action supports pagination by returning a token if there are more pages to retrieve. To get the next page, call this action again with the returned token as the NextToken parameter. https:/autoscaling.amazonaws.com?MaxRecords=20 &InstanceIds.member.1=i-78e0d40b &Version=2011-01-01 &Action=DescribeAutoScalingInstances &AUTHPARAMS Healthy my-test-asg us-east-1e i-78e0d40b my-test-lc InService df992dc3-b72f-11e2-81e1-750aa6EXAMPLE.

Documentation

describeAutoScalingInstances :: DescribeAutoScalingInstances

Minimum specification for a DescribeAutoScalingInstances request.

data DescribeAutoScalingInstances

Constructors

DescribeAutoScalingInstances

Fields

_dasitInstanceIds :: [Text]

The list of Auto Scaling instances to describe. If this list is omitted, all auto scaling instances are described. The list of requested instances cannot contain more than 50 items. If unknown instances are requested, they are ignored with no error.

_dasitMaxRecords :: Maybe Integer

The maximum number of Auto Scaling instances to be described with each call.

_dasitNextToken :: Maybe Text

The token returned by a previous call to indicate that there is more data available.

dasitNextToken :: Lens' DescribeAutoScalingInstances (Maybe Text)

dasitMaxRecords :: Lens' DescribeAutoScalingInstances (Maybe Integer)

dasitInstanceIds :: Lens' DescribeAutoScalingInstances [Text]

data DescribeAutoScalingInstancesResponse

Constructors

DescribeAutoScalingInstancesResponse

Fields

_asitAutoScalingInstances :: [AutoScalingInstanceDetails]

A list of Auto Scaling instances.

_asitNextToken :: Maybe Text

A string that marks the start of the next batch of returned results.

asitNextToken :: Lens' DescribeAutoScalingInstancesResponse (Maybe Text)

asitAutoScalingInstances :: Lens' DescribeAutoScalingInstancesResponse [AutoScalingInstanceDetails]