Network.AWS.AutoScaling.V2011_01_01.DescribeMetricCollectionTypes

Description

Returns a list of metrics and a corresponding list of granularities for each metric. https:/autoscaling.amazonaws.com?Version=2011-01-01&Action=DescribeMetricCollectionTypes &AUTHPARAMS oc2011-01-01"> GroupMinSize GroupMaxSize GroupDesiredCapacity GroupInServiceInstances GroupPendingInstances GroupStandyInstances GroupTerminatingInstances GroupTotalInstances 1Minute 07f3fea2-bf3c-11e2-9b6f-f3cdbb80c073 The GroupStandbyInstances metric is not returned by default. You must explicitly request it when calling EnableMetricsCollection.

Request

data DescribeMetricCollectionTypes

Request constructor

mkDescribeMetricCollectionTypes :: DescribeMetricCollectionTypes

Smart constructor for the minimum required parameters to construct a valid DescribeMetricCollectionTypes request.

Response

data DescribeMetricCollectionTypesResponse

The output of the DescribeMetricCollectionTypes action.

Response constructor

mkDescribeMetricCollectionTypesResponse :: DescribeMetricCollectionTypesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeMetricCollectionTypesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Metrics :: [MetricCollectionType]

Granularities :: [MetricGranularityType]

Response lenses

dmctrMetrics :: Lens' DescribeMetricCollectionTypesResponse [MetricCollectionType]

The list of Metrics collected. The following metrics are supported: GroupMinSize GroupMaxSize GroupDesiredCapacity GroupInServiceInstances GroupPendingInstances GroupStandbyInstances GroupTerminatingInstances GroupTotalInstances The GroupStandbyInstances metric is not returned by default. You must explicitly request it when calling EnableMetricsCollection.

dmctrGranularities :: Lens' DescribeMetricCollectionTypesResponse [MetricGranularityType]

A list of granularities for the listed Metrics.