Network.AWS.AutoScaling.V2011_01_01.DescribeNotificationConfigurations

Description

Returns a list of notification actions associated with Auto Scaling groups for specified events.

Request

data DescribeNotificationConfigurations

Request constructor

mkDescribeNotificationConfigurations :: DescribeNotificationConfigurations

Smart constructor for the minimum required parameters to construct a valid DescribeNotificationConfigurations request.

The fields accessible through corresponding lenses are:

AutoScalingGroupNames :: [Text]

NextToken :: Maybe Text

MaxRecords :: Maybe Integer

Request lenses

dnc1AutoScalingGroupNames :: Lens' DescribeNotificationConfigurations [Text]

The name of the Auto Scaling group.

dnc1NextToken :: Lens' DescribeNotificationConfigurations (Maybe Text)

A string that is used to mark the start of the next batch of returned results for pagination.

dnc1MaxRecords :: Lens' DescribeNotificationConfigurations (Maybe Integer)

Maximum number of records to be returned.

Response

data DescribeNotificationConfigurationsResponse

The output of the DescribeNotificationConfigurations action.

Response constructor

mkDescribeNotificationConfigurationsResponse

Arguments

:: [NotificationConfiguration]
dncrNotificationConfigurations

-> DescribeNotificationConfigurationsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeNotificationConfigurationsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

NotificationConfigurations :: [NotificationConfiguration]

NextToken :: Maybe Text

Response lenses

dncrNotificationConfigurations :: Lens' DescribeNotificationConfigurationsResponse [NotificationConfiguration]

The list of notification configurations.

dncrNextToken :: Lens' DescribeNotificationConfigurationsResponse (Maybe Text)

A string that is used to mark the start of the next batch of returned results for pagination.