Network.AWS.AutoScaling.V2011_01_01.DescribeScheduledActions

Description

Lists all the actions scheduled for your Auto Scaling group that haven't been executed. To see a list of actions already executed, see the activity record returned in DescribeScalingActivities.

Request

data DescribeScheduledActions

Request constructor

mkDescribeScheduledActions :: DescribeScheduledActions

Smart constructor for the minimum required parameters to construct a valid DescribeScheduledActions request.

The fields accessible through corresponding lenses are:

AutoScalingGroupName :: Maybe Text

ScheduledActionNames :: [Text]

StartTime :: Maybe ISO8601

EndTime :: Maybe ISO8601

NextToken :: Maybe Text

MaxRecords :: Maybe Integer

Request lenses

dsa2AutoScalingGroupName :: Lens' DescribeScheduledActions (Maybe Text)

The name of the Auto Scaling group.

dsa2ScheduledActionNames :: Lens' DescribeScheduledActions [Text]

A list of scheduled actions to be described. If this list is omitted, all scheduled actions are described. The list of requested scheduled actions cannot contain more than 50 items. If an auto scaling group name is provided, the results are limited to that group. If unknown scheduled actions are requested, they are ignored with no error.

dsa2StartTime :: Lens' DescribeScheduledActions (Maybe ISO8601)

The earliest scheduled start time to return. If scheduled action names are provided, this field will be ignored.

dsa2EndTime :: Lens' DescribeScheduledActions (Maybe ISO8601)

The latest scheduled start time to return. If scheduled action names are provided, this field is ignored.

dsa2NextToken :: Lens' DescribeScheduledActions (Maybe Text)

A string that marks the start of the next batch of returned results.

dsa2MaxRecords :: Lens' DescribeScheduledActions (Maybe Integer)

The maximum number of scheduled actions to return.

Response

data DescribeScheduledActionsResponse

A scaling action that is scheduled for a future time and date. An action can be scheduled up to thirty days in advance. Starting with API version 2011-01-01, you can use recurrence to specify that a scaling action occurs regularly on a schedule.

Response constructor

mkDescribeScheduledActionsResponse :: DescribeScheduledActionsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeScheduledActionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ScheduledUpdateGroupActions :: [ScheduledUpdateGroupAction]

NextToken :: Maybe Text

Response lenses

dsarrScheduledUpdateGroupActions :: Lens' DescribeScheduledActionsResponse [ScheduledUpdateGroupAction]

A list of scheduled actions designed to update an Auto Scaling group.

dsarrNextToken :: Lens' DescribeScheduledActionsResponse (Maybe Text)

A string that marks the start of the next batch of returned results.