Network.AWS.AutoScaling.V2011_01_01.ExitStandby

Description

Move an instance out of Standby mode. To learn more about how to put instances that are in a Standby mode back into service, see Auto Scaling InService State. https:/autoscaling.amazonaws.com?InstanceIds.member.1=i-5b73d709&AutoScalingGroupName=my-asg&Version=2011-01-01&Action=ExitStandby&SignatureVersion=2&SignatureMet hod=HmacSHA256&Timestamp=2014-06-13T22%3A43%3A53.182Z&AUTHPARAMS dca4efcf-eea6-4844-8064-cab1fecd1aa2 30 PreInService 2014-06-13T22:43:53.523Z At 2014-06-13T22:43:53Z instance i-5b73d709 was moved out of standby in response to a user request, increasing the capacity from 3 to 4. my-asg {"Availability Zone":"us-east-1a"} Moving EC2 instance out of Standby: i-5b73d709 321a11c8-f34c-11e3-a434-7f10009d5849.

Request

data ExitStandby

Request constructor

mkExitStandby

Arguments

:: Text
es1AutoScalingGroupName

-> ExitStandby

Smart constructor for the minimum required parameters to construct a valid ExitStandby request.

The fields accessible through corresponding lenses are:

InstanceIds :: [Text]

AutoScalingGroupName :: Text

Request lenses

es1InstanceIds :: Lens' ExitStandby [Text]

A list of instances to move out of Standby mode. You must specify at least one instance ID.

es1AutoScalingGroupName :: Lens' ExitStandby Text

The name of the Auto Scaling group from which to move instances out of Standby mode.

Response

data ExitStandbyResponse

The output of the ExitStandby action.

Response constructor

mkExitStandbyResponse :: ExitStandbyResponse

Smart constructor for the minimum required parameters to construct a valid ExitStandbyResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Activities :: [Activity]

Response lenses

esrrActivities :: Lens' ExitStandbyResponse [Activity]

A list describing the activities related to moving instances out of Standby mode.