Network.AWS.AutoScaling.V2011_01_01.PutNotificationConfiguration

Description

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to this topic can have messages for events delivered to an endpoint such as a web server or email address. For more information see Get Email Notifications When Your Auto Scaling Group Changes A new PutNotificationConfiguration overwrites an existing configuration.

Documentation

data PutNotificationConfiguration

Constructors

PutNotificationConfiguration

Fields

_pnctNotificationTypes :: [Text]

The type of event that will cause the notification to be sent. For details about notification types supported by Auto Scaling, see DescribeAutoScalingNotificationTypes.

_pnctTopicARN :: Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

_pnctAutoScalingGroupName :: Text

The name of the Auto Scaling group.

pnctTopicARN :: Lens' PutNotificationConfiguration Text

pnctNotificationTypes :: Lens' PutNotificationConfiguration [Text]

pnctAutoScalingGroupName :: Lens' PutNotificationConfiguration Text

data PutNotificationConfigurationResponse

Constructors

PutNotificationConfigurationResponse