Network.AWS.AutoScaling.V2011_01_01.PutScalingPolicy

Description

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameter(s) you want to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &ScalingAdjustment=30 &AdjustmentType=PercentChangeInCapacity &PolicyName=my-scaleout-policy &Version=2011-01-01 &Action=PutScalingPolicy &AUTHPARAMS arn:aws:autoscaling:us-east-1:803981987763:scalingPolicy:b0dcf5e8 -02e6-4e31-9719-0675d0dc31ae:autoScalingGroupNamemy-test-asg:policyNamemy-scal eout-policy 3cfc6fef-c08b-11e2-a697-2922EXAMPLE.

Documentation

putScalingPolicy

Arguments

:: Integer
_psptScalingAdjustment

-> Text
_psptAutoScalingGroupName

-> Text
_psptPolicyName

-> Text
_psptAdjustmentType

-> PutScalingPolicy

Minimum specification for a PutScalingPolicy request.

data PutScalingPolicy

Constructors

PutScalingPolicy

Fields

_psptScalingAdjustment :: Integer

The number of instances by which to scale. AdjustmentType determines the interpretation of this number (e.g., as an absolute number or as a percentage of the existing Auto Scaling group size). A positive increment adds to the current capacity and a negative value removes from the current capacity.

_psptAutoScalingGroupName :: Text

The name or ARN of the Auto Scaling group.

_psptPolicyName :: Text

The name of the policy you want to create or update.

_psptAdjustmentType :: Text

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity. For more information about the adjustment types supported by Auto Scaling, see Scale Based on Demand.

_psptCooldown :: Maybe Integer

The amount of time, in seconds, after a scaling activity completes and before the next scaling activity can start. For more information, see Cooldown Period.

_psptMinAdjustmentStep :: Maybe Integer

Used with AdjustmentType with the value PercentChangeInCapacity, the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least the number of instances specified in the value. You will get a ValidationError if you use MinAdjustmentStep on a policy with an AdjustmentType other than PercentChangeInCapacity.

psptScalingAdjustment :: Lens' PutScalingPolicy Integer

psptPolicyName :: Lens' PutScalingPolicy Text

psptMinAdjustmentStep :: Lens' PutScalingPolicy (Maybe Integer)

psptCooldown :: Lens' PutScalingPolicy (Maybe Integer)

psptAutoScalingGroupName :: Lens' PutScalingPolicy Text

psptAdjustmentType :: Lens' PutScalingPolicy Text

data PutScalingPolicyResponse

Constructors

PutScalingPolicyResponse

Fields

_parntPolicyARN :: Maybe Text

A policy's Amazon Resource Name (ARN).

parntPolicyARN :: Iso' PutScalingPolicyResponse (Maybe Text)