Network.AWS.AutoScaling.V2011_01_01.RecordLifecycleActionHeartbeat

Description

Records a heartbeat for the lifecycle action associated with a specific token. This extends the timeout by the length of time defined by the HeartbeatTimeout parameter of the PutLifecycleHook operation. This operation is a part of the basic sequence for adding a lifecycle hook to an Auto Scaling group: Create a notification target. A target can be either an Amazon SQS queue or an Amazon SNS topic. Create an IAM role. This role allows Auto Scaling to publish lifecycle notifications to the designated SQS queue or SNS topic. Create the lifecycle hook. You can create a hook that acts when instances launch or when instances terminate. If necessary, record the lifecycle action heartbeat to keep the instance in a pending state. Complete the lifecycle action. To learn more, see Auto Scaling Pending State and Auto Scaling Terminating State.

Documentation

data RecordLifecycleActionHeartbeat

Constructors

RecordLifecycleActionHeartbeat

Fields

_rlahtLifecycleHookName :: Text

The name of the lifecycle hook.

_rlahtLifecycleActionToken :: Text

A token that uniquely identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

_rlahtAutoScalingGroupName :: Text

The name of the Auto Scaling group to which the hook belongs.

rlahtLifecycleHookName :: Lens' RecordLifecycleActionHeartbeat Text

rlahtLifecycleActionToken :: Lens' RecordLifecycleActionHeartbeat Text

rlahtAutoScalingGroupName :: Lens' RecordLifecycleActionHeartbeat Text

data RecordLifecycleActionHeartbeatResponse

Constructors

RecordLifecycleActionHeartbeatResponse