Network.AWS.AutoScaling.V2011_01_01.SetDesiredCapacity

Description

Sets the desired size of the specified AutoScalingGroup. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &HonorCooldown=false &DesiredCapacity=2 &Version=2011-01-01 &Action=SetDesiredCapacity &AUTHPARAMS 9fb7e2db-6998-11e2-a985-57c82EXAMPLE.

Documentation

setDesiredCapacity

Arguments

:: Integer
_sdctDesiredCapacity

-> Text
_sdctAutoScalingGroupName

-> SetDesiredCapacity

Minimum specification for a SetDesiredCapacity request.

data SetDesiredCapacity

Constructors

SetDesiredCapacity

Fields

_sdctDesiredCapacity :: Integer

The new capacity setting for the Auto Scaling group.

_sdctAutoScalingGroupName :: Text

The name of the Auto Scaling group.

_sdctHonorCooldown :: Maybe Bool

By default, SetDesiredCapacity overrides any cooldown period associated with the Auto Scaling group. Set to True if you want Auto Scaling to wait for the cooldown period associated with the Auto Scaling group to complete before initiating a scaling activity to set your Auto Scaling group to the new capacity setting.

sdctHonorCooldown :: Lens' SetDesiredCapacity (Maybe Bool)

sdctDesiredCapacity :: Lens' SetDesiredCapacity Integer

sdctAutoScalingGroupName :: Lens' SetDesiredCapacity Text

data SetDesiredCapacityResponse

Constructors

SetDesiredCapacityResponse