Network.AWS.AutoScaling.V2011_01_01.SetInstanceHealth

Description

Sets the health status of a specified instance that belongs to any of your Auto Scaling groups. For more information, see Configure Health Checks for Your Auto Scaling group.

Documentation

setInstanceHealth

Arguments

:: Text
_sihqInstanceId

-> Text
_sihqHealthStatus

-> SetInstanceHealth

Minimum specification for a SetInstanceHealth request.

data SetInstanceHealth

Constructors

SetInstanceHealth

Fields

_sihqInstanceId :: Text

The identifier of the Amazon EC2 instance.

_sihqHealthStatus :: Text

The health status of the instance. Set to Healthy if you want the instance to remain in service. Set to Unhealthy if you want the instance to be out of service. Auto Scaling will terminate and replace the unhealthy instance.

_sihqShouldRespectGracePeriod :: Maybe Bool

If the Auto Scaling group of the specified instance has a HealthCheckGracePeriod specified for the group, by default, this call will respect the grace period. Set this to False, if you do not want the call to respect the grace period associated with the group. For more information, see the HealthCheckGracePeriod parameter description in the CreateAutoScalingGroup action.

sihqShouldRespectGracePeriod :: Lens' SetInstanceHealth (Maybe Bool)

sihqInstanceId :: Lens' SetInstanceHealth Text

sihqHealthStatus :: Lens' SetInstanceHealth Text

data SetInstanceHealthResponse

Constructors

SetInstanceHealthResponse