Network.AWS.AutoScaling.V2011_01_01.SuspendProcesses

Description

Suspends Auto Scaling processes for an Auto Scaling group. To suspend specific process types, specify them by name with the ScalingProcesses.member.N parameter. To suspend all process types, omit the ScalingProcesses.member.N parameter. Suspending either of the two primary process types, Launch or Terminate, can prevent other process types from functioning properly. To resume processes that have been suspended, use ResumeProcesses For more information on suspending and resuming Auto Scaling process, see Suspend and Resume Auto Scaling Process.

Request

data SuspendProcesses

Request constructor

mkSuspendProcesses

Arguments

:: Text
sp1AutoScalingGroupName

-> SuspendProcesses

Smart constructor for the minimum required parameters to construct a valid SuspendProcesses request.

Request lenses

sp1AutoScalingGroupName :: Lens' SuspendProcesses Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

sp1ScalingProcesses :: Lens' SuspendProcesses [Text]

The processes that you want to suspend or resume, which can include one or more of the following: Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer To suspend all process types, omit this parameter.

Response

data SuspendProcessesResponse

Response constructor

mkSuspendProcessesResponse :: SuspendProcessesResponse

Smart constructor for the minimum required parameters to construct a valid SuspendProcessesResponse response.

This constructor is provided for convenience and testing purposes.