Network.AWS.AutoScaling.V2011_01_01.TerminateInstanceInAutoScalingGroup

Description

Terminates the specified instance. Optionally, the desired group size can be adjusted. This call simply registers a termination request. The termination of the instance cannot happen immediately.

Request

data TerminateInstanceInAutoScalingGroup

Request constructor

mkTerminateInstanceInAutoScalingGroup

Arguments

:: Text
tiiasgInstanceId

-> Bool
tiiasgShouldDecrementDesiredCapacity

-> TerminateInstanceInAutoScalingGroup

Smart constructor for the minimum required parameters to construct a valid TerminateInstanceInAutoScalingGroup request.

Request lenses

tiiasgInstanceId :: Lens' TerminateInstanceInAutoScalingGroup Text

The ID of the Amazon EC2 instance to be terminated.

tiiasgShouldDecrementDesiredCapacity :: Lens' TerminateInstanceInAutoScalingGroup Bool

Specifies whether (true) or not (false) terminating this instance should also decrement the size of the AutoScalingGroup.

Response

data TerminateInstanceInAutoScalingGroupResponse

The output for the TerminateInstanceInAutoScalingGroup action.

Response lenses

tiiasgrsActivity :: Lens' TerminateInstanceInAutoScalingGroupResponse (Maybe Activity)

A scaling Activity.