Network.AWS.AutoScaling.V2011_01_01.Types

Description

Auto Scaling allows you to scale your Amazon EC2 capacity up or down automatically according to conditions you define. With Auto Scaling, you can ensure that the number of Amazon EC2 instances you’re using increases seamlessly during demand spikes to maintain performance, and decreases automatically during demand lulls to minimize costs. Auto Scaling is particularly well suited for applications that experience hourly, daily, or weekly variability in usage. Auto Scaling is enabled by Amazon CloudWatch and available at no additional charge beyond Amazon CloudWatch fees.

Documentation

data AutoScaling

Supported version (2011-01-01) of the Auto Scaling service.

xmlOptions :: Tagged a XMLOptions

data LifecycleState

Contains a description of the current lifecycle state. The Quarantined lifecycle state is currently not used.

Constructors

LifecycleStateDetached
Detached

LifecycleStateDetaching
Detaching

LifecycleStateEnteringStandby
EnteringStandby

LifecycleStateInService
InService

LifecycleStatePending
Pending

LifecycleStatePendingProceed
Pending:Proceed

LifecycleStatePendingWait
Pending:Wait

LifecycleStateQuarantined
Quarantined

LifecycleStateStandby
Standby

LifecycleStateTerminated
Terminated

LifecycleStateTerminating
Terminating

LifecycleStateTerminatingProceed
Terminating:Proceed

LifecycleStateTerminatingWait
Terminating:Wait

data ScalingActivityStatusCode

Contains the current status of the activity.

Constructors

ScalingActivityStatusCodeCancelled
Cancelled

ScalingActivityStatusCodeFailed
Failed

ScalingActivityStatusCodeInProgress
InProgress

ScalingActivityStatusCodeMidLifecycleAction
MidLifecycleAction

ScalingActivityStatusCodePreInService
PreInService

ScalingActivityStatusCodeSuccessful
Successful

ScalingActivityStatusCodeWaitingForELBConnectionDraining
WaitingForELBConnectionDraining

ScalingActivityStatusCodeWaitingForInstanceId
WaitingForInstanceId

ScalingActivityStatusCodeWaitingForSpotInstanceId
WaitingForSpotInstanceId

ScalingActivityStatusCodeWaitingForSpotInstanceRequestId
WaitingForSpotInstanceRequestId

newtype AdjustmentType

Specifies whether the PutScalingPolicy ScalingAdjustment parameter is an absolute number or a percentage of the current capacity.

Constructors

AdjustmentType

Fields

_axAdjustmentType :: Maybe Text

A policy adjustment type. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

newtype InstanceMonitoring

Controls whether instances in this group are launched with detailed monitoring or not.

Constructors

InstanceMonitoring

Fields

_inEnabled :: Maybe Bool

If True, instance monitoring is enabled.

newtype MetricCollectionType

The MetricCollectionType data type.

Constructors

MetricCollectionType

Fields

_mcuMetric :: Maybe Text

newtype MetricGranularityType

The MetricGranularityType data type.

Constructors

MetricGranularityType

Fields

_mguGranularity :: Maybe Text

The granularity of a Metric.

newtype ProcessType

There are two primary Auto Scaling process types--Launch and Terminate. The Launch process creates a new Amazon EC2 instance for an Auto Scaling group, and the Terminate process removes an existing Amazon EC2 instance. The remaining Auto Scaling process types relate to specific Auto Scaling features: AddToLoadBalancer AlarmNotification AZRebalance HealthCheck ReplaceUnhealthy ScheduledActions If you suspend Launch or Terminate, all other process types are affected to varying degrees. The following descriptions discuss how each process type is affected by a suspension of Launch or Terminate. The AddToLoadBalancer process type adds instances to the load balancer when the instances are launched. If you suspend this process, Auto Scaling will launch the instances but will not add them to the load balancer. If you resume the AddToLoadBalancer process, Auto Scaling will also resume adding new instances to the load balancer when they are launched. However, Auto Scaling will not add running instances that were launched while the process was suspended; those instances must be added manually using the the RegisterInstancesWithLoadBalancer call in the Elastic Load Balancing API Reference. The AlarmNotification process type accepts notifications from Amazon CloudWatch alarms that are associated with the Auto Scaling group. If you suspend the AlarmNotification process type, Auto Scaling will not automatically execute scaling policies that would be triggered by alarms. Although the AlarmNotification process type is not directly affected by a suspension of Launch or Terminate, alarm notifications are often used to signal that a change in the size of the Auto Scaling group is warranted. If you suspend Launch or Terminate, Auto Scaling might not be able to implement the alarm's associated policy. The AZRebalance process type seeks to maintain a balanced number of instances across Availability Zones within a Region. If you remove an Availability Zone from your Auto Scaling group or an Availability Zone otherwise becomes unhealthy or unavailable, Auto Scaling launches new instances in an unaffected Availability Zone before terminating the unhealthy or unavailable instances. When the unhealthy Availability Zone returns to a healthy state, Auto Scaling automatically redistributes the application instances evenly across all of the designated Availability Zones. If you call SuspendProcesses on the launch process type, the AZRebalance process will neither launch new instances nor terminate existing instances. This is because the AZRebalance process terminates existing instances only after launching the replacement instances. If you call SuspendProcesses on the terminate process type, the AZRebalance process can cause your Auto Scaling group to grow up to ten percent larger than the maximum size. This is because Auto Scaling allows groups to temporarily grow larger than the maximum size during rebalancing activities. If Auto Scaling cannot terminate instances, your Auto Scaling group could remain up to ten percent larger than the maximum size until you resume the terminate process type. The HealthCheck process type checks the health of the instances. Auto Scaling marks an instance as unhealthy if Amazon EC2 or Elastic Load Balancing informs Auto Scaling that the instance is unhealthy. The HealthCheck process can override the health status of an instance that you set with SetInstanceHealth. The ReplaceUnhealthy process type terminates instances that are marked as unhealthy and subsequently creates new instances to replace them. This process calls both of the primary process types--first Terminate and then Launch. The HealthCheck process type works in conjunction with the ReplaceUnhealthly process type to provide health check functionality. If you suspend either Launch or Terminate, the ReplaceUnhealthy process type will not function properly. The ScheduledActions process type performs scheduled actions that you create with PutScheduledUpdateGroupAction. Scheduled actions often involve launching new instances or terminating existing instances. If you suspend either Launch or Terminate, your scheduled actions might not function as expected.

Constructors

ProcessType

Fields

_puProcessName :: Text

The name of a process.

data Activity

A scaling Activity is a long-running process that represents a change to your AutoScalingGroup, such as changing the size of the group. It can also be a process to replace an instance, or a process to perform any other long-running operations supported by the API.

Constructors

Activity

Fields

_hProgress :: Maybe Integer

Specifies a value between 0 and 100 that indicates the progress of the activity.

_hStartTime :: ISO8601

Provides the start time of this activity.

_hActivityId :: Text

Specifies the ID of the activity.

_hCause :: Text

Contains the reason the activity was begun.

_hStatusMessage :: Maybe Text

Contains a friendly, more verbose description of the activity status.

_hAutoScalingGroupName :: Text

The name of the Auto Scaling group.

_hEndTime :: Maybe ISO8601

Provides the end time of this activity.

_hDetails :: Maybe Text

Contains details of the scaling activity.

_hDescription :: Maybe Text

Contains a friendly, more verbose description of the scaling activity.

_hStatusCode :: ScalingActivityStatusCode

Contains the current status of the activity.

data Alarm

The Alarm data type.

Constructors

Alarm

Fields

_amAlarmName :: Maybe Text

The name of the alarm.

_amAlarmARN :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

data AutoScalingGroup

The AutoScalingGroup data type.

Constructors

AutoScalingGroup

Fields

_ashStatus :: Maybe Text

The current state of the Auto Scaling group when a DeleteAutoScalingGroup action is in progress.

_ashTerminationPolicies :: [Text]

A standalone termination policy or a list of termination policies for this Auto Scaling group.

_ashCreatedTime :: ISO8601

Specifies the date and time the Auto Scaling group was created.

_ashHealthCheckGracePeriod :: Maybe Integer

The length of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

_ashVPCZoneIdentifier :: Maybe Text

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list. When you specify VPCZoneIdentifier with AvailabilityZones, ensure that the subnets' Availability Zones match the values you specify for AvailabilityZones.

_ashDefaultCooldown :: Integer

The number of seconds after a scaling activity completes before any further scaling activities can start.

_ashMaxSize :: Integer

Contains the maximum size of the Auto Scaling group.

_ashAvailabilityZones :: [Text]

Contains a list of Availability Zones for the group.

_ashDesiredCapacity :: Integer

Specifies the desired capacity for the Auto Scaling group.

_ashMinSize :: Integer

Contains the minimum size of the Auto Scaling group.

_ashEnabledMetrics :: [EnabledMetric]

A list of metrics enabled for this Auto Scaling group.

_ashAutoScalingGroupName :: Text

Specifies the name of the group.

_ashLaunchConfigurationName :: Text

Specifies the name of the associated LaunchConfiguration.

_ashInstances :: [Instance]

Provides a summary list of Amazon EC2 instances.

_ashHealthCheckType :: Text

The service of interest for the health status check, either EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

_ashAutoScalingGroupARN :: Maybe Text

The Amazon Resource Name (ARN) of the Auto Scaling group.

_ashPlacementGroup :: Maybe Text

The name of the cluster placement group, if applicable. For more information, go to Using Cluster Instances in the Amazon EC2 User Guide.

_ashSuspendedProcesses :: [SuspendedProcess]

Suspended processes associated with this Auto Scaling group.

_ashLoadBalancerNames :: [Text]

A list of load balancers associated with this Auto Scaling group.

_ashTags :: [TagDescription]

A list of tags for the Auto Scaling group.

data AutoScalingInstanceDetails

The AutoScalingInstanceDetails data type.

Constructors

AutoScalingInstanceDetails

Fields

_asidInstanceId :: Text

The instance ID of the Amazon EC2 instance.

_asidAvailabilityZone :: Text

The Availability Zone in which this instance resides.

_asidAutoScalingGroupName :: Text

The name of the Auto Scaling group associated with this instance.

_asidLaunchConfigurationName :: Text

The launch configuration associated with this instance.

_asidHealthStatus :: Text

The health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy. Auto Scaling should terminate and replace it.

_asidLifecycleState :: Text

The life cycle state of this instance. for more information, see Instance Lifecycle State in the Auto Scaling Developer Guide.

data BlockDeviceMapping

The BlockDeviceMapping data type.

Constructors

BlockDeviceMapping

Fields

_bdnVirtualName :: Maybe Text

The virtual name associated with the device.

_bdnNoDevice :: Maybe Bool

Suppresses the device mapping. If NoDevice is set to true for the root device, the instance might fail the EC2 health check. Auto Scaling launches a replacement instance if the instance fails the health check.

_bdnEbs :: Maybe Ebs

The Elastic Block Storage volume information.

_bdnDeviceName :: Text

The name of the device within Amazon EC2 (for example, devsdh or xvdh).

data Ebs

The Elastic Block Storage volume information.

Constructors

Ebs

Fields

_esDeleteOnTermination :: Maybe Bool

Indicates whether to delete the volume on instance termination. Default: true.

_esVolumeSize :: Maybe Integer

The volume size, in gigabytes. Valid values: If the volume type is io1, the minimum size of the volume is 10. Default: If you're creating the volume from a snapshot, and you don't specify a volume size, the default is the snapshot size. Required: Required when the volume type is io1.

_esIops :: Maybe Integer

The number of I/O operations per second (IOPS) that the volume supports. The maximum ratio of IOPS to volume size is 30.0 Valid Values: Range is 100 to 4000. Default: None.

_esVolumeType :: Maybe Text

The volume type. Valid values: standard | io1 Default: standard.

_esSnapshotId :: Maybe Text

The snapshot ID.

data EnabledMetric

The EnabledMetric data type.

Constructors

EnabledMetric

Fields

_enGranularity :: Maybe Text

The granularity of the enabled metric.

_enMetric :: Maybe Text

The name of the enabled metric.

data Filter

The Filter data type.

Constructors

Filter

Fields

_gValues :: [Text]

The value of the filter.

_gName :: Maybe Text

The name of the filter. Valid Name values are: "auto-scaling-group", "key", "value", and "propagate-at-launch".

data Instance

The Instance data type.

Constructors

Instance

Fields

_ieInstanceId :: Text

Specifies the ID of the Amazon EC2 instance.

_ieAvailabilityZone :: Text

Availability Zones associated with this instance.

_ieLaunchConfigurationName :: Text

The launch configuration associated with this instance.

_ieHealthStatus :: Text

The instance's health status.

_ieLifecycleState :: LifecycleState

Contains a description of the current lifecycle state. The Quarantined lifecycle state is currently not used.

data LaunchConfiguration

The LaunchConfiguration data type.

Constructors

LaunchConfiguration

Fields

_ldAssociatePublicIpAddress :: Maybe Bool

Specifies whether the instance is associated with a public IP address (true) or not (false).

_ldSecurityGroups :: [Text]

A description of the security groups to associate with the Amazon EC2 instances.

_ldSpotPrice :: Maybe Text

Specifies the price to bid when launching Spot Instances.

_ldCreatedTime :: ISO8601

Provides the creation date and time for this launch configuration.

_ldInstanceMonitoring :: Maybe InstanceMonitoring

Controls whether instances in this group are launched with detailed monitoring or not.

_ldKeyName :: Maybe Text

Provides the name of the Amazon EC2 key pair.

_ldRamdiskId :: Maybe Text

Provides ID of the RAM disk associated with the Amazon EC2 AMI.

_ldKernelId :: Maybe Text

Provides the ID of the kernel associated with the Amazon EC2 AMI.

_ldInstanceType :: Text

Specifies the instance type of the Amazon EC2 instance.

_ldEbsOptimized :: Maybe Bool

Specifies whether the instance is optimized for EBS I/O (true) or not (false).

_ldUserData :: Maybe Text

The user data available to the launched Amazon EC2 instances.

_ldIamInstanceProfile :: Maybe Text

Provides the name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

_ldImageId :: Text

Provides the unique ID of the Amazon Machine Image (AMI) that was assigned during registration.

_ldLaunchConfigurationName :: Text

Specifies the name of the launch configuration.

_ldLaunchConfigurationARN :: Maybe Text

The launch configuration's Amazon Resource Name (ARN).

_ldPlacementTenancy :: Maybe Text

Specifies the tenancy of the instance. It can be either default or dedicated. An instance with dedicated tenancy runs in an isolated, single-tenant hardware and it can only be launched in a VPC.

_ldBlockDeviceMappings :: [BlockDeviceMapping]

Specifies how block devices are exposed to the instance. Each mapping is made up of a virtualName and a deviceName.

data LifecycleHook

A lifecycle hook tells Auto Scaling that you want to perform an action when an instance launches or terminates. When you have a lifecycle hook in place, the Auto Scaling group will either: Pause the instance after it launches, but before it is put into service Pause the instance as it terminates, but before it is fully terminated To learn more, see Auto Scaling Pending State and Auto Scaling Terminating State.

Constructors

LifecycleHook

Fields

_liDefaultResult :: Maybe Text

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The value for this parameter can be either CONTINUE or ABANDON. The default value for this parameter is CONTINUE.

_liLifecycleHookName :: Maybe Text

The name of the lifecycle action hook.

_liHeartbeatTimeout :: Maybe Integer

Defines the amount of time that can elapse before the lifecycle hook times out. When the lifecycle hook times out, Auto Scaling performs the action defined in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

_liAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group to which the lifecycle action belongs.

_liNotificationMetadata :: Maybe Text

Contains additional information that you want to include any time Auto Scaling sends a message to the notification target.

_liGlobalTimeout :: Maybe Integer

The maximum length of time an instance can remain in a Pending:Wait or Terminating:Wait state. Currently, this value is set at 48 hours.

_liNotificationTargetARN :: Maybe Text

The ARN of the notification target that Auto Scaling will use to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target will include: Lifecycle action token User account ID Name of the Auto Scaling group Lifecycle hook name EC2 instance ID Lifecycle transition Notification metadata.

_liLifecycleTransition :: Maybe Text

The Amazon EC2 instance state to which you want to attach the lifecycle hook. See DescribeLifecycleHooks for a list of available lifecycle hook types.

_liRoleARN :: Maybe Text

The ARN of the Amazon IAM role that allows the Auto Scaling group to publish to the specified notification target.

data NotificationConfiguration

The NotificationConfiguration data type.

Constructors

NotificationConfiguration

Fields

_neTopicARN :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

_neAutoScalingGroupName :: Maybe Text

Specifies the Auto Scaling group name.

_neNotificationType :: Maybe Text

The types of events for an action to start.

data ScalingPolicy

The ScalingPolicy data type.

Constructors

ScalingPolicy

Fields

_ssMinAdjustmentStep :: Maybe Integer

Changes the DesiredCapacity of the Auto Scaling group by at least the specified number of instances.

_ssPolicyName :: Maybe Text

The name of the scaling policy.

_ssAdjustmentType :: Maybe Text

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

_ssAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group associated with this scaling policy.

_ssScalingAdjustment :: Maybe Integer

The number associated with the specified adjustment type. A positive value adds to the current capacity and a negative value removes from the current capacity.

_ssCooldown :: Maybe Integer

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

_ssPolicyARN :: Maybe Text

The Amazon Resource Name (ARN) of the policy.

_ssAlarms :: [Alarm]

A list of CloudWatch Alarms related to the policy.

data ScheduledUpdateGroupAction

This data type stores information about a scheduled update to an Auto Scaling group.

Constructors

ScheduledUpdateGroupAction

Fields

_sugbScheduledActionARN :: Maybe Text

The Amazon Resource Name (ARN) of this scheduled action.

_sugbStartTime :: Maybe ISO8601

The time that the action is scheduled to begin. This value can be up to one month in the future. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

_sugbTime :: Maybe ISO8601

Time is deprecated. The time that the action is scheduled to begin. Time is an alias for StartTime.

_sugbScheduledActionName :: Maybe Text

The name of this scheduled action.

_sugbMaxSize :: Maybe Integer

The maximum size of the Auto Scaling group.

_sugbRecurrence :: Maybe Text

The regular schedule that an action occurs.

_sugbDesiredCapacity :: Maybe Integer

The number of instances you prefer to maintain in your Auto Scaling group.

_sugbMinSize :: Maybe Integer

The minimum size of the Auto Scaling group.

_sugbAutoScalingGroupName :: Maybe Text

The name of the Auto Scaling group to be updated.

_sugbEndTime :: Maybe ISO8601

The time that the action is scheduled to end. This value can be up to one month in the future.

data SuspendedProcess

An Auto Scaling process that has been suspended. For more information, see ProcessType.

Constructors

SuspendedProcess

Fields

_sqProcessName :: Maybe Text

The name of the suspended process.

_sqSuspensionReason :: Maybe Text

The reason that the process was suspended.

data Tag

The tag applied to an Auto Scaling group.

Constructors

Tag

Fields

_tgResourceId :: Maybe Text

The name of the Auto Scaling group.

_tgResourceType :: Maybe Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

_tgValue :: Maybe Text

The value of the tag.

_tgKey :: Text

The key of the tag.

_tgPropagateAtLaunch :: Maybe Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

data TagDescription

The tag applied to an Auto Scaling group.

Constructors

TagDescription

Fields

_tdResourceId :: Maybe Text

The name of the Auto Scaling group.

_tdResourceType :: Maybe Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

_tdValue :: Maybe Text

The value of the tag.

_tdKey :: Maybe Text

The key of the tag.

_tdPropagateAtLaunch :: Maybe Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

axAdjustmentType :: Iso' AdjustmentType (Maybe Text)

inEnabled :: Iso' InstanceMonitoring (Maybe Bool)

mcuMetric :: Iso' MetricCollectionType (Maybe Text)

mguGranularity :: Iso' MetricGranularityType (Maybe Text)

puProcessName :: Iso' ProcessType Text

hStatusMessage :: Lens' Activity (Maybe Text)

hStatusCode :: Lens' Activity ScalingActivityStatusCode

hStartTime :: Lens' Activity ISO8601

hProgress :: Lens' Activity (Maybe Integer)

hEndTime :: Lens' Activity (Maybe ISO8601)

hDetails :: Lens' Activity (Maybe Text)

hDescription :: Lens' Activity (Maybe Text)

hCause :: Lens' Activity Text

hAutoScalingGroupName :: Lens' Activity Text

hActivityId :: Lens' Activity Text

amAlarmName :: Lens' Alarm (Maybe Text)

amAlarmARN :: Lens' Alarm (Maybe Text)

ashVPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

ashTerminationPolicies :: Lens' AutoScalingGroup [Text]

ashTags :: Lens' AutoScalingGroup [TagDescription]

ashSuspendedProcesses :: Lens' AutoScalingGroup [SuspendedProcess]

ashStatus :: Lens' AutoScalingGroup (Maybe Text)

ashPlacementGroup :: Lens' AutoScalingGroup (Maybe Text)

ashMinSize :: Lens' AutoScalingGroup Integer

ashMaxSize :: Lens' AutoScalingGroup Integer

ashLoadBalancerNames :: Lens' AutoScalingGroup [Text]

ashLaunchConfigurationName :: Lens' AutoScalingGroup Text

ashInstances :: Lens' AutoScalingGroup [Instance]

ashHealthCheckType :: Lens' AutoScalingGroup Text

ashHealthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Integer)

ashEnabledMetrics :: Lens' AutoScalingGroup [EnabledMetric]

ashDesiredCapacity :: Lens' AutoScalingGroup Integer

ashDefaultCooldown :: Lens' AutoScalingGroup Integer

ashCreatedTime :: Lens' AutoScalingGroup ISO8601

ashAvailabilityZones :: Lens' AutoScalingGroup [Text]

ashAutoScalingGroupName :: Lens' AutoScalingGroup Text

ashAutoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

asidLifecycleState :: Lens' AutoScalingInstanceDetails Text

asidLaunchConfigurationName :: Lens' AutoScalingInstanceDetails Text

asidInstanceId :: Lens' AutoScalingInstanceDetails Text

asidHealthStatus :: Lens' AutoScalingInstanceDetails Text

asidAvailabilityZone :: Lens' AutoScalingInstanceDetails Text

asidAutoScalingGroupName :: Lens' AutoScalingInstanceDetails Text

bdnVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

bdnNoDevice :: Lens' BlockDeviceMapping (Maybe Bool)

bdnEbs :: Lens' BlockDeviceMapping (Maybe Ebs)

bdnDeviceName :: Lens' BlockDeviceMapping Text

esVolumeType :: Lens' Ebs (Maybe Text)

esVolumeSize :: Lens' Ebs (Maybe Integer)

esSnapshotId :: Lens' Ebs (Maybe Text)

esIops :: Lens' Ebs (Maybe Integer)

esDeleteOnTermination :: Lens' Ebs (Maybe Bool)

enMetric :: Lens' EnabledMetric (Maybe Text)

enGranularity :: Lens' EnabledMetric (Maybe Text)

gValues :: Lens' Filter [Text]

gName :: Lens' Filter (Maybe Text)

ieLifecycleState :: Lens' Instance LifecycleState

ieLaunchConfigurationName :: Lens' Instance Text

ieInstanceId :: Lens' Instance Text

ieHealthStatus :: Lens' Instance Text

ieAvailabilityZone :: Lens' Instance Text

ldUserData :: Lens' LaunchConfiguration (Maybe Text)

ldSpotPrice :: Lens' LaunchConfiguration (Maybe Text)

ldSecurityGroups :: Lens' LaunchConfiguration [Text]

ldRamdiskId :: Lens' LaunchConfiguration (Maybe Text)

ldPlacementTenancy :: Lens' LaunchConfiguration (Maybe Text)

ldLaunchConfigurationName :: Lens' LaunchConfiguration Text

ldLaunchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

ldKeyName :: Lens' LaunchConfiguration (Maybe Text)

ldKernelId :: Lens' LaunchConfiguration (Maybe Text)

ldInstanceType :: Lens' LaunchConfiguration Text

ldInstanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

ldImageId :: Lens' LaunchConfiguration Text

ldIamInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

ldEbsOptimized :: Lens' LaunchConfiguration (Maybe Bool)

ldCreatedTime :: Lens' LaunchConfiguration ISO8601

ldBlockDeviceMappings :: Lens' LaunchConfiguration [BlockDeviceMapping]

ldAssociatePublicIpAddress :: Lens' LaunchConfiguration (Maybe Bool)

liRoleARN :: Lens' LifecycleHook (Maybe Text)

liNotificationTargetARN :: Lens' LifecycleHook (Maybe Text)

liNotificationMetadata :: Lens' LifecycleHook (Maybe Text)

liLifecycleTransition :: Lens' LifecycleHook (Maybe Text)

liLifecycleHookName :: Lens' LifecycleHook (Maybe Text)

liHeartbeatTimeout :: Lens' LifecycleHook (Maybe Integer)

liGlobalTimeout :: Lens' LifecycleHook (Maybe Integer)

liDefaultResult :: Lens' LifecycleHook (Maybe Text)

liAutoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

neTopicARN :: Lens' NotificationConfiguration (Maybe Text)

neNotificationType :: Lens' NotificationConfiguration (Maybe Text)

neAutoScalingGroupName :: Lens' NotificationConfiguration (Maybe Text)

ssScalingAdjustment :: Lens' ScalingPolicy (Maybe Integer)

ssPolicyName :: Lens' ScalingPolicy (Maybe Text)

ssPolicyARN :: Lens' ScalingPolicy (Maybe Text)

ssMinAdjustmentStep :: Lens' ScalingPolicy (Maybe Integer)

ssCooldown :: Lens' ScalingPolicy (Maybe Integer)

ssAutoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

ssAlarms :: Lens' ScalingPolicy [Alarm]

ssAdjustmentType :: Lens' ScalingPolicy (Maybe Text)

sugbTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

sugbStartTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

sugbScheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

sugbScheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

sugbRecurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

sugbMinSize :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

sugbMaxSize :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

sugbEndTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

sugbDesiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

sugbAutoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

sqSuspensionReason :: Lens' SuspendedProcess (Maybe Text)

sqProcessName :: Lens' SuspendedProcess (Maybe Text)

tgValue :: Lens' Tag (Maybe Text)

tgResourceType :: Lens' Tag (Maybe Text)

tgResourceId :: Lens' Tag (Maybe Text)

tgPropagateAtLaunch :: Lens' Tag (Maybe Bool)

tgKey :: Lens' Tag Text

tdValue :: Lens' TagDescription (Maybe Text)

tdResourceType :: Lens' TagDescription (Maybe Text)

tdResourceId :: Lens' TagDescription (Maybe Text)

tdPropagateAtLaunch :: Lens' TagDescription (Maybe Bool)

tdKey :: Lens' TagDescription (Maybe Text)