Network.AWS.AutoScaling.CompleteLifecycleAction

Description

Completes the lifecycle action for the specified token or instance with the specified result.

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

Creating a Request

completeLifecycleAction

Arguments

:: Text
claLifecycleHookName

-> Text
claAutoScalingGroupName

-> Text
claLifecycleActionResult

-> CompleteLifecycleAction

Creates a value of CompleteLifecycleAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

claInstanceId - The ID of the instance.

claLifecycleActionToken - A universally unique identifier (UUID) that identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

claLifecycleHookName - The name of the lifecycle hook.

claAutoScalingGroupName - The name of the group for the lifecycle hook.

claLifecycleActionResult - The action for the group to take. This parameter can be either CONTINUE or ABANDON .

data CompleteLifecycleAction

Contains the parameters for CompleteLifecycleAction.

See: completeLifecycleAction smart constructor.

Request Lenses

claInstanceId :: Lens' CompleteLifecycleAction (Maybe Text)

The ID of the instance.

claLifecycleActionToken :: Lens' CompleteLifecycleAction (Maybe Text)

A universally unique identifier (UUID) that identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

claLifecycleHookName :: Lens' CompleteLifecycleAction Text

The name of the lifecycle hook.

claAutoScalingGroupName :: Lens' CompleteLifecycleAction Text

The name of the group for the lifecycle hook.

claLifecycleActionResult :: Lens' CompleteLifecycleAction Text

The action for the group to take. This parameter can be either CONTINUE or ABANDON .

Destructuring the Response

completeLifecycleActionResponse

Arguments

:: Int
clarsResponseStatus

-> CompleteLifecycleActionResponse

Creates a value of CompleteLifecycleActionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clarsResponseStatus - -- | The response status code.

data CompleteLifecycleActionResponse

Contains the output of CompleteLifecycleAction.

See: completeLifecycleActionResponse smart constructor.

Response Lenses

clarsResponseStatus :: Lens' CompleteLifecycleActionResponse Int

- | The response status code.