Network.AWS.AutoScaling.CreateAutoScalingGroup

Description

Creates an Auto Scaling group with the specified name and attributes.

If you exceed your maximum limit of Auto Scaling groups, which by default is 20 per region, the call fails. For information about viewing and updating this limit, see DescribeAccountLimits .

For more information, see Auto Scaling Groups in the Auto Scaling User Guide .

Creating a Request

createAutoScalingGroup

Arguments

:: Text
casgAutoScalingGroupName

-> Int
casgMinSize

-> Int
casgMaxSize

-> CreateAutoScalingGroup

Creates a value of CreateAutoScalingGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

casgInstanceId - The ID of the instance used to create a launch configuration for the group. Alternatively, specify a launch configuration instead of an EC2 instance. When you specify an ID of an instance, Auto Scaling creates a new launch configuration and associates it with the group. This launch configuration derives its attributes from the specified instance, with the exception of the block device mapping. For more information, see Create an Auto Scaling Group Using an EC2 Instance in the Auto Scaling User Guide .

casgTerminationPolicies - One or more termination policies used to select the instance to terminate. These policies are executed in the order that they are listed. For more information, see Controlling Which Instances Auto Scaling Terminates During Scale In in the Auto Scaling User Guide .

casgHealthCheckGracePeriod - The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service. During this time, any health check failures for the instance are ignored. The default is 0. This parameter is required if you are adding an ELB health check. For more information, see Health Checks in the Auto Scaling User Guide .

casgNewInstancesProtectedFromScaleIn - Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

casgVPCZoneIdentifier - A comma-separated list of subnet identifiers for your virtual private cloud (VPC). If you specify subnets and Availability Zones with this call, ensure that the subnets' Availability Zones match the Availability Zones specified. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide .

casgTargetGroupARNs - The Amazon Resource Names (ARN) of the target groups.

casgDefaultCooldown - The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. The default is 300. For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

casgAvailabilityZones - One or more Availability Zones for the group. This parameter is optional if you specify one or more subnets.

casgDesiredCapacity - The number of EC2 instances that should be running in the group. This number must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group.

casgLaunchConfigurationName - The name of the launch configuration. Alternatively, specify an EC2 instance instead of a launch configuration.

casgHealthCheckType - The service to use for the health checks. The valid values are EC2 and ELB . By default, health checks use Amazon EC2 instance status checks to determine the health of an instance. For more information, see Health Checks in the Auto Scaling User Guide .

casgPlacementGroup - The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide .

casgLoadBalancerNames - One or more Classic Load Balancers. To specify an Application Load Balancer, use TargetGroupARNs instead. For more information, see Using a Load Balancer With an Auto Scaling Group in the Auto Scaling User Guide .

casgTags - One or more tags. For more information, see Tagging Auto Scaling Groups and Instances in the Auto Scaling User Guide .

casgAutoScalingGroupName - The name of the group. This name must be unique within the scope of your AWS account.

casgMinSize - The minimum size of the group.

casgMaxSize - The maximum size of the group.

data CreateAutoScalingGroup

Contains the parameters for CreateAutoScalingGroup.

See: createAutoScalingGroup smart constructor.

Request Lenses

casgInstanceId :: Lens' CreateAutoScalingGroup (Maybe Text)

The ID of the instance used to create a launch configuration for the group. Alternatively, specify a launch configuration instead of an EC2 instance. When you specify an ID of an instance, Auto Scaling creates a new launch configuration and associates it with the group. This launch configuration derives its attributes from the specified instance, with the exception of the block device mapping. For more information, see Create an Auto Scaling Group Using an EC2 Instance in the Auto Scaling User Guide .

casgTerminationPolicies :: Lens' CreateAutoScalingGroup [Text]

One or more termination policies used to select the instance to terminate. These policies are executed in the order that they are listed. For more information, see Controlling Which Instances Auto Scaling Terminates During Scale In in the Auto Scaling User Guide .

casgHealthCheckGracePeriod :: Lens' CreateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service. During this time, any health check failures for the instance are ignored. The default is 0. This parameter is required if you are adding an ELB health check. For more information, see Health Checks in the Auto Scaling User Guide .

casgNewInstancesProtectedFromScaleIn :: Lens' CreateAutoScalingGroup (Maybe Bool)

Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

casgVPCZoneIdentifier :: Lens' CreateAutoScalingGroup (Maybe Text)

A comma-separated list of subnet identifiers for your virtual private cloud (VPC). If you specify subnets and Availability Zones with this call, ensure that the subnets' Availability Zones match the Availability Zones specified. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide .

casgTargetGroupARNs :: Lens' CreateAutoScalingGroup [Text]

The Amazon Resource Names (ARN) of the target groups.

casgDefaultCooldown :: Lens' CreateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start. The default is 300. For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

casgAvailabilityZones :: Lens' CreateAutoScalingGroup (Maybe (NonEmpty Text))

One or more Availability Zones for the group. This parameter is optional if you specify one or more subnets.

casgDesiredCapacity :: Lens' CreateAutoScalingGroup (Maybe Int)

The number of EC2 instances that should be running in the group. This number must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group.

casgLaunchConfigurationName :: Lens' CreateAutoScalingGroup (Maybe Text)

The name of the launch configuration. Alternatively, specify an EC2 instance instead of a launch configuration.

casgHealthCheckType :: Lens' CreateAutoScalingGroup (Maybe Text)

The service to use for the health checks. The valid values are EC2 and ELB . By default, health checks use Amazon EC2 instance status checks to determine the health of an instance. For more information, see Health Checks in the Auto Scaling User Guide .

casgPlacementGroup :: Lens' CreateAutoScalingGroup (Maybe Text)

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide .

casgLoadBalancerNames :: Lens' CreateAutoScalingGroup [Text]

One or more Classic Load Balancers. To specify an Application Load Balancer, use TargetGroupARNs instead. For more information, see Using a Load Balancer With an Auto Scaling Group in the Auto Scaling User Guide .

casgTags :: Lens' CreateAutoScalingGroup [Tag]

One or more tags. For more information, see Tagging Auto Scaling Groups and Instances in the Auto Scaling User Guide .

casgAutoScalingGroupName :: Lens' CreateAutoScalingGroup Text

The name of the group. This name must be unique within the scope of your AWS account.

casgMinSize :: Lens' CreateAutoScalingGroup Int

The minimum size of the group.

casgMaxSize :: Lens' CreateAutoScalingGroup Int

The maximum size of the group.

Destructuring the Response

createAutoScalingGroupResponse :: CreateAutoScalingGroupResponse

Creates a value of CreateAutoScalingGroupResponse with the minimum fields required to make a request.

data CreateAutoScalingGroupResponse

See: createAutoScalingGroupResponse smart constructor.