Network.AWS.AutoScaling.CreateAutoScalingGroup

Description

Creates an Auto Scaling group with the specified name and attributes.

If you exceed your maximum limit of Auto Scaling groups, which by default is 20 per region, the call fails. For information about viewing and updating these limits, see DescribeAccountLimits.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_CreateAutoScalingGroup.html

Request

data CreateAutoScalingGroup

Request constructor

createAutoScalingGroup

Arguments

:: Text
casgAutoScalingGroupName

-> Int
casgMinSize

-> Int
casgMaxSize

-> NonEmpty Text
casgAvailabilityZones

-> CreateAutoScalingGroup

CreateAutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

casgAutoScalingGroupName :: Text

casgAvailabilityZones :: NonEmpty Text

casgDefaultCooldown :: Maybe Int

casgDesiredCapacity :: Maybe Int

casgHealthCheckGracePeriod :: Maybe Int

casgHealthCheckType :: Maybe Text

casgInstanceId :: Maybe Text

casgLaunchConfigurationName :: Maybe Text

casgLoadBalancerNames :: [Text]

casgMaxSize :: Int

casgMinSize :: Int

casgPlacementGroup :: Maybe Text

casgTags :: [Tag]

casgTerminationPolicies :: [Text]

casgVPCZoneIdentifier :: Maybe Text

Request lenses

casgAutoScalingGroupName :: Lens' CreateAutoScalingGroup Text

The name of the group. This name must be unique within the scope of your AWS account.

casgAvailabilityZones :: Lens' CreateAutoScalingGroup (NonEmpty Text)

One or more Availability Zones for the group. This parameter is optional if you specify subnets using the VPCZoneIdentifier parameter.

casgDefaultCooldown :: Lens' CreateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start.

If DefaultCooldown is not specified, the default value is 300. For more information, see Understanding Auto Scaling Cooldowns in the Auto ScalingDeveloper Guide.

casgDesiredCapacity :: Lens' CreateAutoScalingGroup (Maybe Int)

The number of EC2 instances that should be running in the group. This value must be greater than or equal to the minimum size of the group and less than or equal to the maximum size of the group.

casgHealthCheckGracePeriod :: Lens' CreateAutoScalingGroup (Maybe Int)

The amount of time, in seconds, after an EC2 instance comes into service that Auto Scaling starts checking its health. During this time, any health check failures for the instance are ignored.

This parameter is required if you are adding an ELB health check. Frequently, new instances need to warm up, briefly, before they can pass a health check. To provide ample warm-up time, set the health check grace period of the group to match the expected startup period of your application.

For more information, see Add an Elastic Load Balancing Health Check to YourAuto Scaling Group in the Auto Scaling Developer Guide.

casgHealthCheckType :: Lens' CreateAutoScalingGroup (Maybe Text)

The service to use for the health checks. The valid values are EC2 and ELB.

By default, health checks use Amazon EC2 instance status checks to determine the health of an instance. For more information, see Health Checks.

casgInstanceId :: Lens' CreateAutoScalingGroup (Maybe Text)

The ID of the EC2 instance used to create a launch configuration for the group. Alternatively, use the LaunchConfigurationName parameter to specify a launch configuration instead of an EC2 instance.

When you specify an ID of an instance, Auto Scaling creates a new launch configuration and associates it with the group. This launch configuration derives its attributes from the specified instance, with the exception of the block device mapping.

For more information, see Create an Auto Scaling Group Using an EC2 InstanceID in the Auto Scaling Developer Guide.

casgLaunchConfigurationName :: Lens' CreateAutoScalingGroup (Maybe Text)

The name of the launch configuration. Alternatively, use the InstanceId parameter to specify an EC2 instance instead of a launch configuration.

casgLoadBalancerNames :: Lens' CreateAutoScalingGroup [Text]

One or more load balancers.

For more information, see Load Balance Your Auto Scaling Group in the AutoScaling Developer Guide.

casgMaxSize :: Lens' CreateAutoScalingGroup Int

The maximum size of the group.

casgMinSize :: Lens' CreateAutoScalingGroup Int

The minimum size of the group.

casgPlacementGroup :: Lens' CreateAutoScalingGroup (Maybe Text)

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups.

casgTags :: Lens' CreateAutoScalingGroup [Tag]

The tag to be created or updated. Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values: key=value, value=value, propagate=true or false. Value and propagate are optional parameters.

For more information, see Add, Modify, or Remove Auto Scaling Group Tags in the Auto Scaling Developer Guide.

casgTerminationPolicies :: Lens' CreateAutoScalingGroup [Text]

One or more termination policies used to select the instance to terminate. These policies are executed in the order that they are listed.

For more information, see Choosing a Termination Policy for Your AutoScaling Group in the Auto Scaling Developer Guide.

casgVPCZoneIdentifier :: Lens' CreateAutoScalingGroup (Maybe Text)

A comma-separated list of subnet identifiers for your virtual private cloud (VPC).

If you specify subnets and Availability Zones with this call, ensure that the subnets' Availability Zones match the Availability Zones specified.

For more information, see Auto Scaling and Amazon VPC in the Auto ScalingDeveloper Guide.

Response

data CreateAutoScalingGroupResponse

Response constructor

createAutoScalingGroupResponse :: CreateAutoScalingGroupResponse

CreateAutoScalingGroupResponse constructor.