Network.AWS.AutoScaling.CreateLaunchConfiguration

Description

Creates a launch configuration.

If you exceed your maximum limit of launch configurations, which by default is 100 per region, the call fails. For information about viewing and updating this limit, see DescribeAccountLimits .

For more information, see Launch Configurations in the Auto Scaling User Guide .

Creating a Request

createLaunchConfiguration

Arguments

:: Text
clcLaunchConfigurationName

-> CreateLaunchConfiguration

Creates a value of CreateLaunchConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

clcInstanceId - The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, with the exception of the block device mapping. To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request. For more information, see Create a Launch Configuration Using an EC2 Instance in the Auto Scaling User Guide .

clcAssociatePublicIPAddress - Used for groups that launch instances into a virtual private cloud (VPC). Specifies whether to assign a public IP address to each instance. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide . If you specify this parameter, be sure to specify at least one subnet when you create your group. Default: If the instance is launched into a default subnet, the default is true . If the instance is launched into a nondefault subnet, the default is false . For more information, see Supported Platforms in the Amazon Elastic Compute Cloud User Guide .

clcSecurityGroups - One or more security groups with which to associate the instances. If your instances are launched in EC2-Classic, you can either specify security group names or the security group IDs. For more information about security groups for EC2-Classic, see Amazon EC2 Security Groups in the Amazon Elastic Compute Cloud User Guide . If your instances are launched into a VPC, specify security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide .

clcSpotPrice - The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot market price. For more information, see Launching Spot Instances in Your Auto Scaling Group in the Auto Scaling User Guide .

clcInstanceMonitoring - Enables detailed monitoring (true ) or basic monitoring (false ) for the Auto Scaling instances.

clcKeyName - The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon Elastic Compute Cloud User Guide .

clcClassicLinkVPCSecurityGroups - The IDs of one or more security groups for the specified ClassicLink-enabled VPC. This parameter is required if you specify a ClassicLink-enabled VPC, and is not supported otherwise. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

clcRAMDiskId - The ID of the RAM disk associated with the AMI.

clcKernelId - The ID of the kernel associated with the AMI.

clcInstanceType - The instance type of the EC2 instance. For information about available instance types, see Available Instance Types in the Amazon Elastic Compute Cloud User Guide.

clcEBSOptimized - Indicates whether the instance is optimized for Amazon EBS IO. By default, the instance is not optimized for EBS IO. The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal IO performance. This optimization is not available with all instance types. Additional usage charges apply. For more information, see <http:docs.aws.amazon.comAWSEC2latestUserGuideEBSOptimized.html Amazon EBS-Optimized Instances> in the Amazon Elastic Compute Cloud User Guide/ .

clcUserData - The user data to make available to the launched EC2 instances. For more information, see Instance Metadata and User Data in the Amazon Elastic Compute Cloud User Guide .

clcClassicLinkVPCId - The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. This parameter is supported only if you are launching EC2-Classic instances. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

clcIAMInstanceProfile - The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. EC2 instances launched with an IAM role will automatically have AWS security credentials available. You can use IAM roles with Auto Scaling to automatically enable applications running on your EC2 instances to securely access other AWS resources. For more information, see Launch Auto Scaling Instances with an IAM Role in the Auto Scaling User Guide .

clcImageId - The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon Elastic Compute Cloud User Guide .

clcPlacementTenancy - The tenancy of the instance. An instance with a tenancy of dedicated runs on single-tenant hardware and can only be launched into a VPC. You must set the value of this parameter to dedicated if want to launch Dedicated Instances into a shared tenancy VPC (VPC with instance placement tenancy attribute set to default ). If you specify this parameter, be sure to specify at least one subnet when you create your group. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide . Valid values: default | dedicated

clcBlockDeviceMappings - One or more mappings that specify how block devices are exposed to the instance. For more information, see Block Device Mapping in the Amazon Elastic Compute Cloud User Guide .

clcLaunchConfigurationName - The name of the launch configuration. This name must be unique within the scope of your AWS account.

data CreateLaunchConfiguration

Contains the parameters for CreateLaunchConfiguration.

See: createLaunchConfiguration smart constructor.

Request Lenses

clcInstanceId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the instance to use to create the launch configuration. The new launch configuration derives attributes from the instance, with the exception of the block device mapping. To create a launch configuration with a block device mapping or override any other instance attributes, specify them as part of the same request. For more information, see Create a Launch Configuration Using an EC2 Instance in the Auto Scaling User Guide .

clcAssociatePublicIPAddress :: Lens' CreateLaunchConfiguration (Maybe Bool)

Used for groups that launch instances into a virtual private cloud (VPC). Specifies whether to assign a public IP address to each instance. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide . If you specify this parameter, be sure to specify at least one subnet when you create your group. Default: If the instance is launched into a default subnet, the default is true . If the instance is launched into a nondefault subnet, the default is false . For more information, see Supported Platforms in the Amazon Elastic Compute Cloud User Guide .

clcSecurityGroups :: Lens' CreateLaunchConfiguration [Text]

One or more security groups with which to associate the instances. If your instances are launched in EC2-Classic, you can either specify security group names or the security group IDs. For more information about security groups for EC2-Classic, see Amazon EC2 Security Groups in the Amazon Elastic Compute Cloud User Guide . If your instances are launched into a VPC, specify security group IDs. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide .

clcSpotPrice :: Lens' CreateLaunchConfiguration (Maybe Text)

The maximum hourly price to be paid for any Spot Instance launched to fulfill the request. Spot Instances are launched when the price you specify exceeds the current Spot market price. For more information, see Launching Spot Instances in Your Auto Scaling Group in the Auto Scaling User Guide .

clcInstanceMonitoring :: Lens' CreateLaunchConfiguration (Maybe InstanceMonitoring)

Enables detailed monitoring (true ) or basic monitoring (false ) for the Auto Scaling instances.

clcKeyName :: Lens' CreateLaunchConfiguration (Maybe Text)

The name of the key pair. For more information, see Amazon EC2 Key Pairs in the Amazon Elastic Compute Cloud User Guide .

clcClassicLinkVPCSecurityGroups :: Lens' CreateLaunchConfiguration [Text]

The IDs of one or more security groups for the specified ClassicLink-enabled VPC. This parameter is required if you specify a ClassicLink-enabled VPC, and is not supported otherwise. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

clcRAMDiskId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the RAM disk associated with the AMI.

clcKernelId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the kernel associated with the AMI.

clcInstanceType :: Lens' CreateLaunchConfiguration (Maybe Text)

The instance type of the EC2 instance. For information about available instance types, see Available Instance Types in the Amazon Elastic Compute Cloud User Guide.

clcEBSOptimized :: Lens' CreateLaunchConfiguration (Maybe Bool)

Indicates whether the instance is optimized for Amazon EBS IO. By default, the instance is not optimized for EBS IO. The optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal IO performance. This optimization is not available with all instance types. Additional usage charges apply. For more information, see <http:docs.aws.amazon.comAWSEC2latestUserGuideEBSOptimized.html Amazon EBS-Optimized Instances> in the Amazon Elastic Compute Cloud User Guide/ .

clcUserData :: Lens' CreateLaunchConfiguration (Maybe Text)

The user data to make available to the launched EC2 instances. For more information, see Instance Metadata and User Data in the Amazon Elastic Compute Cloud User Guide .

clcClassicLinkVPCId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. This parameter is supported only if you are launching EC2-Classic instances. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

clcIAMInstanceProfile :: Lens' CreateLaunchConfiguration (Maybe Text)

The name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. EC2 instances launched with an IAM role will automatically have AWS security credentials available. You can use IAM roles with Auto Scaling to automatically enable applications running on your EC2 instances to securely access other AWS resources. For more information, see Launch Auto Scaling Instances with an IAM Role in the Auto Scaling User Guide .

clcImageId :: Lens' CreateLaunchConfiguration (Maybe Text)

The ID of the Amazon Machine Image (AMI) to use to launch your EC2 instances. For more information, see Finding an AMI in the Amazon Elastic Compute Cloud User Guide .

clcPlacementTenancy :: Lens' CreateLaunchConfiguration (Maybe Text)

The tenancy of the instance. An instance with a tenancy of dedicated runs on single-tenant hardware and can only be launched into a VPC. You must set the value of this parameter to dedicated if want to launch Dedicated Instances into a shared tenancy VPC (VPC with instance placement tenancy attribute set to default ). If you specify this parameter, be sure to specify at least one subnet when you create your group. For more information, see Launching Auto Scaling Instances in a VPC in the Auto Scaling User Guide . Valid values: default | dedicated

clcBlockDeviceMappings :: Lens' CreateLaunchConfiguration [BlockDeviceMapping]

One or more mappings that specify how block devices are exposed to the instance. For more information, see Block Device Mapping in the Amazon Elastic Compute Cloud User Guide .

clcLaunchConfigurationName :: Lens' CreateLaunchConfiguration Text

The name of the launch configuration. This name must be unique within the scope of your AWS account.

Destructuring the Response

createLaunchConfigurationResponse :: CreateLaunchConfigurationResponse

Creates a value of CreateLaunchConfigurationResponse with the minimum fields required to make a request.

data CreateLaunchConfigurationResponse

See: createLaunchConfigurationResponse smart constructor.