Network.AWS.AutoScaling.CreateOrUpdateTags

Description

Creates or updates tags for the specified Auto Scaling group.

When you specify a tag with a key that already exists, the operation overwrites the previous tag definition, and you do not get an error message.

For more information, see Tagging Auto Scaling Groups and Instances in the Auto Scaling User Guide .

Creating a Request

createOrUpdateTags :: CreateOrUpdateTags

Creates a value of CreateOrUpdateTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coutTags - One or more tags.

data CreateOrUpdateTags

Contains the parameters for CreateOrUpdateTags.

See: createOrUpdateTags smart constructor.

Request Lenses

coutTags :: Lens' CreateOrUpdateTags [Tag]

One or more tags.

Destructuring the Response

createOrUpdateTagsResponse :: CreateOrUpdateTagsResponse

Creates a value of CreateOrUpdateTagsResponse with the minimum fields required to make a request.

data CreateOrUpdateTagsResponse

See: createOrUpdateTagsResponse smart constructor.