Network.AWS.AutoScaling.DeleteAutoScalingGroup

Description

Deletes the specified Auto Scaling group.

If the group has instances or scaling activities in progress, you must specify the option to force the deletion in order for it to succeed.

If the group has policies, deleting the group deletes the policies, the underlying alarm actions, and any alarm that no longer has an associated action.

To remove instances from the Auto Scaling group before deleting it, call DetachInstances with the list of instances and the option to decrement the desired capacity so that Auto Scaling does not launch replacement instances.

To terminate all instances before deleting the Auto Scaling group, call UpdateAutoScalingGroup and set the minimum size and desired capacity of the Auto Scaling group to zero.

Creating a Request

deleteAutoScalingGroup

Arguments

:: Text
dasgAutoScalingGroupName

-> DeleteAutoScalingGroup

Creates a value of DeleteAutoScalingGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasgForceDelete - Specifies that the group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any lifecycle actions associated with the group.

dasgAutoScalingGroupName - The name of the group to delete.

data DeleteAutoScalingGroup

Contains the parameters for DeleteAutoScalingGroup.

See: deleteAutoScalingGroup smart constructor.

Request Lenses

dasgForceDelete :: Lens' DeleteAutoScalingGroup (Maybe Bool)

Specifies that the group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any lifecycle actions associated with the group.

dasgAutoScalingGroupName :: Lens' DeleteAutoScalingGroup Text

The name of the group to delete.

Destructuring the Response

deleteAutoScalingGroupResponse :: DeleteAutoScalingGroupResponse

Creates a value of DeleteAutoScalingGroupResponse with the minimum fields required to make a request.

data DeleteAutoScalingGroupResponse

See: deleteAutoScalingGroupResponse smart constructor.