Network.AWS.AutoScaling.DeleteAutoScalingGroup

Description

Deletes the specified Auto Scaling group.

The group must have no instances and no scaling activities in progress.

To remove all instances before calling DeleteAutoScalingGroup, you can call UpdateAutoScalingGroup to set the minimum and maximum size of the AutoScalingGroup to zero.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DeleteAutoScalingGroup.html

Request

data DeleteAutoScalingGroup

Request constructor

deleteAutoScalingGroup

Arguments

:: Text
dasgAutoScalingGroupName

-> DeleteAutoScalingGroup

DeleteAutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

dasgAutoScalingGroupName :: Text

dasgForceDelete :: Maybe Bool

Request lenses

dasgAutoScalingGroupName :: Lens' DeleteAutoScalingGroup Text

The name of the group to delete.

dasgForceDelete :: Lens' DeleteAutoScalingGroup (Maybe Bool)

Specifies that the group will be deleted along with all instances associated with the group, without waiting for all instances to be terminated. This parameter also deletes any lifecycle actions associated with the group.

Response

data DeleteAutoScalingGroupResponse

Response constructor

deleteAutoScalingGroupResponse :: DeleteAutoScalingGroupResponse

DeleteAutoScalingGroupResponse constructor.