Network.AWS.AutoScaling.DeleteLaunchConfiguration

Description

Deletes the specified launch configuration.

The launch configuration must not be attached to an Auto Scaling group. When this call completes, the launch configuration is no longer available for use.

Creating a Request

deleteLaunchConfiguration

Arguments

:: Text
dlcLaunchConfigurationName

-> DeleteLaunchConfiguration

Creates a value of DeleteLaunchConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlcLaunchConfigurationName - The name of the launch configuration.

data DeleteLaunchConfiguration

Contains the parameters for DeleteLaunchConfiguration.

See: deleteLaunchConfiguration smart constructor.

Request Lenses

dlcLaunchConfigurationName :: Lens' DeleteLaunchConfiguration Text

The name of the launch configuration.

Destructuring the Response

deleteLaunchConfigurationResponse :: DeleteLaunchConfigurationResponse

Creates a value of DeleteLaunchConfigurationResponse with the minimum fields required to make a request.

data DeleteLaunchConfigurationResponse

See: deleteLaunchConfigurationResponse smart constructor.