Network.AWS.AutoScaling.DeleteLifecycleHook

Description

Deletes the specified lifecycle hook.

If there are any outstanding lifecycle actions, they are completed first (ABANDON for launching instances, CONTINUE for terminating instances).

Creating a Request

deleteLifecycleHook

Arguments

:: Text
delLifecycleHookName

-> Text
delAutoScalingGroupName

-> DeleteLifecycleHook

Creates a value of DeleteLifecycleHook with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delLifecycleHookName - The name of the lifecycle hook.

delAutoScalingGroupName - The name of the Auto Scaling group for the lifecycle hook.

data DeleteLifecycleHook

Contains the parameters for DeleteLifecycleHook.

See: deleteLifecycleHook smart constructor.

Request Lenses

delLifecycleHookName :: Lens' DeleteLifecycleHook Text

The name of the lifecycle hook.

delAutoScalingGroupName :: Lens' DeleteLifecycleHook Text

The name of the Auto Scaling group for the lifecycle hook.

Destructuring the Response

deleteLifecycleHookResponse

Arguments

:: Int
drsResponseStatus

-> DeleteLifecycleHookResponse

Creates a value of DeleteLifecycleHookResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsResponseStatus - -- | The response status code.

data DeleteLifecycleHookResponse

Contains the output of DeleteLifecycleHook.

See: deleteLifecycleHookResponse smart constructor.

Response Lenses

drsResponseStatus :: Lens' DeleteLifecycleHookResponse Int

- | The response status code.