Network.AWS.AutoScaling.DeleteNotificationConfiguration

Description

Deletes the specified notification.

Creating a Request

deleteNotificationConfiguration

Arguments

:: Text
dncAutoScalingGroupName

-> Text
dncTopicARN

-> DeleteNotificationConfiguration

Creates a value of DeleteNotificationConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dncAutoScalingGroupName - The name of the Auto Scaling group.

dncTopicARN - The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

data DeleteNotificationConfiguration

Contains the parameters for DeleteNotificationConfiguration.

See: deleteNotificationConfiguration smart constructor.

Request Lenses

dncAutoScalingGroupName :: Lens' DeleteNotificationConfiguration Text

The name of the Auto Scaling group.

dncTopicARN :: Lens' DeleteNotificationConfiguration Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

Destructuring the Response

deleteNotificationConfigurationResponse :: DeleteNotificationConfigurationResponse

Creates a value of DeleteNotificationConfigurationResponse with the minimum fields required to make a request.

data DeleteNotificationConfigurationResponse

See: deleteNotificationConfigurationResponse smart constructor.