Network.AWS.AutoScaling.DescribeAccountLimits

Description

Describes the current Auto Scaling resource limits for your AWS account.

For information about requesting an increase in these limits, see AWS Service Limits in the Amazon Web Services General Reference .

Creating a Request

describeAccountLimits :: DescribeAccountLimits

Creates a value of DescribeAccountLimits with the minimum fields required to make a request.

data DescribeAccountLimits

See: describeAccountLimits smart constructor.

Destructuring the Response

describeAccountLimitsResponse

Arguments

:: Int
dalrsResponseStatus

-> DescribeAccountLimitsResponse

Creates a value of DescribeAccountLimitsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dalrsNumberOfLaunchConfigurations - The current number of launch configurations for your AWS account.

dalrsNumberOfAutoScalingGroups - The current number of groups for your AWS account.

dalrsMaxNumberOfAutoScalingGroups - The maximum number of groups allowed for your AWS account. The default limit is 20 per region.

dalrsMaxNumberOfLaunchConfigurations - The maximum number of launch configurations allowed for your AWS account. The default limit is 100 per region.

dalrsResponseStatus - -- | The response status code.

data DescribeAccountLimitsResponse

Contains the parameters for DescribeAccountLimits.

See: describeAccountLimitsResponse smart constructor.

Response Lenses

dalrsNumberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The current number of launch configurations for your AWS account.

dalrsNumberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The current number of groups for your AWS account.

dalrsMaxNumberOfAutoScalingGroups :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of groups allowed for your AWS account. The default limit is 20 per region.

dalrsMaxNumberOfLaunchConfigurations :: Lens' DescribeAccountLimitsResponse (Maybe Int)

The maximum number of launch configurations allowed for your AWS account. The default limit is 100 per region.

dalrsResponseStatus :: Lens' DescribeAccountLimitsResponse Int

- | The response status code.