Network.AWS.AutoScaling.DescribeAutoScalingInstances

Description

Describes one or more Auto Scaling instances.

This operation returns paginated results.

Creating a Request

describeAutoScalingInstances :: DescribeAutoScalingInstances

Creates a value of DescribeAutoScalingInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasiNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dasiInstanceIds - The instances to describe; up to 50 instance IDs. If you omit this parameter, all Auto Scaling instances are described. If you specify an ID that does not exist, it is ignored with no error.

dasiMaxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

data DescribeAutoScalingInstances

Contains the parameters for DescribeAutoScalingInstances.

See: describeAutoScalingInstances smart constructor.

Request Lenses

dasiNextToken :: Lens' DescribeAutoScalingInstances (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dasiInstanceIds :: Lens' DescribeAutoScalingInstances [Text]

The instances to describe; up to 50 instance IDs. If you omit this parameter, all Auto Scaling instances are described. If you specify an ID that does not exist, it is ignored with no error.

dasiMaxRecords :: Lens' DescribeAutoScalingInstances (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Destructuring the Response

describeAutoScalingInstancesResponse

Arguments

:: Int
dasirsResponseStatus

-> DescribeAutoScalingInstancesResponse

Creates a value of DescribeAutoScalingInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasirsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dasirsAutoScalingInstances - The instances.

dasirsResponseStatus - -- | The response status code.

data DescribeAutoScalingInstancesResponse

Contains the output of DescribeAutoScalingInstances.

See: describeAutoScalingInstancesResponse smart constructor.

Response Lenses

dasirsNextToken :: Lens' DescribeAutoScalingInstancesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dasirsAutoScalingInstances :: Lens' DescribeAutoScalingInstancesResponse [AutoScalingInstanceDetails]

The instances.

dasirsResponseStatus :: Lens' DescribeAutoScalingInstancesResponse Int

- | The response status code.