Network.AWS.AutoScaling.DescribeLaunchConfigurations

Description

Describes one or more launch configurations.

This operation returns paginated results.

Creating a Request

describeLaunchConfigurations :: DescribeLaunchConfigurations

Creates a value of DescribeLaunchConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlcLaunchConfigurationNames - The launch configuration names. If you omit this parameter, all launch configurations are described.

dlcNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dlcMaxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

data DescribeLaunchConfigurations

Contains the parameters for DescribeLaunchConfigurations.

See: describeLaunchConfigurations smart constructor.

Request Lenses

dlcLaunchConfigurationNames :: Lens' DescribeLaunchConfigurations [Text]

The launch configuration names. If you omit this parameter, all launch configurations are described.

dlcNextToken :: Lens' DescribeLaunchConfigurations (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dlcMaxRecords :: Lens' DescribeLaunchConfigurations (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

Destructuring the Response

describeLaunchConfigurationsResponse

Arguments

:: Int
dlcrsResponseStatus

-> DescribeLaunchConfigurationsResponse

Creates a value of DescribeLaunchConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlcrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dlcrsResponseStatus - -- | The response status code.

dlcrsLaunchConfigurations - The launch configurations.

data DescribeLaunchConfigurationsResponse

Contains the output of DescribeLaunchConfigurations.

See: describeLaunchConfigurationsResponse smart constructor.

Response Lenses

dlcrsNextToken :: Lens' DescribeLaunchConfigurationsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dlcrsResponseStatus :: Lens' DescribeLaunchConfigurationsResponse Int

- | The response status code.

dlcrsLaunchConfigurations :: Lens' DescribeLaunchConfigurationsResponse [LaunchConfiguration]

The launch configurations.