Network.AWS.AutoScaling.DescribeLifecycleHooks

Description

Describes the lifecycle hooks for the specified Auto Scaling group.

Creating a Request

describeLifecycleHooks

Arguments

:: Text
dlhAutoScalingGroupName

-> DescribeLifecycleHooks

Creates a value of DescribeLifecycleHooks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlhLifecycleHookNames - The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.

dlhAutoScalingGroupName - The name of the group.

data DescribeLifecycleHooks

Contains the parameters for DescribeLifecycleHooks.

See: describeLifecycleHooks smart constructor.

Request Lenses

dlhLifecycleHookNames :: Lens' DescribeLifecycleHooks [Text]

The names of one or more lifecycle hooks. If you omit this parameter, all lifecycle hooks are described.

dlhAutoScalingGroupName :: Lens' DescribeLifecycleHooks Text

The name of the group.

Destructuring the Response

describeLifecycleHooksResponse

Arguments

:: Int
dlhrsResponseStatus

-> DescribeLifecycleHooksResponse

Creates a value of DescribeLifecycleHooksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlhrsLifecycleHooks - The lifecycle hooks for the specified group.

dlhrsResponseStatus - -- | The response status code.

data DescribeLifecycleHooksResponse

Contains the output of DescribeLifecycleHooks.

See: describeLifecycleHooksResponse smart constructor.

Response Lenses

dlhrsLifecycleHooks :: Lens' DescribeLifecycleHooksResponse [LifecycleHook]

The lifecycle hooks for the specified group.

dlhrsResponseStatus :: Lens' DescribeLifecycleHooksResponse Int

- | The response status code.