Network.AWS.AutoScaling.DescribeLoadBalancers

Description

Describes the load balancers for the specified Auto Scaling group.

Note that this operation describes only Classic Load Balancers. If you have Application Load Balancers, use DescribeLoadBalancerTargetGroups instead.

Creating a Request

describeLoadBalancers

Arguments

:: Text
dlbAutoScalingGroupName

-> DescribeLoadBalancers

Creates a value of DescribeLoadBalancers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dlbMaxRecords - The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

dlbAutoScalingGroupName - The name of the group.

data DescribeLoadBalancers

Contains the parameters for DescribeLoadBalancers.

See: describeLoadBalancers smart constructor.

Request Lenses

dlbNextToken :: Lens' DescribeLoadBalancers (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dlbMaxRecords :: Lens' DescribeLoadBalancers (Maybe Int)

The maximum number of items to return with this call. The default value is 50 and the maximum value is 100.

dlbAutoScalingGroupName :: Lens' DescribeLoadBalancers Text

The name of the group.

Destructuring the Response

describeLoadBalancersResponse

Arguments

:: Int
dlbrsResponseStatus

-> DescribeLoadBalancersResponse

Creates a value of DescribeLoadBalancersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dlbrsLoadBalancers - The load balancers.

dlbrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dlbrsResponseStatus - -- | The response status code.

data DescribeLoadBalancersResponse

Contains the output of DescribeLoadBalancers.

See: describeLoadBalancersResponse smart constructor.

Response Lenses

dlbrsLoadBalancers :: Lens' DescribeLoadBalancersResponse [LoadBalancerState]

The load balancers.

dlbrsNextToken :: Lens' DescribeLoadBalancersResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dlbrsResponseStatus :: Lens' DescribeLoadBalancersResponse Int

- | The response status code.