Network.AWS.AutoScaling.DescribeMetricCollectionTypes

Description

Describes the available CloudWatch metrics for Auto Scaling.

Note that the GroupStandbyInstances metric is not returned by default. You must explicitly request this metric when calling EnableMetricsCollection .

Creating a Request

describeMetricCollectionTypes :: DescribeMetricCollectionTypes

Creates a value of DescribeMetricCollectionTypes with the minimum fields required to make a request.

data DescribeMetricCollectionTypes

See: describeMetricCollectionTypes smart constructor.

Destructuring the Response

describeMetricCollectionTypesResponse

Arguments

:: Int
dmctrsResponseStatus

-> DescribeMetricCollectionTypesResponse

Creates a value of DescribeMetricCollectionTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmctrsMetrics - One or more metrics.

dmctrsGranularities - The granularities for the metrics.

dmctrsResponseStatus - -- | The response status code.

data DescribeMetricCollectionTypesResponse

Contains the output of DescribeMetricsCollectionTypes.

See: describeMetricCollectionTypesResponse smart constructor.

Response Lenses

dmctrsMetrics :: Lens' DescribeMetricCollectionTypesResponse [MetricCollectionType]

One or more metrics.

dmctrsGranularities :: Lens' DescribeMetricCollectionTypesResponse [MetricGranularityType]

The granularities for the metrics.

dmctrsResponseStatus :: Lens' DescribeMetricCollectionTypesResponse Int

- | The response status code.