Network.AWS.AutoScaling.DescribePolicies

Description

Describes the policies for the specified Auto Scaling group.

This operation returns paginated results.

Creating a Request

describePolicies :: DescribePolicies

Creates a value of DescribePolicies with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpsPolicyNames - One or more policy names or policy ARNs to be described. If you omit this parameter, all policy names are described. If an group name is provided, the results are limited to that group. This list is limited to 50 items. If you specify an unknown policy name, it is ignored with no error.

dpsNextToken - The token for the next set of items to return. (You received this token from a previous call.)

dpsAutoScalingGroupName - The name of the group.

dpsMaxRecords - The maximum number of items to be returned with each call. The default value is 50 and the maximum value is 100.

dpsPolicyTypes - One or more policy types. Valid values are SimpleScaling and StepScaling .

data DescribePolicies

Contains the parameters for DescribePolicies.

See: describePolicies smart constructor.

Request Lenses

dpsPolicyNames :: Lens' DescribePolicies [Text]

One or more policy names or policy ARNs to be described. If you omit this parameter, all policy names are described. If an group name is provided, the results are limited to that group. This list is limited to 50 items. If you specify an unknown policy name, it is ignored with no error.

dpsNextToken :: Lens' DescribePolicies (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dpsAutoScalingGroupName :: Lens' DescribePolicies (Maybe Text)

The name of the group.

dpsMaxRecords :: Lens' DescribePolicies (Maybe Int)

The maximum number of items to be returned with each call. The default value is 50 and the maximum value is 100.

dpsPolicyTypes :: Lens' DescribePolicies [Text]

One or more policy types. Valid values are SimpleScaling and StepScaling .

Destructuring the Response

describePoliciesResponse

Arguments

:: Int
dprsResponseStatus

-> DescribePoliciesResponse

Creates a value of DescribePoliciesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dprsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dprsScalingPolicies - The scaling policies.

dprsResponseStatus - -- | The response status code.

data DescribePoliciesResponse

Contains the output of DescribePolicies.

See: describePoliciesResponse smart constructor.

Response Lenses

dprsNextToken :: Lens' DescribePoliciesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dprsScalingPolicies :: Lens' DescribePoliciesResponse [ScalingPolicy]

The scaling policies.

dprsResponseStatus :: Lens' DescribePoliciesResponse Int

- | The response status code.