Network.AWS.AutoScaling.DescribePolicies

Description

Returns descriptions of what each policy does. This action supports pagination. If the response includes a token, there are more records available. To get the additional records, repeat the request with the response token as the NextToken parameter. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-test-asg &MaxRecords=20 &Version=2011-01-01 &Action=DescribePolicies &AUTHPARAMS arn:aws:autoscaling:us-east-1:803981987763:scalingPolicy:c322 761b-3172-4d56-9a21-0ed9d6161d67:autoScalingGroupNamemy-test-asg:policyNameMyScaleDownPolicy ChangeInCapacity -1 MyScaleDownPolicy my-test-asg 60 TestQueue arn:aws:cloudwatch:us-east-1:803981987763:alarm:TestQueue arn:aws:autoscaling:us-east-1:803981987763:scalingPolicy:c55a5cdd-9be0-435b-b60b-a8dd313159f5:autoScalingGroupNamemy-test-asg:policyNameMyScaleUpPolicy ChangeInCapacity 1 MyScaleUpPolicy my-test-asg 60 TestQueue arn:aws:cloudwatch:us-east-1:803981987763:alarm:TestQueue ec3bffad-b739-11e2-b38d-15fbEXAMPLE.

Request

data DescribePolicies

Request constructor

mkDescribePolicies :: DescribePolicies

Smart constructor for the minimum required parameters to construct a valid DescribePolicies request.

The fields accessible through corresponding lenses are:

AutoScalingGroupName :: Maybe Text

PolicyNames :: [Text]

NextToken :: Maybe Text

MaxRecords :: Maybe Integer

Request lenses

dp1AutoScalingGroupName :: Lens' DescribePolicies (Maybe Text)

The name of the Auto Scaling group.

dp1PolicyNames :: Lens' DescribePolicies [Text]

A list of policy names or policy ARNs to be described. If this list is omitted, all policy names are described. If an auto scaling group name is provided, the results are limited to that group. The list of requested policy names cannot contain more than 50 items. If unknown policy names are requested, they are ignored with no error.

dp1NextToken :: Lens' DescribePolicies (Maybe Text)

A string that is used to mark the start of the next batch of returned results for pagination.

dp1MaxRecords :: Lens' DescribePolicies (Maybe Integer)

The maximum number of policies that will be described with each call.

Response

data DescribePoliciesResponse

The PoliciesType data type.

Response constructor

mkDescribePoliciesResponse :: DescribePoliciesResponse

Smart constructor for the minimum required parameters to construct a valid DescribePoliciesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ScalingPolicies :: [ScalingPolicy]

NextToken :: Maybe Text

Response lenses

dprScalingPolicies :: Lens' DescribePoliciesResponse [ScalingPolicy]

A list of scaling policies.

dprNextToken :: Lens' DescribePoliciesResponse (Maybe Text)

A string that marks the start of the next batch of returned results.