Network.AWS.AutoScaling.DescribePolicies

Description

Describes the policies for the specified Auto Scaling group.

You can specify a maximum number of items to be returned with a single call. If there are more items to return, the call returns a token. To get the next set of items, repeat the call with the returned token in the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribePolicies.html

Request

data DescribePolicies

Request constructor

describePolicies :: DescribePolicies

DescribePolicies constructor.

The fields accessible through corresponding lenses are:

dp1AutoScalingGroupName :: Maybe Text

dp1MaxRecords :: Maybe Int

dp1NextToken :: Maybe Text

dp1PolicyNames :: [Text]

Request lenses

dp1AutoScalingGroupName :: Lens' DescribePolicies (Maybe Text)

The name of the group.

dp1MaxRecords :: Lens' DescribePolicies (Maybe Int)

The maximum number of items to be returned with each call.

dp1NextToken :: Lens' DescribePolicies (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.)

dp1PolicyNames :: Lens' DescribePolicies [Text]

One or more policy names or policy ARNs to be described. If you omit this list, all policy names are described. If an group name is provided, the results are limited to that group. This list is limited to 50 items. If you specify an unknown policy name, it is ignored with no error.

Response

data DescribePoliciesResponse

Response constructor

describePoliciesResponse :: DescribePoliciesResponse

DescribePoliciesResponse constructor.

The fields accessible through corresponding lenses are:

dprNextToken :: Maybe Text

dprScalingPolicies :: [ScalingPolicy]

Response lenses

dprNextToken :: Lens' DescribePoliciesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dprScalingPolicies :: Lens' DescribePoliciesResponse [ScalingPolicy]

The scaling policies.