Network.AWS.AutoScaling.DescribeScalingActivities

Description

Describes one or more scaling activities for the specified Auto Scaling group. If you omit the ActivityIds, the call returns all activities from the past six weeks. Activities are sorted by the start time. Activities still in progress appear first on the list. You can specify a maximum number of items to be returned with a single call. If there are more items to return, the call returns a token. To get the next set of items, repeat the call with the returned token in the NextToken parameter.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_DescribeScalingActivities.html

Request

data DescribeScalingActivities

Request constructor

describeScalingActivities :: DescribeScalingActivities

DescribeScalingActivities constructor.

The fields accessible through corresponding lenses are:

dsa2ActivityIds :: [Text]

dsa2AutoScalingGroupName :: Maybe Text

dsa2MaxRecords :: Maybe Int

dsa2NextToken :: Maybe Text

Request lenses

dsa2ActivityIds :: Lens' DescribeScalingActivities [Text]

A list containing the activity IDs of the desired scaling activities. If this list is omitted, all activities are described. If an AutoScalingGroupName is provided, the results are limited to that group. The list of requested activities cannot contain more than 50 items. If unknown activities are requested, they are ignored with no error.

dsa2AutoScalingGroupName :: Lens' DescribeScalingActivities (Maybe Text)

The name of the group.

dsa2MaxRecords :: Lens' DescribeScalingActivities (Maybe Int)

The maximum number of items to return with this call.

dsa2NextToken :: Lens' DescribeScalingActivities (Maybe Text)

The token for the next set of items to return. (You received this token from a previous call.).

Response

data DescribeScalingActivitiesResponse

Response constructor

describeScalingActivitiesResponse :: DescribeScalingActivitiesResponse

DescribeScalingActivitiesResponse constructor.

The fields accessible through corresponding lenses are:

dsar1Activities :: [Activity]

dsar1NextToken :: Maybe Text

Response lenses

dsar1Activities :: Lens' DescribeScalingActivitiesResponse [Activity]

The scaling activities.

dsar1NextToken :: Lens' DescribeScalingActivitiesResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.