Network.AWS.AutoScaling.DescribeScalingProcessTypes

Description

Describes the scaling process types for use with ResumeProcesses and SuspendProcesses .

Creating a Request

describeScalingProcessTypes :: DescribeScalingProcessTypes

Creates a value of DescribeScalingProcessTypes with the minimum fields required to make a request.

data DescribeScalingProcessTypes

See: describeScalingProcessTypes smart constructor.

Destructuring the Response

describeScalingProcessTypesResponse

Arguments

:: Int
dsptrsResponseStatus

-> DescribeScalingProcessTypesResponse

Creates a value of DescribeScalingProcessTypesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsptrsProcesses - The names of the process types.

dsptrsResponseStatus - -- | The response status code.

data DescribeScalingProcessTypesResponse

Contains the output of DescribeScalingProcessTypes.

See: describeScalingProcessTypesResponse smart constructor.

Response Lenses

dsptrsProcesses :: Lens' DescribeScalingProcessTypesResponse [ProcessType]

The names of the process types.

dsptrsResponseStatus :: Lens' DescribeScalingProcessTypesResponse Int

- | The response status code.