Network.AWS.AutoScaling.DisableMetricsCollection

Description

Disables group metrics for the specified Auto Scaling group.

Creating a Request

disableMetricsCollection

Arguments

:: Text
dmcAutoScalingGroupName

-> DisableMetricsCollection

Creates a value of DisableMetricsCollection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmcMetrics - One or more of the following metrics. If you omit this parameter, all metrics are disabled. * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

dmcAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the group.

data DisableMetricsCollection

Contains the parameters for DisableMetricsCollection.

See: disableMetricsCollection smart constructor.

Request Lenses

dmcMetrics :: Lens' DisableMetricsCollection [Text]

One or more of the following metrics. If you omit this parameter, all metrics are disabled. * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

dmcAutoScalingGroupName :: Lens' DisableMetricsCollection Text

The name or Amazon Resource Name (ARN) of the group.

Destructuring the Response

disableMetricsCollectionResponse :: DisableMetricsCollectionResponse

Creates a value of DisableMetricsCollectionResponse with the minimum fields required to make a request.

data DisableMetricsCollectionResponse

See: disableMetricsCollectionResponse smart constructor.