Network.AWS.AutoScaling.EnterStandby

Description

Moves the specified instances into Standby mode.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

Creating a Request

enterStandby

Arguments

:: Text
esAutoScalingGroupName

-> Bool
esShouldDecrementDesiredCapacity

-> EnterStandby

Creates a value of EnterStandby with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esInstanceIds - One or more instances to move into Standby mode. You must specify at least one instance ID.

esAutoScalingGroupName - The name of the Auto Scaling group.

esShouldDecrementDesiredCapacity - Specifies whether the instances moved to Standby mode count as part of the Auto Scaling group's desired capacity. If set, the desired capacity for the Auto Scaling group decrements by the number of instances moved to Standby mode.

data EnterStandby

Contains the parameters for EnteStandby.

See: enterStandby smart constructor.

Request Lenses

esInstanceIds :: Lens' EnterStandby [Text]

One or more instances to move into Standby mode. You must specify at least one instance ID.

esAutoScalingGroupName :: Lens' EnterStandby Text

The name of the Auto Scaling group.

esShouldDecrementDesiredCapacity :: Lens' EnterStandby Bool

Specifies whether the instances moved to Standby mode count as part of the Auto Scaling group's desired capacity. If set, the desired capacity for the Auto Scaling group decrements by the number of instances moved to Standby mode.

Destructuring the Response

enterStandbyResponse

Arguments

:: Int
ersResponseStatus

-> EnterStandbyResponse

Creates a value of EnterStandbyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ersActivities - The activities related to moving instances into Standby mode.

ersResponseStatus - -- | The response status code.

data EnterStandbyResponse

Contains the output of EnterStandby.

See: enterStandbyResponse smart constructor.

Response Lenses

ersActivities :: Lens' EnterStandbyResponse [Activity]

The activities related to moving instances into Standby mode.

ersResponseStatus :: Lens' EnterStandbyResponse Int

- | The response status code.