Network.AWS.AutoScaling.ExecutePolicy

Description

Executes the specified policy.

Creating a Request

executePolicy

Arguments

:: Text
epPolicyName

-> ExecutePolicy

Creates a value of ExecutePolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

epHonorCooldown - If this parameter is true, Auto Scaling waits for the cooldown period to complete before executing the policy. Otherwise, Auto Scaling executes the policy without waiting for the cooldown period to complete. This parameter is not supported if the policy type is StepScaling . For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

epMetricValue - The metric value to compare to BreachThreshold . This enables you to execute a policy of type StepScaling and determine which step adjustment to use. For example, if the breach threshold is 50 and you want to use a step adjustment with a lower bound of 0 and an upper bound of 10, you can set the metric value to 59. If you specify a metric value that doesn't correspond to a step adjustment for the policy, the call returns an error. This parameter is required if the policy type is StepScaling and not supported otherwise.

epAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the Auto Scaling group.

epBreachThreshold - The breach threshold for the alarm. This parameter is required if the policy type is StepScaling and not supported otherwise.

epPolicyName - The name or ARN of the policy.

data ExecutePolicy

Contains the parameters for ExecutePolicy.

See: executePolicy smart constructor.

Request Lenses

epHonorCooldown :: Lens' ExecutePolicy (Maybe Bool)

If this parameter is true, Auto Scaling waits for the cooldown period to complete before executing the policy. Otherwise, Auto Scaling executes the policy without waiting for the cooldown period to complete. This parameter is not supported if the policy type is StepScaling . For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

epMetricValue :: Lens' ExecutePolicy (Maybe Double)

The metric value to compare to BreachThreshold . This enables you to execute a policy of type StepScaling and determine which step adjustment to use. For example, if the breach threshold is 50 and you want to use a step adjustment with a lower bound of 0 and an upper bound of 10, you can set the metric value to 59. If you specify a metric value that doesn't correspond to a step adjustment for the policy, the call returns an error. This parameter is required if the policy type is StepScaling and not supported otherwise.

epAutoScalingGroupName :: Lens' ExecutePolicy (Maybe Text)

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

epBreachThreshold :: Lens' ExecutePolicy (Maybe Double)

The breach threshold for the alarm. This parameter is required if the policy type is StepScaling and not supported otherwise.

epPolicyName :: Lens' ExecutePolicy Text

The name or ARN of the policy.

Destructuring the Response

executePolicyResponse :: ExecutePolicyResponse

Creates a value of ExecutePolicyResponse with the minimum fields required to make a request.

data ExecutePolicyResponse

See: executePolicyResponse smart constructor.