Network.AWS.AutoScaling.PutNotificationConfiguration

Description

Configures an Auto Scaling group to send notifications when specified events take place. Subscribers to the specified topic can have messages delivered to an endpoint such as a web server or an email address.

This configuration overwrites any existing configuration.

For more information see Getting SNS Notifications When Your Auto Scaling Group Scales in the Auto Scaling User Guide .

Creating a Request

putNotificationConfiguration

Arguments

:: Text
pncAutoScalingGroupName

-> Text
pncTopicARN

-> PutNotificationConfiguration

Creates a value of PutNotificationConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pncAutoScalingGroupName - The name of the Auto Scaling group.

pncTopicARN - The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

pncNotificationTypes - The type of event that will cause the notification to be sent. For details about notification types supported by Auto Scaling, see DescribeAutoScalingNotificationTypes .

data PutNotificationConfiguration

Contains the parameters for PutNotificationConfiguration.

See: putNotificationConfiguration smart constructor.

Request Lenses

pncAutoScalingGroupName :: Lens' PutNotificationConfiguration Text

The name of the Auto Scaling group.

pncTopicARN :: Lens' PutNotificationConfiguration Text

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

pncNotificationTypes :: Lens' PutNotificationConfiguration [Text]

The type of event that will cause the notification to be sent. For details about notification types supported by Auto Scaling, see DescribeAutoScalingNotificationTypes .

Destructuring the Response

putNotificationConfigurationResponse :: PutNotificationConfigurationResponse

Creates a value of PutNotificationConfigurationResponse with the minimum fields required to make a request.

data PutNotificationConfigurationResponse

See: putNotificationConfigurationResponse smart constructor.