Network.AWS.AutoScaling.PutScalingPolicy

Description

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameters you want to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request.

If you exceed your maximum limit of step adjustments, which by default is 20 per region, the call fails. For information about updating this limit, see AWS Service Limits in the Amazon Web Services General Reference .

Creating a Request

putScalingPolicy

Arguments

:: Text
pspAutoScalingGroupName

-> Text
pspPolicyName

-> PutScalingPolicy

Creates a value of PutScalingPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pspMinAdjustmentStep - Available for backward compatibility. Use MinAdjustmentMagnitude instead.

pspEstimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. The default is to use the value specified for the default cooldown period for the group. This parameter is supported if the policy type is StepScaling or TargetTrackingScaling .

pspPolicyType - The policy type. The valid values are SimpleScaling , StepScaling , and TargetTrackingScaling . If the policy type is null, the value is treated as SimpleScaling .

pspStepAdjustments - A set of adjustments that enable you to scale based on the size of the alarm breach. This parameter is required if the policy type is StepScaling and not supported otherwise.

pspTargetTrackingConfiguration - The configuration of a target tracking policy. This parameter is required if the policy type is TargetTrackingScaling and not supported otherwise.

pspAdjustmentType - The adjustment type. The valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity . This parameter is supported if the policy type is SimpleScaling or StepScaling . For more information, see Dynamic Scaling in the Auto Scaling User Guide .

pspScalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. This parameter is required if the policy type is SimpleScaling and not supported otherwise.

pspCooldown - The amount of time, in seconds, after a scaling activity completes and before the next scaling activity can start. If this parameter is not specified, the default cooldown period for the group applies. This parameter is supported if the policy type is SimpleScaling . For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

pspMetricAggregationType - The aggregation type for the CloudWatch metrics. The valid values are Minimum , Maximum , and Average . If the aggregation type is null, the value is treated as Average . This parameter is supported if the policy type is StepScaling .

pspMinAdjustmentMagnitude - The minimum number of instances to scale. If the value of AdjustmentType is PercentChangeInCapacity , the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least this many instances. Otherwise, the error is ValidationError . This parameter is supported if the policy type is SimpleScaling or StepScaling .

pspAutoScalingGroupName - The name or ARN of the group.

pspPolicyName - The name of the policy.

data PutScalingPolicy

Contains the parameters for PutScalingPolicy.

See: putScalingPolicy smart constructor.

Request Lenses

pspMinAdjustmentStep :: Lens' PutScalingPolicy (Maybe Int)

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

pspEstimatedInstanceWarmup :: Lens' PutScalingPolicy (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics. The default is to use the value specified for the default cooldown period for the group. This parameter is supported if the policy type is StepScaling or TargetTrackingScaling .

pspPolicyType :: Lens' PutScalingPolicy (Maybe Text)

The policy type. The valid values are SimpleScaling , StepScaling , and TargetTrackingScaling . If the policy type is null, the value is treated as SimpleScaling .

pspStepAdjustments :: Lens' PutScalingPolicy [StepAdjustment]

A set of adjustments that enable you to scale based on the size of the alarm breach. This parameter is required if the policy type is StepScaling and not supported otherwise.

pspTargetTrackingConfiguration :: Lens' PutScalingPolicy (Maybe TargetTrackingConfiguration)

The configuration of a target tracking policy. This parameter is required if the policy type is TargetTrackingScaling and not supported otherwise.

pspAdjustmentType :: Lens' PutScalingPolicy (Maybe Text)

The adjustment type. The valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity . This parameter is supported if the policy type is SimpleScaling or StepScaling . For more information, see Dynamic Scaling in the Auto Scaling User Guide .

pspScalingAdjustment :: Lens' PutScalingPolicy (Maybe Int)

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity. This parameter is required if the policy type is SimpleScaling and not supported otherwise.

pspCooldown :: Lens' PutScalingPolicy (Maybe Int)

The amount of time, in seconds, after a scaling activity completes and before the next scaling activity can start. If this parameter is not specified, the default cooldown period for the group applies. This parameter is supported if the policy type is SimpleScaling . For more information, see Auto Scaling Cooldowns in the Auto Scaling User Guide .

pspMetricAggregationType :: Lens' PutScalingPolicy (Maybe Text)

The aggregation type for the CloudWatch metrics. The valid values are Minimum , Maximum , and Average . If the aggregation type is null, the value is treated as Average . This parameter is supported if the policy type is StepScaling .

pspMinAdjustmentMagnitude :: Lens' PutScalingPolicy (Maybe Int)

The minimum number of instances to scale. If the value of AdjustmentType is PercentChangeInCapacity , the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least this many instances. Otherwise, the error is ValidationError . This parameter is supported if the policy type is SimpleScaling or StepScaling .

pspAutoScalingGroupName :: Lens' PutScalingPolicy Text

The name or ARN of the group.

pspPolicyName :: Lens' PutScalingPolicy Text

The name of the policy.

Destructuring the Response

putScalingPolicyResponse

Arguments

:: Int
psprsResponseStatus

-> PutScalingPolicyResponse

Creates a value of PutScalingPolicyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psprsPolicyARN - The Amazon Resource Name (ARN) of the policy.

psprsAlarms - The CloudWatch alarms created for the target tracking policy. This parameter will be empty if the policy type is anything other than TargetTrackingScaling .

psprsResponseStatus - -- | The response status code.

data PutScalingPolicyResponse

Contains the output of PutScalingPolicy.

See: putScalingPolicyResponse smart constructor.

Response Lenses

psprsPolicyARN :: Lens' PutScalingPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

psprsAlarms :: Lens' PutScalingPolicyResponse [Alarm]

The CloudWatch alarms created for the target tracking policy. This parameter will be empty if the policy type is anything other than TargetTrackingScaling .

psprsResponseStatus :: Lens' PutScalingPolicyResponse Int

- | The response status code.