Network.AWS.AutoScaling.PutScalingPolicy

Description

Creates or updates a policy for an Auto Scaling group. To update an existing policy, use the existing policy name and set the parameters you want to change. Any existing parameter not changed in an update to an existing policy is not changed in this update request.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutScalingPolicy.html

Request

data PutScalingPolicy

Request constructor

putScalingPolicy

Arguments

:: Text
pspAutoScalingGroupName

-> Text
pspPolicyName

-> Int
pspScalingAdjustment

-> Text
pspAdjustmentType

-> PutScalingPolicy

PutScalingPolicy constructor.

The fields accessible through corresponding lenses are:

pspAdjustmentType :: Text

pspAutoScalingGroupName :: Text

pspCooldown :: Maybe Int

pspMinAdjustmentStep :: Maybe Int

pspPolicyName :: Text

pspScalingAdjustment :: Int

Request lenses

pspAdjustmentType :: Lens' PutScalingPolicy Text

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

For more information, see Dynamic Scaling in the Auto Scaling Developer Guide .

pspAutoScalingGroupName :: Lens' PutScalingPolicy Text

The name or ARN of the group.

pspCooldown :: Lens' PutScalingPolicy (Maybe Int)

The amount of time, in seconds, after a scaling activity completes and before the next scaling activity can start.

For more information, see Understanding Auto Scaling Cooldowns in the AutoScaling Developer Guide.

pspMinAdjustmentStep :: Lens' PutScalingPolicy (Maybe Int)

Used with AdjustmentType with the value PercentChangeInCapacity, the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least the number of instances specified in the value.

You will get a ValidationError if you use MinAdjustmentStep on a policy with an AdjustmentType other than PercentChangeInCapacity.

pspPolicyName :: Lens' PutScalingPolicy Text

The name of the policy.

pspScalingAdjustment :: Lens' PutScalingPolicy Int

The number of instances by which to scale. AdjustmentType determines the interpretation of this number (e.g., as an absolute number or as a percentage of the existing Auto Scaling group size). A positive increment adds to the current capacity and a negative value removes from the current capacity.

Response

data PutScalingPolicyResponse

Response constructor

putScalingPolicyResponse :: PutScalingPolicyResponse

PutScalingPolicyResponse constructor.

The fields accessible through corresponding lenses are:

psprPolicyARN :: Maybe Text

Response lenses

psprPolicyARN :: Lens' PutScalingPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the policy.