Network.AWS.AutoScaling.PutScheduledUpdateGroupAction

Description

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged.

For more information, see Scheduled Scaling in the Auto Scaling User Guide .

Creating a Request

putScheduledUpdateGroupAction

Arguments

:: Text
psugaAutoScalingGroupName

-> Text
psugaScheduledActionName

-> PutScheduledUpdateGroupAction

Creates a value of PutScheduledUpdateGroupAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psugaStartTime - The time for this action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only (for example, 2014-06-01T00:00:00Z ). If you specify Recurrence and StartTime , Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence. If you try to schedule your action in the past, Auto Scaling returns an error message.

psugaTime - This parameter is deprecated.

psugaMaxSize - The maximum size for the Auto Scaling group.

psugaRecurrence - The recurring schedule for this action, in Unix cron syntax format. For more information, see Cron in Wikipedia.

psugaDesiredCapacity - The number of EC2 instances that should be running in the group.

psugaMinSize - The minimum size for the Auto Scaling group.

psugaEndTime - The time for the recurring schedule to end. Auto Scaling does not perform the action after this time.

psugaAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the Auto Scaling group.

psugaScheduledActionName - The name of this scaling action.

data PutScheduledUpdateGroupAction

Contains the parameters for PutScheduledUpdateGroupAction.

See: putScheduledUpdateGroupAction smart constructor.

Request Lenses

psugaStartTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The time for this action to start, in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only (for example, 2014-06-01T00:00:00Z ). If you specify Recurrence and StartTime , Auto Scaling performs the action at this time, and then performs the action based on the specified recurrence. If you try to schedule your action in the past, Auto Scaling returns an error message.

psugaTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

This parameter is deprecated.

psugaMaxSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The maximum size for the Auto Scaling group.

psugaRecurrence :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

The recurring schedule for this action, in Unix cron syntax format. For more information, see Cron in Wikipedia.

psugaDesiredCapacity :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The number of EC2 instances that should be running in the group.

psugaMinSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The minimum size for the Auto Scaling group.

psugaEndTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The time for the recurring schedule to end. Auto Scaling does not perform the action after this time.

psugaAutoScalingGroupName :: Lens' PutScheduledUpdateGroupAction Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

psugaScheduledActionName :: Lens' PutScheduledUpdateGroupAction Text

The name of this scaling action.

Destructuring the Response

putScheduledUpdateGroupActionResponse :: PutScheduledUpdateGroupActionResponse

Creates a value of PutScheduledUpdateGroupActionResponse with the minimum fields required to make a request.

data PutScheduledUpdateGroupActionResponse

See: putScheduledUpdateGroupActionResponse smart constructor.