Network.AWS.AutoScaling.PutScheduledUpdateGroupAction

Description

Creates or updates a scheduled scaling action for an Auto Scaling group. When updating a scheduled scaling action, if you leave a parameter unspecified, the corresponding value remains unchanged in the affected Auto Scaling group.

For more information, see Scheduled Scaling in the Auto Scaling DeveloperGuide.

Auto Scaling supports the date and time expressed in YYYY-MM-DDThh:mm:ssZ format in UTC/GMT only.

http://docs.aws.amazon.com/AutoScaling/latest/APIReference/API_PutScheduledUpdateGroupAction.html

Request

data PutScheduledUpdateGroupAction

Request constructor

putScheduledUpdateGroupAction

Arguments

:: Text
psugaAutoScalingGroupName

-> Text
psugaScheduledActionName

-> PutScheduledUpdateGroupAction

PutScheduledUpdateGroupAction constructor.

The fields accessible through corresponding lenses are:

psugaAutoScalingGroupName :: Text

psugaDesiredCapacity :: Maybe Int

psugaEndTime :: Maybe UTCTime

psugaMaxSize :: Maybe Int

psugaMinSize :: Maybe Int

psugaRecurrence :: Maybe Text

psugaScheduledActionName :: Text

psugaStartTime :: Maybe UTCTime

psugaTime :: Maybe UTCTime

Request lenses

psugaAutoScalingGroupName :: Lens' PutScheduledUpdateGroupAction Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

psugaDesiredCapacity :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The number of Amazon EC2 instances that should be running in the group.

psugaEndTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The time for this action to end.

psugaMaxSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The maximum size for the Auto Scaling group.

psugaMinSize :: Lens' PutScheduledUpdateGroupAction (Maybe Int)

The minimum size for the new Auto Scaling group.

psugaRecurrence :: Lens' PutScheduledUpdateGroupAction (Maybe Text)

The time when recurring future actions will start. Start time is specified by the user following the Unix cron syntax format. For information about cron syntax, go to Wikipedia, The Free Encyclopedia.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

psugaScheduledActionName :: Lens' PutScheduledUpdateGroupAction Text

The name of this scaling action.

psugaStartTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

The time for this action to start, as in '--start-time 2010-06-01T00:00:00Z'.

If you try to schedule your action in the past, Auto Scaling returns an error message.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

psugaTime :: Lens' PutScheduledUpdateGroupAction (Maybe UTCTime)

Time is deprecated.

The time for this action to start. Time is an alias for StartTime and can be specified instead of StartTime, or vice versa. If both Time and StartTime are specified, their values should be identical. Otherwise, PutScheduledUpdateGroupAction will return an error.

Response

data PutScheduledUpdateGroupActionResponse

Response constructor

putScheduledUpdateGroupActionResponse :: PutScheduledUpdateGroupActionResponse

PutScheduledUpdateGroupActionResponse constructor.