Network.AWS.AutoScaling.RecordLifecycleActionHeartbeat

Description

Records a heartbeat for the lifecycle action associated with the specified token or instance. This extends the timeout by the length of time defined using PutLifecycleHook .

This step is a part of the procedure for adding a lifecycle hook to an Auto Scaling group:

(Optional) Create a Lambda function and a rule that allows CloudWatch Events to invoke your Lambda function when Auto Scaling launches or terminates instances.

(Optional) Create a notification target and an IAM role. The target can be either an Amazon SQS queue or an Amazon SNS topic. The role allows Auto Scaling to publish lifecycle notifications to the target.

Create the lifecycle hook. Specify whether the hook is used when the instances launch or terminate.

If you need more time, record the lifecycle action heartbeat to keep the instance in a pending state.

If you finish before the timeout period ends, complete the lifecycle action.

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

Creating a Request

recordLifecycleActionHeartbeat

Arguments

:: Text
rlahLifecycleHookName

-> Text
rlahAutoScalingGroupName

-> RecordLifecycleActionHeartbeat

Creates a value of RecordLifecycleActionHeartbeat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rlahInstanceId - The ID of the instance.

rlahLifecycleActionToken - A token that uniquely identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

rlahLifecycleHookName - The name of the lifecycle hook.

rlahAutoScalingGroupName - The name of the Auto Scaling group for the hook.

data RecordLifecycleActionHeartbeat

Contains the parameters for RecordLifecycleActionHeartbeat.

See: recordLifecycleActionHeartbeat smart constructor.

Request Lenses

rlahInstanceId :: Lens' RecordLifecycleActionHeartbeat (Maybe Text)

The ID of the instance.

rlahLifecycleActionToken :: Lens' RecordLifecycleActionHeartbeat (Maybe Text)

A token that uniquely identifies a specific lifecycle action associated with an instance. Auto Scaling sends this token to the notification target you specified when you created the lifecycle hook.

rlahLifecycleHookName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the lifecycle hook.

rlahAutoScalingGroupName :: Lens' RecordLifecycleActionHeartbeat Text

The name of the Auto Scaling group for the hook.

Destructuring the Response

recordLifecycleActionHeartbeatResponse

Arguments

:: Int
rlahrsResponseStatus

-> RecordLifecycleActionHeartbeatResponse

Creates a value of RecordLifecycleActionHeartbeatResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rlahrsResponseStatus - -- | The response status code.

data RecordLifecycleActionHeartbeatResponse

Contains the output of RecordLifecycleActionHeartBeat.

See: recordLifecycleActionHeartbeatResponse smart constructor.

Response Lenses

rlahrsResponseStatus :: Lens' RecordLifecycleActionHeartbeatResponse Int

- | The response status code.