Network.AWS.AutoScaling.ResumeProcesses

Description

Resumes the specified suspended Auto Scaling processes, or all suspended process, for the specified Auto Scaling group.

For more information, see Suspending and Resuming Auto Scaling Processes in the Auto Scaling User Guide .

Creating a Request

resumeProcesses

Arguments

:: Text
rpAutoScalingGroupName

-> ResumeProcesses

Creates a value of ResumeProcesses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpScalingProcesses - One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

rpAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the Auto Scaling group.

data ResumeProcesses

Contains the parameters for SuspendProcesses and ResumeProcesses.

See: resumeProcesses smart constructor.

Request Lenses

rpScalingProcesses :: Lens' ResumeProcesses [Text]

One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

rpAutoScalingGroupName :: Lens' ResumeProcesses Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

Destructuring the Response

resumeProcessesResponse :: ResumeProcessesResponse

Creates a value of ResumeProcessesResponse with the minimum fields required to make a request.

data ResumeProcessesResponse

See: resumeProcessesResponse smart constructor.