Network.AWS.AutoScaling.SetInstanceProtection

Description

Updates the instance protection settings of the specified instances.

For more information, see Instance Protection in the Auto Scaling User Guide .

Creating a Request

setInstanceProtection

Arguments

:: Text
sipAutoScalingGroupName

-> Bool
sipProtectedFromScaleIn

-> SetInstanceProtection

Creates a value of SetInstanceProtection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sipInstanceIds - One or more instance IDs.

sipAutoScalingGroupName - The name of the group.

sipProtectedFromScaleIn - Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

data SetInstanceProtection

Contains the parameters for SetInstanceProtection.

See: setInstanceProtection smart constructor.

Request Lenses

sipInstanceIds :: Lens' SetInstanceProtection [Text]

One or more instance IDs.

sipAutoScalingGroupName :: Lens' SetInstanceProtection Text

The name of the group.

sipProtectedFromScaleIn :: Lens' SetInstanceProtection Bool

Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

Destructuring the Response

setInstanceProtectionResponse

Arguments

:: Int
siprsResponseStatus

-> SetInstanceProtectionResponse

Creates a value of SetInstanceProtectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siprsResponseStatus - -- | The response status code.

data SetInstanceProtectionResponse

Contains the output of SetInstanceProtection.

See: setInstanceProtectionResponse smart constructor.

Response Lenses

siprsResponseStatus :: Lens' SetInstanceProtectionResponse Int

- | The response status code.