Network.AWS.AutoScaling.TerminateInstanceInAutoScalingGroup

Description

Terminates the specified instance and optionally adjusts the desired group size.

This call simply makes a termination request. The instance is not terminated immediately.

Creating a Request

terminateInstanceInAutoScalingGroup

Arguments

:: Text
tiiasgInstanceId

-> Bool
tiiasgShouldDecrementDesiredCapacity

-> TerminateInstanceInAutoScalingGroup

Creates a value of TerminateInstanceInAutoScalingGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiiasgInstanceId - The ID of the instance.

tiiasgShouldDecrementDesiredCapacity - If true , terminating the instance also decrements the size of the Auto Scaling group.

data TerminateInstanceInAutoScalingGroup

Contains the parameters for TerminateInstanceInAutoScalingGroup.

See: terminateInstanceInAutoScalingGroup smart constructor.

Request Lenses

tiiasgInstanceId :: Lens' TerminateInstanceInAutoScalingGroup Text

The ID of the instance.

tiiasgShouldDecrementDesiredCapacity :: Lens' TerminateInstanceInAutoScalingGroup Bool

If true , terminating the instance also decrements the size of the Auto Scaling group.

Destructuring the Response

terminateInstanceInAutoScalingGroupResponse

Arguments

:: Int
tiiasgrsResponseStatus

-> TerminateInstanceInAutoScalingGroupResponse

Creates a value of TerminateInstanceInAutoScalingGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiiasgrsActivity - A scaling activity.

tiiasgrsResponseStatus - -- | The response status code.

data TerminateInstanceInAutoScalingGroupResponse

Contains the output of TerminateInstancesInAutoScalingGroup.

See: terminateInstanceInAutoScalingGroupResponse smart constructor.

Response Lenses

tiiasgrsActivity :: Lens' TerminateInstanceInAutoScalingGroupResponse (Maybe Activity)

A scaling activity.

tiiasgrsResponseStatus :: Lens' TerminateInstanceInAutoScalingGroupResponse Int

- | The response status code.