Network.AWS.AutoScaling.Types

Description

Service Configuration

autoScaling :: Service

API version 2011-01-01 of the Amazon Auto Scaling SDK configuration.

Errors

_AlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

You already have an Auto Scaling group or launch configuration with this name.

_LimitExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

You have already reached a limit for your Auto Scaling resources (for example, groups, launch configurations, or lifecycle hooks). For more information, see DescribeAccountLimits .

_ResourceInUseFault :: AsError a => Getting (First ServiceError) a ServiceError

The operation can't be performed because the resource is in use.

_InvalidNextToken :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken value is not valid.

_ScalingActivityInProgressFault :: AsError a => Getting (First ServiceError) a ServiceError

The operation can't be performed because there are scaling activities in progress.

_ResourceContentionFault :: AsError a => Getting (First ServiceError) a ServiceError

You already have a pending update to an Auto Scaling resource (for example, a group, instance, or load balancer).

LifecycleState

data LifecycleState

Constructors

Detached

Detaching

EnteringStandby

InService

Pending

PendingProceed

PendingWait

Quarantined

Standby

Terminated

Terminating

TerminatingProceed

TerminatingWait

MetricStatistic

data MetricStatistic

Constructors

Average

Maximum

Minimum

SampleCount

Sum

MetricType

data MetricType

Constructors

ALBRequestCountPerTarget

ASGAverageCPUUtilization

ASGAverageNetworkIn

ASGAverageNetworkOut

ScalingActivityStatusCode

data ScalingActivityStatusCode

Constructors

Cancelled

Failed

InProgress

MidLifecycleAction

PendingSpotBidPlacement

PreInService

Successful

WaitingForELBConnectionDraining

WaitingForInstanceId

WaitingForInstanceWarmup

WaitingForSpotInstanceId

WaitingForSpotInstanceRequestId

Activity

data Activity

Describes scaling activity, which is a long-running process that represents a change to your Auto Scaling group, such as changing its size or replacing an instance.

See: activity smart constructor.

activity

Arguments

:: Text
aActivityId

-> Text
aAutoScalingGroupName

-> Text
aCause

-> UTCTime
aStartTime

-> ScalingActivityStatusCode
aStatusCode

-> Activity

Creates a value of Activity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aProgress - A value between 0 and 100 that indicates the progress of the activity.

aStatusMessage - A friendly, more verbose description of the activity status.

aEndTime - The end time of the activity.

aDetails - The details about the activity.

aDescription - A friendly, more verbose description of the activity.

aActivityId - The ID of the activity.

aAutoScalingGroupName - The name of the Auto Scaling group.

aCause - The reason the activity began.

aStartTime - The start time of the activity.

aStatusCode - The current status of the activity.

aProgress :: Lens' Activity (Maybe Int)

A value between 0 and 100 that indicates the progress of the activity.

aStatusMessage :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the activity status.

aEndTime :: Lens' Activity (Maybe UTCTime)

The end time of the activity.

aDetails :: Lens' Activity (Maybe Text)

The details about the activity.

aDescription :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the activity.

aActivityId :: Lens' Activity Text

The ID of the activity.

aAutoScalingGroupName :: Lens' Activity Text

The name of the Auto Scaling group.

aCause :: Lens' Activity Text

The reason the activity began.

aStartTime :: Lens' Activity UTCTime

The start time of the activity.

aStatusCode :: Lens' Activity ScalingActivityStatusCode

The current status of the activity.

AdjustmentType

data AdjustmentType

Describes a policy adjustment type.

For more information, see Dynamic Scaling in the Auto Scaling User Guide .

See: adjustmentType smart constructor.

adjustmentType :: AdjustmentType

Creates a value of AdjustmentType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atAdjustmentType - The policy adjustment type. The valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity .

atAdjustmentType :: Lens' AdjustmentType (Maybe Text)

The policy adjustment type. The valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity .

Alarm

data Alarm

Describes an alarm.

See: alarm smart constructor.

alarm :: Alarm

Creates a value of Alarm with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aAlarmName - The name of the alarm.

aAlarmARN - The Amazon Resource Name (ARN) of the alarm.

aAlarmName :: Lens' Alarm (Maybe Text)

The name of the alarm.

aAlarmARN :: Lens' Alarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

AutoScalingGroup

data AutoScalingGroup

Describes an Auto Scaling group.

See: autoScalingGroup smart constructor.

autoScalingGroup

Arguments

:: Text
asgAutoScalingGroupName

-> Int
asgMinSize

-> Int
asgMaxSize

-> Int
asgDesiredCapacity

-> Int
asgDefaultCooldown

-> NonEmpty Text
asgAvailabilityZones

-> Text
asgHealthCheckType

-> UTCTime
asgCreatedTime

-> AutoScalingGroup

Creates a value of AutoScalingGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asgStatus - The current state of the group when DeleteAutoScalingGroup is in progress.

asgTerminationPolicies - The termination policies for the group.

asgHealthCheckGracePeriod - The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

asgNewInstancesProtectedFromScaleIn - Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

asgVPCZoneIdentifier - One or more subnet IDs, if applicable, separated by commas. If you specify VPCZoneIdentifier and AvailabilityZones , ensure that the Availability Zones of the subnets match the values for AvailabilityZones .

asgTargetGroupARNs - The Amazon Resource Names (ARN) of the target groups for your load balancer.

asgEnabledMetrics - The metrics enabled for the group.

asgLaunchConfigurationName - The name of the associated launch configuration.

asgInstances - The EC2 instances associated with the group.

asgAutoScalingGroupARN - The Amazon Resource Name (ARN) of the group.

asgPlacementGroup - The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide .

asgSuspendedProcesses - The suspended processes associated with the group.

asgLoadBalancerNames - One or more load balancers associated with the group.

asgTags - The tags for the group.

asgAutoScalingGroupName - The name of the group.

asgMinSize - The minimum size of the group.

asgMaxSize - The maximum size of the group.

asgDesiredCapacity - The desired size of the group.

asgDefaultCooldown - The amount of time, in seconds, after a scaling activity completes before another scaling activity can start.

asgAvailabilityZones - One or more Availability Zones for the group.

asgHealthCheckType - The service to use for the health checks. The valid values are EC2 and ELB .

asgCreatedTime - The date and time the group was created.

asgStatus :: Lens' AutoScalingGroup (Maybe Text)

The current state of the group when DeleteAutoScalingGroup is in progress.

asgTerminationPolicies :: Lens' AutoScalingGroup [Text]

The termination policies for the group.

asgHealthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Int)

The amount of time, in seconds, that Auto Scaling waits before checking the health status of an EC2 instance that has come into service.

asgNewInstancesProtectedFromScaleIn :: Lens' AutoScalingGroup (Maybe Bool)

Indicates whether newly launched instances are protected from termination by Auto Scaling when scaling in.

asgVPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

One or more subnet IDs, if applicable, separated by commas. If you specify VPCZoneIdentifier and AvailabilityZones , ensure that the Availability Zones of the subnets match the values for AvailabilityZones .

asgTargetGroupARNs :: Lens' AutoScalingGroup [Text]

The Amazon Resource Names (ARN) of the target groups for your load balancer.

asgEnabledMetrics :: Lens' AutoScalingGroup [EnabledMetric]

The metrics enabled for the group.

asgLaunchConfigurationName :: Lens' AutoScalingGroup (Maybe Text)

The name of the associated launch configuration.

asgInstances :: Lens' AutoScalingGroup [Instance]

The EC2 instances associated with the group.

asgAutoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the group.

asgPlacementGroup :: Lens' AutoScalingGroup (Maybe Text)

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups in the Amazon Elastic Compute Cloud User Guide .

asgSuspendedProcesses :: Lens' AutoScalingGroup [SuspendedProcess]

The suspended processes associated with the group.

asgLoadBalancerNames :: Lens' AutoScalingGroup [Text]

One or more load balancers associated with the group.

asgTags :: Lens' AutoScalingGroup [TagDescription]

The tags for the group.

asgAutoScalingGroupName :: Lens' AutoScalingGroup Text

The name of the group.

asgMinSize :: Lens' AutoScalingGroup Int

The minimum size of the group.

asgMaxSize :: Lens' AutoScalingGroup Int

The maximum size of the group.

asgDesiredCapacity :: Lens' AutoScalingGroup Int

The desired size of the group.

asgDefaultCooldown :: Lens' AutoScalingGroup Int

The amount of time, in seconds, after a scaling activity completes before another scaling activity can start.

asgAvailabilityZones :: Lens' AutoScalingGroup (NonEmpty Text)

One or more Availability Zones for the group.

asgHealthCheckType :: Lens' AutoScalingGroup Text

The service to use for the health checks. The valid values are EC2 and ELB .

asgCreatedTime :: Lens' AutoScalingGroup UTCTime

The date and time the group was created.

AutoScalingInstanceDetails

data AutoScalingInstanceDetails

Describes an EC2 instance associated with an Auto Scaling group.

See: autoScalingInstanceDetails smart constructor.

autoScalingInstanceDetails

Arguments

:: Text
asidInstanceId

-> Text
asidAutoScalingGroupName

-> Text
asidAvailabilityZone

-> Text
asidLifecycleState

-> Text
asidHealthStatus

-> Bool
asidProtectedFromScaleIn

-> AutoScalingInstanceDetails

Creates a value of AutoScalingInstanceDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asidLaunchConfigurationName - The launch configuration used to launch the instance. This value is not available if you attached the instance to the Auto Scaling group.

asidInstanceId - The ID of the instance.

asidAutoScalingGroupName - The name of the Auto Scaling group associated with the instance.

asidAvailabilityZone - The Availability Zone for the instance.

asidLifecycleState - The lifecycle state for the instance. For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

asidHealthStatus - The last reported health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy and Auto Scaling should terminate and replace it.

asidProtectedFromScaleIn - Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

asidLaunchConfigurationName :: Lens' AutoScalingInstanceDetails (Maybe Text)

The launch configuration used to launch the instance. This value is not available if you attached the instance to the Auto Scaling group.

asidInstanceId :: Lens' AutoScalingInstanceDetails Text

The ID of the instance.

asidAutoScalingGroupName :: Lens' AutoScalingInstanceDetails Text

The name of the Auto Scaling group associated with the instance.

asidAvailabilityZone :: Lens' AutoScalingInstanceDetails Text

The Availability Zone for the instance.

asidLifecycleState :: Lens' AutoScalingInstanceDetails Text

The lifecycle state for the instance. For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

asidHealthStatus :: Lens' AutoScalingInstanceDetails Text

The last reported health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy and Auto Scaling should terminate and replace it.

asidProtectedFromScaleIn :: Lens' AutoScalingInstanceDetails Bool

Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

BlockDeviceMapping

data BlockDeviceMapping

Describes a block device mapping.

See: blockDeviceMapping smart constructor.

blockDeviceMapping

Arguments

:: Text
bdmDeviceName

-> BlockDeviceMapping

Creates a value of BlockDeviceMapping with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdmVirtualName - The name of the virtual device (for example, ephemeral0 ).

bdmNoDevice - Suppresses a device mapping. If this parameter is true for the root device, the instance might fail the EC2 health check. Auto Scaling launches a replacement instance if the instance fails the health check.

bdmEBS - The information about the Amazon EBS volume.

bdmDeviceName - The device name exposed to the EC2 instance (for example, devsdh or xvdh ).

bdmVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

The name of the virtual device (for example, ephemeral0 ).

bdmNoDevice :: Lens' BlockDeviceMapping (Maybe Bool)

Suppresses a device mapping. If this parameter is true for the root device, the instance might fail the EC2 health check. Auto Scaling launches a replacement instance if the instance fails the health check.

bdmEBS :: Lens' BlockDeviceMapping (Maybe EBS)

The information about the Amazon EBS volume.

bdmDeviceName :: Lens' BlockDeviceMapping Text

The device name exposed to the EC2 instance (for example, devsdh or xvdh ).

CustomizedMetricSpecification

data CustomizedMetricSpecification

Configures a customized metric for a target tracking policy.

See: customizedMetricSpecification smart constructor.

customizedMetricSpecification

Arguments

:: Text
cmsMetricName

-> Text
cmsNamespace

-> MetricStatistic
cmsStatistic

-> CustomizedMetricSpecification

Creates a value of CustomizedMetricSpecification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmsDimensions - The dimensions of the metric.

cmsUnit - The unit of the metric.

cmsMetricName - The name of the metric.

cmsNamespace - The namespace of the metric.

cmsStatistic - The statistic of the metric.

cmsDimensions :: Lens' CustomizedMetricSpecification [MetricDimension]

The dimensions of the metric.

cmsUnit :: Lens' CustomizedMetricSpecification (Maybe Text)

The unit of the metric.

cmsMetricName :: Lens' CustomizedMetricSpecification Text

The name of the metric.

cmsNamespace :: Lens' CustomizedMetricSpecification Text

The namespace of the metric.

cmsStatistic :: Lens' CustomizedMetricSpecification MetricStatistic

The statistic of the metric.

EBS

data EBS

Describes an Amazon EBS volume.

See: ebs smart constructor.

ebs :: EBS

Creates a value of EBS with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ebsDeleteOnTermination - Indicates whether the volume is deleted on instance termination. Default: true

ebsVolumeSize - The volume size, in GiB. For standard volumes, specify a value from 1 to 1,024. For io1 volumes, specify a value from 4 to 16,384. For gp2 volumes, specify a value from 1 to 16,384. If you specify a snapshot, the volume size must be equal to or larger than the snapshot size. Default: If you create a volume from a snapshot and you don't specify a volume size, the default is the snapshot size.

ebsIOPS - The number of I/O operations per second (IOPS) to provision for the volume. Constraint: Required when the volume type is io1 .

ebsEncrypted - Indicates whether the volume should be encrypted. Encrypted EBS volumes must be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or an unencrypted volume from an encrypted snapshot. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

ebsVolumeType - The volume type. For more information, see Amazon EBS Volume Types in the Amazon Elastic Compute Cloud User Guide . Valid values: standard | io1 | gp2 Default: standard

ebsSnapshotId - The ID of the snapshot.

ebsDeleteOnTermination :: Lens' EBS (Maybe Bool)

Indicates whether the volume is deleted on instance termination. Default: true

ebsVolumeSize :: Lens' EBS (Maybe Natural)

The volume size, in GiB. For standard volumes, specify a value from 1 to 1,024. For io1 volumes, specify a value from 4 to 16,384. For gp2 volumes, specify a value from 1 to 16,384. If you specify a snapshot, the volume size must be equal to or larger than the snapshot size. Default: If you create a volume from a snapshot and you don't specify a volume size, the default is the snapshot size.

ebsIOPS :: Lens' EBS (Maybe Natural)

The number of I/O operations per second (IOPS) to provision for the volume. Constraint: Required when the volume type is io1 .

ebsEncrypted :: Lens' EBS (Maybe Bool)

Indicates whether the volume should be encrypted. Encrypted EBS volumes must be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or an unencrypted volume from an encrypted snapshot. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

ebsVolumeType :: Lens' EBS (Maybe Text)

The volume type. For more information, see Amazon EBS Volume Types in the Amazon Elastic Compute Cloud User Guide . Valid values: standard | io1 | gp2 Default: standard

ebsSnapshotId :: Lens' EBS (Maybe Text)

The ID of the snapshot.

EnabledMetric

data EnabledMetric

Describes an enabled metric.

See: enabledMetric smart constructor.

enabledMetric :: EnabledMetric

Creates a value of EnabledMetric with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

emGranularity - The granularity of the metric. The only valid value is 1Minute .

emMetric - One of the following metrics: * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

emGranularity :: Lens' EnabledMetric (Maybe Text)

The granularity of the metric. The only valid value is 1Minute .

emMetric :: Lens' EnabledMetric (Maybe Text)

One of the following metrics: * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

Filter

data Filter

Describes a filter.

See: filter' smart constructor.

filter'

Arguments

:: Text
fName

-> Filter

Creates a value of Filter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fValues - The value of the filter.

fName - The name of the filter. The valid values are: "auto-scaling-group" , "key" , "value" , and "propagate-at-launch" .

fValues :: Lens' Filter [Text]

The value of the filter.

fName :: Lens' Filter Text

The name of the filter. The valid values are: "auto-scaling-group" , "key" , "value" , and "propagate-at-launch" .

Instance

data Instance

Describes an EC2 instance.

See: instance' smart constructor.

instance'

Arguments

:: Text
iInstanceId

-> Text
iAvailabilityZone

-> LifecycleState
iLifecycleState

-> Text
iHealthStatus

-> Bool
iProtectedFromScaleIn

-> Instance

Creates a value of Instance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iLaunchConfigurationName - The launch configuration associated with the instance.

iInstanceId - The ID of the instance.

iAvailabilityZone - The Availability Zone in which the instance is running.

iLifecycleState - A description of the current lifecycle state. Note that the Quarantined state is not used.

iHealthStatus - The last reported health status of the instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy and Auto Scaling should terminate and replace it.

iProtectedFromScaleIn - Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

iLaunchConfigurationName :: Lens' Instance (Maybe Text)

The launch configuration associated with the instance.

iInstanceId :: Lens' Instance Text

The ID of the instance.

iAvailabilityZone :: Lens' Instance Text

The Availability Zone in which the instance is running.

iLifecycleState :: Lens' Instance LifecycleState

A description of the current lifecycle state. Note that the Quarantined state is not used.

iHealthStatus :: Lens' Instance Text

The last reported health status of the instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy and Auto Scaling should terminate and replace it.

iProtectedFromScaleIn :: Lens' Instance Bool

Indicates whether the instance is protected from termination by Auto Scaling when scaling in.

InstanceMonitoring

data InstanceMonitoring

Describes whether instance monitoring is enabled.

See: instanceMonitoring smart constructor.

instanceMonitoring :: InstanceMonitoring

Creates a value of InstanceMonitoring with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

imEnabled - If True , instance monitoring is enabled.

imEnabled :: Lens' InstanceMonitoring (Maybe Bool)

If True , instance monitoring is enabled.

LaunchConfiguration

data LaunchConfiguration

Describes a launch configuration.

See: launchConfiguration smart constructor.

launchConfiguration

Arguments

:: Text
lcLaunchConfigurationName

-> Text
lcImageId

-> Text
lcInstanceType

-> UTCTime
lcCreatedTime

-> LaunchConfiguration

Creates a value of LaunchConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcAssociatePublicIPAddress - [EC2-VPC] Indicates whether to assign a public IP address to each instance.

lcSecurityGroups - The security groups to associate with the instances.

lcSpotPrice - The price to bid when launching Spot Instances.

lcInstanceMonitoring - Controls whether instances in this group are launched with detailed (true ) or basic (false ) monitoring.

lcKeyName - The name of the key pair.

lcClassicLinkVPCSecurityGroups - The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId . This parameter is required if you specify a ClassicLink-enabled VPC, and cannot be used otherwise. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

lcRAMDiskId - The ID of the RAM disk associated with the AMI.

lcKernelId - The ID of the kernel associated with the AMI.

lcEBSOptimized - Controls whether the instance is optimized for EBS I/O (true ) or not (false ).

lcUserData - The user data available to the instances.

lcClassicLinkVPCId - The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. This parameter can only be used if you are launching EC2-Classic instances. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

lcIAMInstanceProfile - The name or Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance.

lcLaunchConfigurationARN - The Amazon Resource Name (ARN) of the launch configuration.

lcPlacementTenancy - The tenancy of the instance, either default or dedicated . An instance with dedicated tenancy runs in an isolated, single-tenant hardware and can only be launched into a VPC.

lcBlockDeviceMappings - A block device mapping, which specifies the block devices for the instance.

lcLaunchConfigurationName - The name of the launch configuration.

lcImageId - The ID of the Amazon Machine Image (AMI).

lcInstanceType - The instance type for the instances.

lcCreatedTime - The creation date and time for the launch configuration.

lcAssociatePublicIPAddress :: Lens' LaunchConfiguration (Maybe Bool)

EC2-VPC

Indicates whether to assign a public IP address to each instance.

lcSecurityGroups :: Lens' LaunchConfiguration [Text]

The security groups to associate with the instances.

lcSpotPrice :: Lens' LaunchConfiguration (Maybe Text)

The price to bid when launching Spot Instances.

lcInstanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed (true ) or basic (false ) monitoring.

lcKeyName :: Lens' LaunchConfiguration (Maybe Text)

The name of the key pair.

lcClassicLinkVPCSecurityGroups :: Lens' LaunchConfiguration [Text]

The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId . This parameter is required if you specify a ClassicLink-enabled VPC, and cannot be used otherwise. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

lcRAMDiskId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the RAM disk associated with the AMI.

lcKernelId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the kernel associated with the AMI.

lcEBSOptimized :: Lens' LaunchConfiguration (Maybe Bool)

Controls whether the instance is optimized for EBS I/O (true ) or not (false ).

lcUserData :: Lens' LaunchConfiguration (Maybe Text)

The user data available to the instances.

lcClassicLinkVPCId :: Lens' LaunchConfiguration (Maybe Text)

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. This parameter can only be used if you are launching EC2-Classic instances. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

lcIAMInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

The name or Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance.

lcLaunchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the launch configuration.

lcPlacementTenancy :: Lens' LaunchConfiguration (Maybe Text)

The tenancy of the instance, either default or dedicated . An instance with dedicated tenancy runs in an isolated, single-tenant hardware and can only be launched into a VPC.

lcBlockDeviceMappings :: Lens' LaunchConfiguration [BlockDeviceMapping]

A block device mapping, which specifies the block devices for the instance.

lcLaunchConfigurationName :: Lens' LaunchConfiguration Text

The name of the launch configuration.

lcImageId :: Lens' LaunchConfiguration Text

The ID of the Amazon Machine Image (AMI).

lcInstanceType :: Lens' LaunchConfiguration Text

The instance type for the instances.

lcCreatedTime :: Lens' LaunchConfiguration UTCTime

The creation date and time for the launch configuration.

LifecycleHook

data LifecycleHook

Describes a lifecycle hook, which tells Auto Scaling that you want to perform an action when an instance launches or terminates. When you have a lifecycle hook in place, the Auto Scaling group will either:

Pause the instance after it launches, but before it is put into service

Pause the instance as it terminates, but before it is fully terminated

For more information, see Auto Scaling Lifecycle in the Auto Scaling User Guide .

See: lifecycleHook smart constructor.

lifecycleHook :: LifecycleHook

Creates a value of LifecycleHook with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lhDefaultResult - Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON . The default value is CONTINUE .

lhLifecycleHookName - The name of the lifecycle hook.

lhHeartbeatTimeout - The maximum time, in seconds, that can elapse before the lifecycle hook times out. The default is 3600 seconds (1 hour). When the lifecycle hook times out, Auto Scaling performs the default action. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat .

lhAutoScalingGroupName - The name of the Auto Scaling group for the lifecycle hook.

lhNotificationMetadata - Additional information that you want to include any time Auto Scaling sends a message to the notification target.

lhGlobalTimeout - The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout , whichever is smaller.

lhNotificationTargetARN - The ARN of the notification target that Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target includes the following: * Lifecycle action token * User account ID * Name of the Auto Scaling group * Lifecycle hook name * EC2 instance ID * Lifecycle transition * Notification metadata

lhLifecycleTransition - The state of the EC2 instance to which you want to attach the lifecycle hook. For a list of lifecycle hook types, see DescribeLifecycleHookTypes .

lhRoleARN - The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

lhDefaultResult :: Lens' LifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON . The default value is CONTINUE .

lhLifecycleHookName :: Lens' LifecycleHook (Maybe Text)

The name of the lifecycle hook.

lhHeartbeatTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum time, in seconds, that can elapse before the lifecycle hook times out. The default is 3600 seconds (1 hour). When the lifecycle hook times out, Auto Scaling performs the default action. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat .

lhAutoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

The name of the Auto Scaling group for the lifecycle hook.

lhNotificationMetadata :: Lens' LifecycleHook (Maybe Text)

Additional information that you want to include any time Auto Scaling sends a message to the notification target.

lhGlobalTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum time, in seconds, that an instance can remain in a Pending:Wait or Terminating:Wait state. The maximum is 172800 seconds (48 hours) or 100 times HeartbeatTimeout , whichever is smaller.

lhNotificationTargetARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the notification target that Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target includes the following: * Lifecycle action token * User account ID * Name of the Auto Scaling group * Lifecycle hook name * EC2 instance ID * Lifecycle transition * Notification metadata

lhLifecycleTransition :: Lens' LifecycleHook (Maybe Text)

The state of the EC2 instance to which you want to attach the lifecycle hook. For a list of lifecycle hook types, see DescribeLifecycleHookTypes .

lhRoleARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

LoadBalancerState

data LoadBalancerState

Describes the state of a Classic Load Balancer.

If you specify a load balancer when creating the Auto Scaling group, the state of the load balancer is InService .

If you attach a load balancer to an existing Auto Scaling group, the initial state is Adding . The state transitions to Added after all instances in the group are registered with the load balancer. If ELB health checks are enabled for the load balancer, the state transitions to InService after at least one instance in the group passes the health check. If EC2 health checks are enabled instead, the load balancer remains in the Added state.

See: loadBalancerState smart constructor.

loadBalancerState :: LoadBalancerState

Creates a value of LoadBalancerState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbsState - One of the following load balancer states: * Adding - The instances in the group are being registered with the load balancer. * Added - All instances in the group are registered with the load balancer. * InService - At least one instance in the group passed an ELB health check. * Removing - The instances in the group are being deregistered from the load balancer. If connection draining is enabled, Elastic Load Balancing waits for in-flight requests to complete before deregistering the instances. * Removed - All instances in the group are deregistered from the load balancer.

lbsLoadBalancerName - The name of the load balancer.

lbsState :: Lens' LoadBalancerState (Maybe Text)

One of the following load balancer states: * Adding - The instances in the group are being registered with the load balancer. * Added - All instances in the group are registered with the load balancer. * InService - At least one instance in the group passed an ELB health check. * Removing - The instances in the group are being deregistered from the load balancer. If connection draining is enabled, Elastic Load Balancing waits for in-flight requests to complete before deregistering the instances. * Removed - All instances in the group are deregistered from the load balancer.

lbsLoadBalancerName :: Lens' LoadBalancerState (Maybe Text)

The name of the load balancer.

LoadBalancerTargetGroupState

data LoadBalancerTargetGroupState

Describes the state of a target group.

If you attach a target group to an existing Auto Scaling group, the initial state is Adding . The state transitions to Added after all Auto Scaling instances are registered with the target group. If ELB health checks are enabled, the state transitions to InService after at least one Auto Scaling instance passes the health check. If EC2 health checks are enabled instead, the target group remains in the Added state.

See: loadBalancerTargetGroupState smart constructor.

loadBalancerTargetGroupState :: LoadBalancerTargetGroupState

Creates a value of LoadBalancerTargetGroupState with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbtgsState - The state of the target group. * Adding - The Auto Scaling instances are being registered with the target group. * Added - All Auto Scaling instances are registered with the target group. * InService - At least one Auto Scaling instance passed an ELB health check. * Removing - The Auto Scaling instances are being deregistered from the target group. If connection draining is enabled, Elastic Load Balancing waits for in-flight requests to complete before deregistering the instances. * Removed - All Auto Scaling instances are deregistered from the target group.

lbtgsLoadBalancerTargetGroupARN - The Amazon Resource Name (ARN) of the target group.

lbtgsState :: Lens' LoadBalancerTargetGroupState (Maybe Text)

The state of the target group. * Adding - The Auto Scaling instances are being registered with the target group. * Added - All Auto Scaling instances are registered with the target group. * InService - At least one Auto Scaling instance passed an ELB health check. * Removing - The Auto Scaling instances are being deregistered from the target group. If connection draining is enabled, Elastic Load Balancing waits for in-flight requests to complete before deregistering the instances. * Removed - All Auto Scaling instances are deregistered from the target group.

lbtgsLoadBalancerTargetGroupARN :: Lens' LoadBalancerTargetGroupState (Maybe Text)

The Amazon Resource Name (ARN) of the target group.

MetricCollectionType

data MetricCollectionType

Describes a metric.

See: metricCollectionType smart constructor.

metricCollectionType :: MetricCollectionType

Creates a value of MetricCollectionType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mctMetric - One of the following metrics: * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

mctMetric :: Lens' MetricCollectionType (Maybe Text)

One of the following metrics: * GroupMinSize * GroupMaxSize * GroupDesiredCapacity * GroupInServiceInstances * GroupPendingInstances * GroupStandbyInstances * GroupTerminatingInstances * GroupTotalInstances

MetricDimension

data MetricDimension

Describes the dimension of a metric.

See: metricDimension smart constructor.

metricDimension

Arguments

:: Text
mdName

-> Text
mdValue

-> MetricDimension

Creates a value of MetricDimension with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdName - The name of the dimension.

mdValue - The value of the dimension.

mdName :: Lens' MetricDimension Text

The name of the dimension.

mdValue :: Lens' MetricDimension Text

The value of the dimension.

MetricGranularityType

data MetricGranularityType

Describes a granularity of a metric.

See: metricGranularityType smart constructor.

metricGranularityType :: MetricGranularityType

Creates a value of MetricGranularityType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mgtGranularity - The granularity. The only valid value is 1Minute .

mgtGranularity :: Lens' MetricGranularityType (Maybe Text)

The granularity. The only valid value is 1Minute .

NotificationConfiguration

data NotificationConfiguration

Describes a notification.

See: notificationConfiguration smart constructor.

notificationConfiguration :: NotificationConfiguration

Creates a value of NotificationConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ncTopicARN - The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

ncAutoScalingGroupName - The name of the group.

ncNotificationType - One of the following event notification types: * autoscaling:EC2_INSTANCE_LAUNCH * autoscaling:EC2_INSTANCE_LAUNCH_ERROR * autoscaling:EC2_INSTANCE_TERMINATE * autoscaling:EC2_INSTANCE_TERMINATE_ERROR * autoscaling:TEST_NOTIFICATION

ncTopicARN :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

ncAutoScalingGroupName :: Lens' NotificationConfiguration (Maybe Text)

The name of the group.

ncNotificationType :: Lens' NotificationConfiguration (Maybe Text)

One of the following event notification types: * autoscaling:EC2_INSTANCE_LAUNCH * autoscaling:EC2_INSTANCE_LAUNCH_ERROR * autoscaling:EC2_INSTANCE_TERMINATE * autoscaling:EC2_INSTANCE_TERMINATE_ERROR * autoscaling:TEST_NOTIFICATION

PredefinedMetricSpecification

data PredefinedMetricSpecification

Configures a predefined metric for a target tracking policy. The following predefined metrics are available:

ASGAverageCPUUtilization - average CPU utilization of the Auto Scaling group

ASGAverageNetworkIn - average number of bytes received on all network interfaces by the Auto Scaling group

ASGAverageNetworkOut - average number of bytes sent out on all network interfaces by the Auto Scaling group

ALBRequestCountPerTarget - number of requests completed per target in an Application Load Balancer target group

See: predefinedMetricSpecification smart constructor.

predefinedMetricSpecification

Arguments

:: MetricType
pmsPredefinedMetricType

-> PredefinedMetricSpecification

Creates a value of PredefinedMetricSpecification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pmsResourceLabel - Identifies the resource associated with the metric type. For predefined metric types ASGAverageCPUUtilization , ASGAverageNetworkIn and ASGAverageNetworkOut , the parameter must not be specified as the resource associated with the metric type is the Auto Scaling group. For predefined metric type ALBRequestCountPerTarget , the parameter must be specified in the format app/load-balancer-name /load-balancer-id targetgrouptarget-group-name /target-group-id , where app/load-balancer-name /load-balancer-id is the final portion of the load balancer ARN, and targetgroup/target-group-name /target-group-id is the final portion of the target group ARN. The target group must be attached to the Auto Scaling group.

pmsPredefinedMetricType - The metric type.

pmsResourceLabel :: Lens' PredefinedMetricSpecification (Maybe Text)

Identifies the resource associated with the metric type. For predefined metric types ASGAverageCPUUtilization , ASGAverageNetworkIn and ASGAverageNetworkOut , the parameter must not be specified as the resource associated with the metric type is the Auto Scaling group. For predefined metric type ALBRequestCountPerTarget , the parameter must be specified in the format app/load-balancer-name /load-balancer-id targetgrouptarget-group-name /target-group-id , where app/load-balancer-name /load-balancer-id is the final portion of the load balancer ARN, and targetgroup/target-group-name /target-group-id is the final portion of the target group ARN. The target group must be attached to the Auto Scaling group.

pmsPredefinedMetricType :: Lens' PredefinedMetricSpecification MetricType

The metric type.

ProcessType

data ProcessType

Describes a process type.

For more information, see Auto Scaling Processes in the Auto Scaling User Guide .

See: processType smart constructor.

processType

Arguments

:: Text
ptProcessName

-> ProcessType

Creates a value of ProcessType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ptProcessName - One of the following processes: * Launch * Terminate * AddToLoadBalancer * AlarmNotification * AZRebalance * HealthCheck * ReplaceUnhealthy * ScheduledActions

ptProcessName :: Lens' ProcessType Text

One of the following processes: * Launch * Terminate * AddToLoadBalancer * AlarmNotification * AZRebalance * HealthCheck * ReplaceUnhealthy * ScheduledActions

ScalingPolicy

data ScalingPolicy

Describes a scaling policy.

See: scalingPolicy smart constructor.

scalingPolicy :: ScalingPolicy

Creates a value of ScalingPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sMinAdjustmentStep - Available for backward compatibility. Use MinAdjustmentMagnitude instead.

sEstimatedInstanceWarmup - The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

sPolicyName - The name of the scaling policy.

sPolicyType - The policy type. Valid values are SimpleScaling and StepScaling .

sStepAdjustments - A set of adjustments that enable you to scale based on the size of the alarm breach.

sTargetTrackingConfiguration - A target tracking policy.

sAdjustmentType - The adjustment type, which specifies how ScalingAdjustment is interpreted. Valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity .

sAutoScalingGroupName - The name of the Auto Scaling group associated with this scaling policy.

sScalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

sCooldown - The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sPolicyARN - The Amazon Resource Name (ARN) of the policy.

sAlarms - The CloudWatch alarms related to the policy.

sMetricAggregationType - The aggregation type for the CloudWatch metrics. Valid values are Minimum , Maximum , and Average .

sMinAdjustmentMagnitude - The minimum number of instances to scale. If the value of AdjustmentType is PercentChangeInCapacity , the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least this many instances. Otherwise, the error is ValidationError .

sMinAdjustmentStep :: Lens' ScalingPolicy (Maybe Int)

Available for backward compatibility. Use MinAdjustmentMagnitude instead.

sEstimatedInstanceWarmup :: Lens' ScalingPolicy (Maybe Int)

The estimated time, in seconds, until a newly launched instance can contribute to the CloudWatch metrics.

sPolicyName :: Lens' ScalingPolicy (Maybe Text)

The name of the scaling policy.

sPolicyType :: Lens' ScalingPolicy (Maybe Text)

The policy type. Valid values are SimpleScaling and StepScaling .

sStepAdjustments :: Lens' ScalingPolicy [StepAdjustment]

A set of adjustments that enable you to scale based on the size of the alarm breach.

sTargetTrackingConfiguration :: Lens' ScalingPolicy (Maybe TargetTrackingConfiguration)

A target tracking policy.

sAdjustmentType :: Lens' ScalingPolicy (Maybe Text)

The adjustment type, which specifies how ScalingAdjustment is interpreted. Valid values are ChangeInCapacity , ExactCapacity , and PercentChangeInCapacity .

sAutoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

The name of the Auto Scaling group associated with this scaling policy.

sScalingAdjustment :: Lens' ScalingPolicy (Maybe Int)

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

sCooldown :: Lens' ScalingPolicy (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sPolicyARN :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

sAlarms :: Lens' ScalingPolicy [Alarm]

The CloudWatch alarms related to the policy.

sMetricAggregationType :: Lens' ScalingPolicy (Maybe Text)

The aggregation type for the CloudWatch metrics. Valid values are Minimum , Maximum , and Average .

sMinAdjustmentMagnitude :: Lens' ScalingPolicy (Maybe Int)

The minimum number of instances to scale. If the value of AdjustmentType is PercentChangeInCapacity , the scaling policy changes the DesiredCapacity of the Auto Scaling group by at least this many instances. Otherwise, the error is ValidationError .

ScalingProcessQuery

data ScalingProcessQuery

Contains the parameters for SuspendProcesses and ResumeProcesses.

See: scalingProcessQuery smart constructor.

scalingProcessQuery

Arguments

:: Text
spqAutoScalingGroupName

-> ScalingProcessQuery

Creates a value of ScalingProcessQuery with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spqScalingProcesses - One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

spqAutoScalingGroupName - The name or Amazon Resource Name (ARN) of the Auto Scaling group.

spqScalingProcesses :: Lens' ScalingProcessQuery [Text]

One or more of the following processes. If you omit this parameter, all processes are specified. * Launch * Terminate * HealthCheck * ReplaceUnhealthy * AZRebalance * AlarmNotification * ScheduledActions * AddToLoadBalancer

spqAutoScalingGroupName :: Lens' ScalingProcessQuery Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

ScheduledUpdateGroupAction

data ScheduledUpdateGroupAction

Describes a scheduled update to an Auto Scaling group.

See: scheduledUpdateGroupAction smart constructor.

scheduledUpdateGroupAction :: ScheduledUpdateGroupAction

Creates a value of ScheduledUpdateGroupAction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sugaScheduledActionARN - The Amazon Resource Name (ARN) of the scheduled action.

sugaStartTime - The date and time that the action is scheduled to begin. This date and time can be up to one month in the future. When StartTime and EndTime are specified with Recurrence , they form the boundaries of when the recurring action will start and stop.

sugaTime - This parameter is deprecated.

sugaScheduledActionName - The name of the scheduled action.

sugaMaxSize - The maximum size of the group.

sugaRecurrence - The recurring schedule for the action.

sugaDesiredCapacity - The number of instances you prefer to maintain in the group.

sugaMinSize - The minimum size of the group.

sugaAutoScalingGroupName - The name of the group.

sugaEndTime - The date and time that the action is scheduled to end. This date and time can be up to one month in the future.

sugaScheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The Amazon Resource Name (ARN) of the scheduled action.

sugaStartTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The date and time that the action is scheduled to begin. This date and time can be up to one month in the future. When StartTime and EndTime are specified with Recurrence , they form the boundaries of when the recurring action will start and stop.

sugaTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

This parameter is deprecated.

sugaScheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the scheduled action.

sugaMaxSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The maximum size of the group.

sugaRecurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The recurring schedule for the action.

sugaDesiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The number of instances you prefer to maintain in the group.

sugaMinSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The minimum size of the group.

sugaAutoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the group.

sugaEndTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The date and time that the action is scheduled to end. This date and time can be up to one month in the future.

StepAdjustment

data StepAdjustment

Describes an adjustment based on the difference between the value of the aggregated CloudWatch metric and the breach threshold that you've defined for the alarm.

For the following examples, suppose that you have an alarm with a breach threshold of 50:

If you want the adjustment to be triggered when the metric is greater than or equal to 50 and less than 60, specify a lower bound of 0 and an upper bound of 10.

If you want the adjustment to be triggered when the metric is greater than 40 and less than or equal to 50, specify a lower bound of -10 and an upper bound of 0.

There are a few rules for the step adjustments for your step policy:

The ranges of your step adjustments can't overlap or have a gap.

At most one step adjustment can have a null lower bound. If one step adjustment has a negative lower bound, then there must be a step adjustment with a null lower bound.

At most one step adjustment can have a null upper bound. If one step adjustment has a positive upper bound, then there must be a step adjustment with a null upper bound.

The upper and lower bound can't be null in the same step adjustment.

See: stepAdjustment smart constructor.

stepAdjustment

Arguments

:: Int
saScalingAdjustment

-> StepAdjustment

Creates a value of StepAdjustment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

saMetricIntervalLowerBound - The lower bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the lower bound is inclusive (the metric must be greater than or equal to the threshold plus the lower bound). Otherwise, it is exclusive (the metric must be greater than the threshold plus the lower bound). A null value indicates negative infinity.

saMetricIntervalUpperBound - The upper bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the upper bound is exclusive (the metric must be less than the threshold plus the upper bound). Otherwise, it is inclusive (the metric must be less than or equal to the threshold plus the upper bound). A null value indicates positive infinity. The upper bound must be greater than the lower bound.

saScalingAdjustment - The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

saMetricIntervalLowerBound :: Lens' StepAdjustment (Maybe Double)

The lower bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the lower bound is inclusive (the metric must be greater than or equal to the threshold plus the lower bound). Otherwise, it is exclusive (the metric must be greater than the threshold plus the lower bound). A null value indicates negative infinity.

saMetricIntervalUpperBound :: Lens' StepAdjustment (Maybe Double)

The upper bound for the difference between the alarm threshold and the CloudWatch metric. If the metric value is above the breach threshold, the upper bound is exclusive (the metric must be less than the threshold plus the upper bound). Otherwise, it is inclusive (the metric must be less than or equal to the threshold plus the upper bound). A null value indicates positive infinity. The upper bound must be greater than the lower bound.

saScalingAdjustment :: Lens' StepAdjustment Int

The amount by which to scale, based on the specified adjustment type. A positive value adds to the current capacity while a negative number removes from the current capacity.

SuspendedProcess

data SuspendedProcess

Describes an Auto Scaling process that has been suspended. For more information, see ProcessType .

See: suspendedProcess smart constructor.

suspendedProcess :: SuspendedProcess

Creates a value of SuspendedProcess with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

spProcessName - The name of the suspended process.

spSuspensionReason - The reason that the process was suspended.

spProcessName :: Lens' SuspendedProcess (Maybe Text)

The name of the suspended process.

spSuspensionReason :: Lens' SuspendedProcess (Maybe Text)

The reason that the process was suspended.

Tag

data Tag

Describes a tag for an Auto Scaling group.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagResourceId

-> Text
tagResourceType

-> Bool
tagPropagateAtLaunch

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - The tag key.

tagResourceId - The name of the group.

tagResourceType - The type of resource. The only supported value is auto-scaling-group .

tagPropagateAtLaunch - Determines whether the tag is added to new instances as they are launched in the group.

tagValue - The tag value.

tagKey :: Lens' Tag Text

The tag key.

tagResourceId :: Lens' Tag Text

The name of the group.

tagResourceType :: Lens' Tag Text

The type of resource. The only supported value is auto-scaling-group .

tagPropagateAtLaunch :: Lens' Tag Bool

Determines whether the tag is added to new instances as they are launched in the group.

tagValue :: Lens' Tag Text

The tag value.

TagDescription

data TagDescription

Describes a tag for an Auto Scaling group.

See: tagDescription smart constructor.

tagDescription

Arguments

:: Text
tdResourceId

-> Text
tdResourceType

-> Text
tdKey

-> Bool
tdPropagateAtLaunch

-> Text
tdValue

-> TagDescription

Creates a value of TagDescription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tdResourceId - The name of the group.

tdResourceType - The type of resource. The only supported value is auto-scaling-group .

tdKey - The tag key.

tdPropagateAtLaunch - Determines whether the tag is added to new instances as they are launched in the group.

tdValue - The tag value.

tdResourceId :: Lens' TagDescription Text

The name of the group.

tdResourceType :: Lens' TagDescription Text

The type of resource. The only supported value is auto-scaling-group .

tdKey :: Lens' TagDescription Text

The tag key.

tdPropagateAtLaunch :: Lens' TagDescription Bool

Determines whether the tag is added to new instances as they are launched in the group.

tdValue :: Lens' TagDescription Text

The tag value.

TargetTrackingConfiguration

data TargetTrackingConfiguration

Represents a target tracking policy configuration.

See: targetTrackingConfiguration smart constructor.

targetTrackingConfiguration

Arguments

:: Double
ttcTargetValue

-> TargetTrackingConfiguration

Creates a value of TargetTrackingConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttcPredefinedMetricSpecification - A predefined metric. You can specify either a predefined metric or a customized metric.

ttcCustomizedMetricSpecification - A customized metric.

ttcDisableScaleIn - If the parameter is true, then scale-in will be disabled for the target tracking policy, i.e. the target tracking policy will not scale in the Auto Scaling group. The default value is false.

ttcTargetValue - The target value for the metric.

ttcPredefinedMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe PredefinedMetricSpecification)

A predefined metric. You can specify either a predefined metric or a customized metric.

ttcCustomizedMetricSpecification :: Lens' TargetTrackingConfiguration (Maybe CustomizedMetricSpecification)

A customized metric.

ttcDisableScaleIn :: Lens' TargetTrackingConfiguration (Maybe Bool)

If the parameter is true, then scale-in will be disabled for the target tracking policy, i.e. the target tracking policy will not scale in the Auto Scaling group. The default value is false.

ttcTargetValue :: Lens' TargetTrackingConfiguration Double

The target value for the metric.