Network.AWS.AutoScaling.Types

Service

data AutoScaling

Version 2011-01-01 of the Amazon Auto Scaling service.

Error

data RESTError :: *

XML

ns :: Text

TagDescription

data TagDescription

tagDescription

Arguments

:: Text
tdResourceId

-> Text
tdResourceType

-> Text
tdKey

-> Text
tdValue

-> Bool
tdPropagateAtLaunch

-> TagDescription

TagDescription constructor.

The fields accessible through corresponding lenses are:

tdKey :: Text

tdPropagateAtLaunch :: Bool

tdResourceId :: Text

tdResourceType :: Text

tdValue :: Text

tdKey :: Lens' TagDescription Text

The tag key.

tdPropagateAtLaunch :: Lens' TagDescription Bool

Specifies whether the tag is applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, it is applied to all instances launched after you made the change.

tdResourceId :: Lens' TagDescription Text

The name of the group.

tdResourceType :: Lens' TagDescription Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the 'auto-scaling-group' resource type.

tdValue :: Lens' TagDescription Text

The tag value.

Tag

data Tag

tag

Arguments

:: Text
tagResourceId

-> Text
tagResourceType

-> Text
tagKey

-> Text
tagValue

-> Bool
tagPropagateAtLaunch

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagPropagateAtLaunch :: Bool

tagResourceId :: Text

tagResourceType :: Text

tagValue :: Text

tagKey :: Lens' Tag Text

The tag key.

tagPropagateAtLaunch :: Lens' Tag Bool

Specifies whether the tag is applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, it is applied to all instances launched after you made the change.

tagResourceId :: Lens' Tag Text

The name of the group.

tagResourceType :: Lens' Tag Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the 'auto-scaling-group' resource type.

tagValue :: Lens' Tag Text

The tag value.

NotificationConfiguration

data NotificationConfiguration

notificationConfiguration :: NotificationConfiguration

NotificationConfiguration constructor.

The fields accessible through corresponding lenses are:

ncAutoScalingGroupName :: Maybe Text

ncNotificationType :: Maybe Text

ncTopicARN :: Maybe Text

ncAutoScalingGroupName :: Lens' NotificationConfiguration (Maybe Text)

The name of the group.

ncNotificationType :: Lens' NotificationConfiguration (Maybe Text)

The types of events for an action to start.

ncTopicARN :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

BlockDeviceMapping

data BlockDeviceMapping

blockDeviceMapping

Arguments

:: Text
bdmDeviceName

-> BlockDeviceMapping

BlockDeviceMapping constructor.

The fields accessible through corresponding lenses are:

bdmDeviceName :: Text

bdmEbs :: Maybe Ebs

bdmNoDevice :: Maybe Bool

bdmVirtualName :: Maybe Text

bdmDeviceName :: Lens' BlockDeviceMapping Text

The device name exposed to the EC2 instance (for example, '/dev/sdh' or xvdh).

bdmEbs :: Lens' BlockDeviceMapping (Maybe Ebs)

The information about the Amazon EBS volume.

bdmNoDevice :: Lens' BlockDeviceMapping (Maybe Bool)

Suppresses a device mapping.

If NoDevice is set to true for the root device, the instance might fail the EC2 health check. Auto Scaling launches a replacement instance if the instance fails the health check.

bdmVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

The name of the virtual device, ephemeral0 to ephemeral3.

LaunchConfiguration

data LaunchConfiguration

launchConfiguration

Arguments

:: Text
lcLaunchConfigurationName

-> Text
lcImageId

-> Text
lcInstanceType

-> UTCTime
lcCreatedTime

-> LaunchConfiguration

LaunchConfiguration constructor.

The fields accessible through corresponding lenses are:

lcAssociatePublicIpAddress :: Maybe Bool

lcBlockDeviceMappings :: [BlockDeviceMapping]

lcClassicLinkVPCId :: Maybe Text

lcClassicLinkVPCSecurityGroups :: [Text]

lcCreatedTime :: UTCTime

lcEbsOptimized :: Maybe Bool

lcIamInstanceProfile :: Maybe Text

lcImageId :: Text

lcInstanceMonitoring :: Maybe InstanceMonitoring

lcInstanceType :: Text

lcKernelId :: Maybe Text

lcKeyName :: Maybe Text

lcLaunchConfigurationARN :: Maybe Text

lcLaunchConfigurationName :: Text

lcPlacementTenancy :: Maybe Text

lcRamdiskId :: Maybe Text

lcSecurityGroups :: [Text]

lcSpotPrice :: Maybe Text

lcUserData :: Maybe Text

lcAssociatePublicIpAddress :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the EC2 instances are associated with a public IP address (true) or not (false).

lcBlockDeviceMappings :: Lens' LaunchConfiguration [BlockDeviceMapping]

A block device mapping that specifies how block devices are exposed to the instance. Each mapping is made up of a virtualName and a deviceName.

lcClassicLinkVPCId :: Lens' LaunchConfiguration (Maybe Text)

The ID of a ClassicLink-enabled VPC to link your EC2-Classic instances to. This parameter can only be used if you are launching EC2-Classic instances. For more information, see ClassicLink in the Amazon Elastic Compute CloudUser Guide.

lcClassicLinkVPCSecurityGroups :: Lens' LaunchConfiguration [Text]

The IDs of one or more security groups for the VPC specified in ClassicLinkVPCId. This parameter is required if ClassicLinkVPCId is specified, and cannot be used otherwise. For more information, see ClassicLink in the Amazon ElasticCompute Cloud User Guide.

lcCreatedTime :: Lens' LaunchConfiguration UTCTime

The creation date and time for the launch configuration.

lcEbsOptimized :: Lens' LaunchConfiguration (Maybe Bool)

Controls whether the instance is optimized for EBS I/O (true) or not (false).

lcIamInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

The name or Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance.

lcImageId :: Lens' LaunchConfiguration Text

The ID of the Amazon Machine Image (AMI).

lcInstanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed monitoring.

lcInstanceType :: Lens' LaunchConfiguration Text

The instance type for the EC2 instances.

lcKernelId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the kernel associated with the AMI.

lcKeyName :: Lens' LaunchConfiguration (Maybe Text)

The name of the key pair.

lcLaunchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the launch configuration.

lcLaunchConfigurationName :: Lens' LaunchConfiguration Text

The name of the launch configuration.

lcPlacementTenancy :: Lens' LaunchConfiguration (Maybe Text)

The tenancy of the instance, either 'default' or dedicated. An instance with dedicated tenancy runs in an isolated, single-tenant hardware and can only be launched in a VPC.

lcRamdiskId :: Lens' LaunchConfiguration (Maybe Text)

The ID of the RAM disk associated with the AMI.

lcSecurityGroups :: Lens' LaunchConfiguration [Text]

The security groups to associate with the EC2 instances.

lcSpotPrice :: Lens' LaunchConfiguration (Maybe Text)

The price to bid when launching Spot Instances.

lcUserData :: Lens' LaunchConfiguration (Maybe Text)

The user data available to the EC2 instances.

AutoScalingGroup

data AutoScalingGroup

autoScalingGroup

Arguments

:: Text
asgAutoScalingGroupName

-> Text
asgLaunchConfigurationName

-> Int
asgMinSize

-> Int
asgMaxSize

-> Int
asgDesiredCapacity

-> Int
asgDefaultCooldown

-> NonEmpty Text
asgAvailabilityZones

-> Text
asgHealthCheckType

-> UTCTime
asgCreatedTime

-> AutoScalingGroup

AutoScalingGroup constructor.

The fields accessible through corresponding lenses are:

asgAutoScalingGroupARN :: Maybe Text

asgAutoScalingGroupName :: Text

asgAvailabilityZones :: NonEmpty Text

asgCreatedTime :: UTCTime

asgDefaultCooldown :: Int

asgDesiredCapacity :: Int

asgEnabledMetrics :: [EnabledMetric]

asgHealthCheckGracePeriod :: Maybe Int

asgHealthCheckType :: Text

asgInstances :: [Instance]

asgLaunchConfigurationName :: Text

asgLoadBalancerNames :: [Text]

asgMaxSize :: Int

asgMinSize :: Int

asgPlacementGroup :: Maybe Text

asgStatus :: Maybe Text

asgSuspendedProcesses :: [SuspendedProcess]

asgTags :: [TagDescription]

asgTerminationPolicies :: [Text]

asgVPCZoneIdentifier :: Maybe Text

asgAutoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the group.

asgAutoScalingGroupName :: Lens' AutoScalingGroup Text

The name of the group.

asgAvailabilityZones :: Lens' AutoScalingGroup (NonEmpty Text)

One or more Availability Zones for the group.

asgCreatedTime :: Lens' AutoScalingGroup UTCTime

The date and time the group was created.

asgDefaultCooldown :: Lens' AutoScalingGroup Int

The number of seconds after a scaling activity completes before any further scaling activities can start.

asgDesiredCapacity :: Lens' AutoScalingGroup Int

The size of the group.

asgEnabledMetrics :: Lens' AutoScalingGroup [EnabledMetric]

The metrics enabled for this Auto Scaling group.

asgHealthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Int)

The amount of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

asgHealthCheckType :: Lens' AutoScalingGroup Text

The service of interest for the health status check, which can be either EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

asgInstances :: Lens' AutoScalingGroup [Instance]

The EC2 instances associated with the group.

asgLaunchConfigurationName :: Lens' AutoScalingGroup Text

The name of the associated launch configuration.

asgLoadBalancerNames :: Lens' AutoScalingGroup [Text]

One or more load balancers associated with the group.

asgMaxSize :: Lens' AutoScalingGroup Int

The maximum size of the group.

asgMinSize :: Lens' AutoScalingGroup Int

The minimum size of the group.

asgPlacementGroup :: Lens' AutoScalingGroup (Maybe Text)

The name of the placement group into which you'll launch your instances, if any. For more information, see Placement Groups.

asgStatus :: Lens' AutoScalingGroup (Maybe Text)

The current state of the Auto Scaling group when a DeleteAutoScalingGroup action is in progress.

asgSuspendedProcesses :: Lens' AutoScalingGroup [SuspendedProcess]

The suspended processes associated with the group.

asgTags :: Lens' AutoScalingGroup [TagDescription]

The tags for the Auto Scaling group.

asgTerminationPolicies :: Lens' AutoScalingGroup [Text]

The termination policies for this Auto Scaling group.

asgVPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

One or more subnet IDs, if applicable, separated by commas.

If you specify VPCZoneIdentifier and AvailabilityZones, ensure that the Availability Zones of the subnets match the values for AvailabilityZones.

ScalingPolicy

data ScalingPolicy

scalingPolicy :: ScalingPolicy

ScalingPolicy constructor.

The fields accessible through corresponding lenses are:

sp1AdjustmentType :: Maybe Text

sp1Alarms :: [Alarm]

sp1AutoScalingGroupName :: Maybe Text

sp1Cooldown :: Maybe Int

sp1MinAdjustmentStep :: Maybe Int

sp1PolicyARN :: Maybe Text

sp1PolicyName :: Maybe Text

sp1ScalingAdjustment :: Maybe Int

sp1AdjustmentType :: Lens' ScalingPolicy (Maybe Text)

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

sp1Alarms :: Lens' ScalingPolicy [Alarm]

The CloudWatch Alarms related to the policy.

sp1AutoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

The name of the Auto Scaling group associated with this scaling policy.

sp1Cooldown :: Lens' ScalingPolicy (Maybe Int)

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sp1MinAdjustmentStep :: Lens' ScalingPolicy (Maybe Int)

Changes the DesiredCapacity of the Auto Scaling group by at least the specified number of instances.

sp1PolicyARN :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

sp1PolicyName :: Lens' ScalingPolicy (Maybe Text)

The name of the scaling policy.

sp1ScalingAdjustment :: Lens' ScalingPolicy (Maybe Int)

The number associated with the specified adjustment type. A positive value adds to the current capacity and a negative value removes from the current capacity.

InstanceMonitoring

data InstanceMonitoring

instanceMonitoring :: InstanceMonitoring

InstanceMonitoring constructor.

The fields accessible through corresponding lenses are:

imEnabled :: Maybe Bool

imEnabled :: Lens' InstanceMonitoring (Maybe Bool)

If True, instance monitoring is enabled.

ScheduledUpdateGroupAction

data ScheduledUpdateGroupAction

scheduledUpdateGroupAction :: ScheduledUpdateGroupAction

ScheduledUpdateGroupAction constructor.

The fields accessible through corresponding lenses are:

sugaAutoScalingGroupName :: Maybe Text

sugaDesiredCapacity :: Maybe Int

sugaEndTime :: Maybe UTCTime

sugaMaxSize :: Maybe Int

sugaMinSize :: Maybe Int

sugaRecurrence :: Maybe Text

sugaScheduledActionARN :: Maybe Text

sugaScheduledActionName :: Maybe Text

sugaStartTime :: Maybe UTCTime

sugaTime :: Maybe UTCTime

sugaAutoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the group.

sugaDesiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The number of instances you prefer to maintain in the group.

sugaEndTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The time that the action is scheduled to end. This value can be up to one month in the future.

sugaMaxSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The maximum size of the group.

sugaMinSize :: Lens' ScheduledUpdateGroupAction (Maybe Int)

The minimum size of the group.

sugaRecurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The regular schedule that an action occurs.

sugaScheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The Amazon Resource Name (ARN) of the scheduled action.

sugaScheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the scheduled action.

sugaStartTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

The time that the action is scheduled to begin. This value can be up to one month in the future.

When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

sugaTime :: Lens' ScheduledUpdateGroupAction (Maybe UTCTime)

Time is deprecated.

The time that the action is scheduled to begin. Time is an alias for StartTime.

ScalingProcessQuery

data ScalingProcessQuery

scalingProcessQuery

Arguments

:: Text
spqAutoScalingGroupName

-> ScalingProcessQuery

ScalingProcessQuery constructor.

The fields accessible through corresponding lenses are:

spqAutoScalingGroupName :: Text

spqScalingProcesses :: [Text]

spqAutoScalingGroupName :: Lens' ScalingProcessQuery Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

spqScalingProcesses :: Lens' ScalingProcessQuery [Text]

One or more of the following processes:

Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer

Ebs

data Ebs

ebs :: Ebs

Ebs constructor.

The fields accessible through corresponding lenses are:

ebsDeleteOnTermination :: Maybe Bool

ebsIops :: Maybe Natural

ebsSnapshotId :: Maybe Text

ebsVolumeSize :: Maybe Natural

ebsVolumeType :: Maybe Text

ebsDeleteOnTermination :: Lens' Ebs (Maybe Bool)

Indicates whether to delete the volume on instance termination.

Default: true

ebsIops :: Lens' Ebs (Maybe Natural)

For Provisioned IOPS (SSD) volumes only. The number of I/O operations per second (IOPS) to provision for the volume.

Valid values: Range is 100 to 4000.

Default: None

ebsSnapshotId :: Lens' Ebs (Maybe Text)

The ID of the snapshot.

ebsVolumeSize :: Lens' Ebs (Maybe Natural)

The volume size, in gigabytes.

Valid values: If the volume type is io1, the minimum size of the volume is 10 GiB. If you specify SnapshotId and VolumeSize, VolumeSize must be equal to or larger than the size of the snapshot.

Default: If you create a volume from a snapshot and you don't specify a volume size, the default is the size of the snapshot.

Required: Required when the volume type is io1.

ebsVolumeType :: Lens' Ebs (Maybe Text)

The volume type.

Valid values: 'standard | io1 | gp2'

Default: standard

AdjustmentType

data AdjustmentType

adjustmentType :: AdjustmentType

AdjustmentType constructor.

The fields accessible through corresponding lenses are:

atAdjustmentType :: Maybe Text

atAdjustmentType :: Lens' AdjustmentType (Maybe Text)

The policy adjustment type. The valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

For more information, see Dynamic Scaling in the Auto Scaling Developer Guide .

MetricCollectionType

data MetricCollectionType

metricCollectionType :: MetricCollectionType

MetricCollectionType constructor.

The fields accessible through corresponding lenses are:

mctMetric :: Maybe Text

mctMetric :: Lens' MetricCollectionType (Maybe Text)

The metric.

LifecycleHook

data LifecycleHook

lifecycleHook :: LifecycleHook

LifecycleHook constructor.

The fields accessible through corresponding lenses are:

lhAutoScalingGroupName :: Maybe Text

lhDefaultResult :: Maybe Text

lhGlobalTimeout :: Maybe Int

lhHeartbeatTimeout :: Maybe Int

lhLifecycleHookName :: Maybe Text

lhLifecycleTransition :: Maybe Text

lhNotificationMetadata :: Maybe Text

lhNotificationTargetARN :: Maybe Text

lhRoleARN :: Maybe Text

lhAutoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

The name of the Auto Scaling group for the lifecycle hook.

lhDefaultResult :: Lens' LifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The valid values are CONTINUE and ABANDON. The default value is CONTINUE.

lhGlobalTimeout :: Lens' LifecycleHook (Maybe Int)

The maximum length of time an instance can remain in a 'Pending:Wait' or 'Terminating:Wait' state. Currently, this value is set at 48 hours.

lhHeartbeatTimeout :: Lens' LifecycleHook (Maybe Int)

The amount of time that can elapse before the lifecycle hook times out. When the lifecycle hook times out, Auto Scaling performs the action defined in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

lhLifecycleHookName :: Lens' LifecycleHook (Maybe Text)

The name of the lifecycle hook.

lhLifecycleTransition :: Lens' LifecycleHook (Maybe Text)

The state of the EC2 instance to which you want to attach the lifecycle hook. For a list of lifecycle hook types, see DescribeLifecycleHooks.

lhNotificationMetadata :: Lens' LifecycleHook (Maybe Text)

Additional information that you want to include any time Auto Scaling sends a message to the notification target.

lhNotificationTargetARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the notification target that Auto Scaling uses to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target includes the following:

Lifecycle action token User account ID Name of the Auto Scaling group Lifecycle hook name EC2 instance ID Lifecycle transition Notification metadata

lhRoleARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the IAM role that allows the Auto Scaling group to publish to the specified notification target.

Activity

data Activity

activity

Arguments

:: Text
aActivityId

-> Text
aAutoScalingGroupName

-> Text
aCause

-> UTCTime
aStartTime

-> ScalingActivityStatusCode
aStatusCode

-> Activity

Activity constructor.

The fields accessible through corresponding lenses are:

aActivityId :: Text

aAutoScalingGroupName :: Text

aCause :: Text

aDescription :: Maybe Text

aDetails :: Maybe Text

aEndTime :: Maybe UTCTime

aProgress :: Maybe Int

aStartTime :: UTCTime

aStatusCode :: ScalingActivityStatusCode

aStatusMessage :: Maybe Text

aActivityId :: Lens' Activity Text

The ID of the activity.

aAutoScalingGroupName :: Lens' Activity Text

The name of the Auto Scaling group.

aCause :: Lens' Activity Text

The reason the activity was begun.

aDescription :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the scaling activity.

aDetails :: Lens' Activity (Maybe Text)

The details about the scaling activity.

aEndTime :: Lens' Activity (Maybe UTCTime)

The end time of this activity.

aProgress :: Lens' Activity (Maybe Int)

A value between 0 and 100 that indicates the progress of the activity.

aStartTime :: Lens' Activity UTCTime

The start time of this activity.

aStatusCode :: Lens' Activity ScalingActivityStatusCode

The current status of the activity.

aStatusMessage :: Lens' Activity (Maybe Text)

A friendly, more verbose description of the activity status.

SuspendedProcess

data SuspendedProcess

suspendedProcess :: SuspendedProcess

SuspendedProcess constructor.

The fields accessible through corresponding lenses are:

spProcessName :: Maybe Text

spSuspensionReason :: Maybe Text

spProcessName :: Lens' SuspendedProcess (Maybe Text)

The name of the suspended process.

spSuspensionReason :: Lens' SuspendedProcess (Maybe Text)

The reason that the process was suspended.

MetricGranularityType

data MetricGranularityType

metricGranularityType :: MetricGranularityType

MetricGranularityType constructor.

The fields accessible through corresponding lenses are:

mgtGranularity :: Maybe Text

mgtGranularity :: Lens' MetricGranularityType (Maybe Text)

The granularity.

Filter

data Filter

filter'

Arguments

:: Text
fName

-> Filter

Filter constructor.

The fields accessible through corresponding lenses are:

fName :: Text

fValues :: [Text]

fName :: Lens' Filter Text

The name of the filter. The valid values are: '"auto-scaling-group"', '"key"', '"value"', and '"propagate-at-launch"'.

fValues :: Lens' Filter [Text]

The value of the filter.

ProcessType

data ProcessType

processType

Arguments

:: Text
ptProcessName

-> ProcessType

ProcessType constructor.

The fields accessible through corresponding lenses are:

ptProcessName :: Text

ptProcessName :: Lens' ProcessType Text

The name of the process.

Alarm

data Alarm

alarm :: Alarm

Alarm constructor.

The fields accessible through corresponding lenses are:

aAlarmARN :: Maybe Text

aAlarmName :: Maybe Text

aAlarmARN :: Lens' Alarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

aAlarmName :: Lens' Alarm (Maybe Text)

The name of the alarm.

EnabledMetric

data EnabledMetric

enabledMetric :: EnabledMetric

EnabledMetric constructor.

The fields accessible through corresponding lenses are:

emGranularity :: Maybe Text

emMetric :: Maybe Text

emGranularity :: Lens' EnabledMetric (Maybe Text)

The granularity of the metric.

emMetric :: Lens' EnabledMetric (Maybe Text)

The name of the metric.

Instance

data Instance

instance'

Arguments

:: Text
iInstanceId

-> Text
iAvailabilityZone

-> LifecycleState
iLifecycleState

-> Text
iHealthStatus

-> Text
iLaunchConfigurationName

-> Instance

Instance constructor.

The fields accessible through corresponding lenses are:

iAvailabilityZone :: Text

iHealthStatus :: Text

iInstanceId :: Text

iLaunchConfigurationName :: Text

iLifecycleState :: LifecycleState

iAvailabilityZone :: Lens' Instance Text

The Availability Zone associated with this instance.

iHealthStatus :: Lens' Instance Text

The health status of the instance.

iInstanceId :: Lens' Instance Text

The ID of the instance.

iLaunchConfigurationName :: Lens' Instance Text

The launch configuration associated with the instance.

iLifecycleState :: Lens' Instance LifecycleState

A description of the current lifecycle state.

The Quarantined lifecycle state is not used.

LifecycleState

data LifecycleState

Constructors

Detached
Detached

Detaching
Detaching

EnteringStandby
EnteringStandby

InService
InService

Pending
Pending

PendingProceed
Pending:Proceed

PendingWait
Pending:Wait

Quarantined
Quarantined

Standby
Standby

Terminated
Terminated

Terminating
Terminating

TerminatingProceed
Terminating:Proceed

TerminatingWait
Terminating:Wait

AutoScalingInstanceDetails

data AutoScalingInstanceDetails

autoScalingInstanceDetails

Arguments

:: Text
asidInstanceId

-> Text
asidAutoScalingGroupName

-> Text
asidAvailabilityZone

-> Text
asidLifecycleState

-> Text
asidHealthStatus

-> Text
asidLaunchConfigurationName

-> AutoScalingInstanceDetails

AutoScalingInstanceDetails constructor.

The fields accessible through corresponding lenses are:

asidAutoScalingGroupName :: Text

asidAvailabilityZone :: Text

asidHealthStatus :: Text

asidInstanceId :: Text

asidLaunchConfigurationName :: Text

asidLifecycleState :: Text

asidAutoScalingGroupName :: Lens' AutoScalingInstanceDetails Text

The name of the Auto Scaling group associated with the instance.

asidAvailabilityZone :: Lens' AutoScalingInstanceDetails Text

The Availability Zone for the instance.

asidHealthStatus :: Lens' AutoScalingInstanceDetails Text

The health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy and Auto Scaling should terminate and replace it.

asidInstanceId :: Lens' AutoScalingInstanceDetails Text

The ID of the instance.

asidLaunchConfigurationName :: Lens' AutoScalingInstanceDetails Text

The launch configuration associated with the instance.

asidLifecycleState :: Lens' AutoScalingInstanceDetails Text

The lifecycle state for the instance. For more information, see Auto ScalingInstance States in the Auto Scaling Developer Guide.

ScalingActivityStatusCode

data ScalingActivityStatusCode

Constructors

Cancelled
Cancelled

Failed
Failed

InProgress
InProgress

MidLifecycleAction
MidLifecycleAction

PreInService
PreInService

Successful
Successful

WaitingForELBConnectionDraining
WaitingForELBConnectionDraining

WaitingForInstanceId
WaitingForInstanceId

WaitingForSpotInstanceId
WaitingForSpotInstanceId

WaitingForSpotInstanceRequestId
WaitingForSpotInstanceRequestId