Network.AWS.AutoScaling.V2011_01_01.DeleteTags

Description

Removes the specified tags or a set of tags from a set of resources.

Documentation

data DeleteTags

Constructors

DeleteTags

Fields

_dtuTags :: [Tag]

Each tag should be defined by its resource type, resource ID, key, value, and a propagate flag. Valid values are: Resource type = auto-scaling-group, Resource ID = AutoScalingGroupName, key=value, value=value, propagate=true or false.

dtuTags :: Iso' DeleteTags [Tag]

data DeleteTagsResponse

Constructors

DeleteTagsResponse