Network.AWS.AutoScaling.V2011_01_01.DetachInstances

Description

Using DetachInstances, you can remove an instance from an Auto Scaling group. After the instances are detached, you can manage them independently from the rest of the Auto Scaling group. To learn more about detaching instances, see Detach Amazon EC2 Instances From Your Auto Scaling Group. https:/autoscaling.amazonaws.com?AutoScalingGroupName=my-asg&ShouldDecrementDesiredCapacity=true&InstanceIds.member.1=i-5f2e8a0d&Version=2011-01-01

&Action=DetachInstances&SignatureVersion=2&SignatureMethod=HmacSHA256&Timestamp=2014-06-14T00%3A07%3A29.962Z&AUTHPARAMS e54ff599-bf05-4076-8b95-a0f090ed90bb 50 InProgress 2014-06-14T00:07:30.280Z At 2014-06-14T00:07:30Z instance i-5f2e8a0d was detached in response to a user request, shrinking the capacity from 4 to 3. my-asg {"Availability Zone":"us-east-1a"} Detaching EC2 instance: i-5f2e8a0d e04f3b11-f357-11e3-a434-7f10009d5849.

Request

data DetachInstances

Request constructor

mkDetachInstances

Arguments

:: Text
diAutoScalingGroupName

-> Bool
diShouldDecrementDesiredCapacity

-> DetachInstances

Smart constructor for the minimum required parameters to construct a valid DetachInstances request.

The fields accessible through corresponding lenses are:

InstanceIds :: [Text]

AutoScalingGroupName :: Text

ShouldDecrementDesiredCapacity :: Bool

Request lenses

diInstanceIds :: Lens' DetachInstances [Text]

A list of instances to detach from the Auto Scaling group. You must specify at least one instance ID.

diAutoScalingGroupName :: Lens' DetachInstances Text

The name of the Auto Scaling group from which to detach instances.

diShouldDecrementDesiredCapacity :: Lens' DetachInstances Bool

Specifies if the detached instance should decrement the desired capacity value for the Auto Scaling group. If set to True, the Auto Scaling group decrements the desired capacity value by the number of instances detached.

Response

data DetachInstancesResponse

The output of the DetachInstances action.

Response constructor

mkDetachInstancesResponse :: DetachInstancesResponse

Smart constructor for the minimum required parameters to construct a valid DetachInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Activities :: [Activity]

Response lenses

dirActivities :: Lens' DetachInstancesResponse [Activity]

A list describing the activities related to detaching the instances from the Auto Scaling group.