Network.AWS.AutoScaling.V2011_01_01.SuspendProcesses

Description

Suspends Auto Scaling processes for an Auto Scaling group. To suspend specific process types, specify them by name with the ScalingProcesses.member.N parameter. To suspend all process types, omit the ScalingProcesses.member.N parameter. Suspending either of the two primary process types, Launch or Terminate, can prevent other process types from functioning properly. To resume processes that have been suspended, use ResumeProcesses For more information on suspending and resuming Auto Scaling process, see Suspend and Resume Auto Scaling Process.

Documentation

suspendProcesses

Arguments

:: Text
_spqAutoScalingGroupName

-> SuspendProcesses

Minimum specification for a SuspendProcesses request.

data SuspendProcesses

Constructors

SuspendProcesses

Fields

_spqAutoScalingGroupName :: Text

The name or Amazon Resource Name (ARN) of the Auto Scaling group.

_spqScalingProcesses :: [Text]

The processes that you want to suspend or resume, which can include one or more of the following: Launch Terminate HealthCheck ReplaceUnhealthy AZRebalance AlarmNotification ScheduledActions AddToLoadBalancer To suspend all process types, omit this parameter.

spqScalingProcesses :: Lens' SuspendProcesses [Text]

spqAutoScalingGroupName :: Lens' SuspendProcesses Text

data SuspendProcessesResponse

Constructors

SuspendProcessesResponse