Network.AWS.AutoScaling.V2011_01_01.Types

Description

Auto Scaling allows you to scale your Amazon EC2 capacity up or down automatically according to conditions you define. With Auto Scaling, you can ensure that the number of Amazon EC2 instances you’re using increases seamlessly during demand spikes to maintain performance, and decreases automatically during demand lulls to minimize costs. Auto Scaling is particularly well suited for applications that experience hourly, daily, or weekly variability in usage. Auto Scaling is enabled by Amazon CloudWatch and available at no additional charge beyond Amazon CloudWatch fees.

Service

data AutoScaling

Supported version (2011-01-01) of the Auto Scaling service.

XML

xmlOptions :: Tagged a XMLOptions

LifecycleState

data LifecycleState

Constructors

LifecycleStateDetached
Detached

LifecycleStateDetaching
Detaching

LifecycleStateEnteringStandby
EnteringStandby

LifecycleStateInService
InService

LifecycleStatePending
Pending

LifecycleStatePendingProceed
Pending:Proceed

LifecycleStatePendingWait
Pending:Wait

LifecycleStateQuarantined
Quarantined

LifecycleStateStandby
Standby

LifecycleStateTerminated
Terminated

LifecycleStateTerminating
Terminating

LifecycleStateTerminatingProceed
Terminating:Proceed

LifecycleStateTerminatingWait
Terminating:Wait

ScalingActivityStatusCode

data ScalingActivityStatusCode

Constructors

ScalingActivityStatusCodeCancelled
Cancelled

ScalingActivityStatusCodeFailed
Failed

ScalingActivityStatusCodeInProgress
InProgress

ScalingActivityStatusCodeMidLifecycleAction
MidLifecycleAction

ScalingActivityStatusCodePreInService
PreInService

ScalingActivityStatusCodeSuccessful
Successful

ScalingActivityStatusCodeWaitingForELBConnectionDraining
WaitingForELBConnectionDraining

ScalingActivityStatusCodeWaitingForInstanceId
WaitingForInstanceId

ScalingActivityStatusCodeWaitingForSpotInstanceId
WaitingForSpotInstanceId

ScalingActivityStatusCodeWaitingForSpotInstanceRequestId
WaitingForSpotInstanceRequestId

AdjustmentType

data AdjustmentType

Specifies whether the PutScalingPolicy ScalingAdjustment parameter is an absolute number or a percentage of the current capacity.

mkAdjustmentType :: AdjustmentType

Smart constructor for the minimum required fields to construct a valid AdjustmentType data type to populate a request.

atAdjustmentType :: Lens' AdjustmentType (Maybe Text)

A policy adjustment type. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

InstanceMonitoring

data InstanceMonitoring

Enables detailed monitoring if it is disabled. Detailed monitoring is enabled by default. When detailed monitoring is enabled, Amazon Cloudwatch will generate metrics every minute and your account will be charged a fee. When you disable detailed monitoring, by specifying False, Cloudwatch will generate metrics every 5 minutes. For more information, see Monitor Your Auto Scaling Instances. For information about Amazon CloudWatch, see the Amazon CloudWatch Developer Guide.

mkInstanceMonitoring :: InstanceMonitoring

Smart constructor for the minimum required fields to construct a valid InstanceMonitoring data type to populate a request.

imEnabled :: Lens' InstanceMonitoring (Maybe Bool)

If True, instance monitoring is enabled.

MetricCollectionType

data MetricCollectionType

The MetricCollectionType data type.

mkMetricCollectionType :: MetricCollectionType

Smart constructor for the minimum required fields to construct a valid MetricCollectionType data type to populate a request.

mctMetric :: Lens' MetricCollectionType (Maybe Text)

MetricGranularityType

data MetricGranularityType

The MetricGranularityType data type.

mkMetricGranularityType :: MetricGranularityType

Smart constructor for the minimum required fields to construct a valid MetricGranularityType data type to populate a request.

mgtGranularity :: Lens' MetricGranularityType (Maybe Text)

The granularity of a Metric.

ProcessType

data ProcessType

There are two primary Auto Scaling process types--Launch and Terminate. The Launch process creates a new Amazon EC2 instance for an Auto Scaling group, and the Terminate process removes an existing Amazon EC2 instance. The remaining Auto Scaling process types relate to specific Auto Scaling features: AddToLoadBalancer AlarmNotification AZRebalance HealthCheck ReplaceUnhealthy ScheduledActions If you suspend Launch or Terminate, all other process types are affected to varying degrees. The following descriptions discuss how each process type is affected by a suspension of Launch or Terminate. The AddToLoadBalancer process type adds instances to the load balancer when the instances are launched. If you suspend this process, Auto Scaling will launch the instances but will not add them to the load balancer. If you resume the AddToLoadBalancer process, Auto Scaling will also resume adding new instances to the load balancer when they are launched. However, Auto Scaling will not add running instances that were launched while the process was suspended; those instances must be added manually using the the RegisterInstancesWithLoadBalancer call in the Elastic Load Balancing API Reference. The AlarmNotification process type accepts notifications from Amazon CloudWatch alarms that are associated with the Auto Scaling group. If you suspend the AlarmNotification process type, Auto Scaling will not automatically execute scaling policies that would be triggered by alarms. Although the AlarmNotification process type is not directly affected by a suspension of Launch or Terminate, alarm notifications are often used to signal that a change in the size of the Auto Scaling group is warranted. If you suspend Launch or Terminate, Auto Scaling might not be able to implement the alarm's associated policy. The AZRebalance process type seeks to maintain a balanced number of instances across Availability Zones within a Region. If you remove an Availability Zone from your Auto Scaling group or an Availability Zone otherwise becomes unhealthy or unavailable, Auto Scaling launches new instances in an unaffected Availability Zone before terminating the unhealthy or unavailable instances. When the unhealthy Availability Zone returns to a healthy state, Auto Scaling automatically redistributes the application instances evenly across all of the designated Availability Zones. If you call SuspendProcesses on the launch process type, the AZRebalance process will neither launch new instances nor terminate existing instances. This is because the AZRebalance process terminates existing instances only after launching the replacement instances. If you call SuspendProcesses on the terminate process type, the AZRebalance process can cause your Auto Scaling group to grow up to ten percent larger than the maximum size. This is because Auto Scaling allows groups to temporarily grow larger than the maximum size during rebalancing activities. If Auto Scaling cannot terminate instances, your Auto Scaling group could remain up to ten percent larger than the maximum size until you resume the terminate process type. The HealthCheck process type checks the health of the instances. Auto Scaling marks an instance as unhealthy if Amazon EC2 or Elastic Load Balancing informs Auto Scaling that the instance is unhealthy. The HealthCheck process can override the health status of an instance that you set with SetInstanceHealth. The ReplaceUnhealthy process type terminates instances that are marked as unhealthy and subsequently creates new instances to replace them. This process calls both of the primary process types--first Terminate and then Launch. The HealthCheck process type works in conjunction with the ReplaceUnhealthly process type to provide health check functionality. If you suspend either Launch or Terminate, the ReplaceUnhealthy process type will not function properly. The ScheduledActions process type performs scheduled actions that you create with PutScheduledUpdateGroupAction. Scheduled actions often involve launching new instances or terminating existing instances. If you suspend either Launch or Terminate, your scheduled actions might not function as expected.

mkProcessType

Arguments

:: Text
ptProcessName

-> ProcessType

Smart constructor for the minimum required fields to construct a valid ProcessType data type to populate a request.

ptProcessName :: Lens' ProcessType Text

The name of a process.

Activity

data Activity

A scaling Activity is a long-running process that represents a change to your AutoScalingGroup, such as changing the size of the group. It can also be a process to replace an instance, or a process to perform any other long-running operations supported by the API.

mkActivity

Arguments

:: Text
arActivityId

-> Text
arAutoScalingGroupName

-> Text
arCause

-> ISO8601
arStartTime

-> ScalingActivityStatusCode
arStatusCode

-> Activity

Smart constructor for the minimum required fields to construct a valid Activity data type to populate a request.

arActivityId :: Lens' Activity Text

Specifies the ID of the activity.

arAutoScalingGroupName :: Lens' Activity Text

The name of the Auto Scaling group.

arDescription :: Lens' Activity (Maybe Text)

Contains a friendly, more verbose description of the scaling activity.

arCause :: Lens' Activity Text

Contains the reason the activity was begun.

arStartTime :: Lens' Activity ISO8601

Provides the start time of this activity.

arEndTime :: Lens' Activity (Maybe ISO8601)

Provides the end time of this activity.

arStatusCode :: Lens' Activity ScalingActivityStatusCode

Contains the current status of the activity.

arStatusMessage :: Lens' Activity (Maybe Text)

Contains a friendly, more verbose description of the activity status.

arProgress :: Lens' Activity (Maybe Integer)

Specifies a value between 0 and 100 that indicates the progress of the activity.

arDetails :: Lens' Activity (Maybe Text)

Contains details of the scaling activity.

Alarm

data Alarm

The Alarm data type.

mkAlarm :: Alarm

Smart constructor for the minimum required fields to construct a valid Alarm data type to populate a request.

aAlarmName :: Lens' Alarm (Maybe Text)

The name of the alarm.

aAlarmARN :: Lens' Alarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

AutoScalingGroup

data AutoScalingGroup

The AutoScalingGroup data type.

mkAutoScalingGroup

Arguments

:: Text
asgAutoScalingGroupName

-> Text
asgHealthCheckType

-> ISO8601
asgCreatedTime

-> Text
asgLaunchConfigurationName

-> Integer
asgMinSize

-> Integer
asgMaxSize

-> Integer
asgDesiredCapacity

-> Integer
asgDefaultCooldown

-> [Text]
asgAvailabilityZones

-> AutoScalingGroup

Smart constructor for the minimum required fields to construct a valid AutoScalingGroup data type to populate a request.

asgAutoScalingGroupName :: Lens' AutoScalingGroup Text

Specifies the name of the group.

asgAutoScalingGroupARN :: Lens' AutoScalingGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Auto Scaling group.

asgLaunchConfigurationName :: Lens' AutoScalingGroup Text

Specifies the name of the associated LaunchConfiguration.

asgMinSize :: Lens' AutoScalingGroup Integer

Contains the minimum size of the Auto Scaling group.

asgMaxSize :: Lens' AutoScalingGroup Integer

Contains the maximum size of the Auto Scaling group.

asgDesiredCapacity :: Lens' AutoScalingGroup Integer

Specifies the desired capacity for the Auto Scaling group.

asgDefaultCooldown :: Lens' AutoScalingGroup Integer

The number of seconds after a scaling activity completes before any further scaling activities can start.

asgAvailabilityZones :: Lens' AutoScalingGroup [Text]

Contains a list of Availability Zones for the group.

asgLoadBalancerNames :: Lens' AutoScalingGroup [Text]

A list of load balancers associated with this Auto Scaling group.

asgHealthCheckType :: Lens' AutoScalingGroup Text

The service of interest for the health status check, either EC2 for Amazon EC2 or ELB for Elastic Load Balancing.

asgHealthCheckGracePeriod :: Lens' AutoScalingGroup (Maybe Integer)

The length of time that Auto Scaling waits before checking an instance's health status. The grace period begins when an instance comes into service.

asgInstances :: Lens' AutoScalingGroup [Instance]

Provides a summary list of Amazon EC2 instances.

asgCreatedTime :: Lens' AutoScalingGroup ISO8601

Specifies the date and time the Auto Scaling group was created.

asgSuspendedProcesses :: Lens' AutoScalingGroup [SuspendedProcess]

Suspended processes associated with this Auto Scaling group.

asgPlacementGroup :: Lens' AutoScalingGroup (Maybe Text)

The name of the cluster placement group, if applicable. For more information, go to Using Cluster Instances in the Amazon EC2 User Guide.

asgVPCZoneIdentifier :: Lens' AutoScalingGroup (Maybe Text)

The subnet identifier for the Amazon VPC connection, if applicable. You can specify several subnets in a comma-separated list. When you specify VPCZoneIdentifier with AvailabilityZones, ensure that the subnets' Availability Zones match the values you specify for AvailabilityZones.

asgEnabledMetrics :: Lens' AutoScalingGroup [EnabledMetric]

A list of metrics enabled for this Auto Scaling group.

asgStatus :: Lens' AutoScalingGroup (Maybe Text)

The current state of the Auto Scaling group when a DeleteAutoScalingGroup action is in progress.

asgTags :: Lens' AutoScalingGroup [TagDescription]

A list of tags for the Auto Scaling group.

asgTerminationPolicies :: Lens' AutoScalingGroup [Text]

A standalone termination policy or a list of termination policies for this Auto Scaling group.

AutoScalingInstanceDetails

data AutoScalingInstanceDetails

The AutoScalingInstanceDetails data type.

mkAutoScalingInstanceDetails

Arguments

:: Text
asidInstanceId

-> Text
asidAutoScalingGroupName

-> Text
asidAvailabilityZone

-> Text
asidLifecycleState

-> Text
asidHealthStatus

-> Text
asidLaunchConfigurationName

-> AutoScalingInstanceDetails

Smart constructor for the minimum required fields to construct a valid AutoScalingInstanceDetails data type to populate a request.

asidInstanceId :: Lens' AutoScalingInstanceDetails Text

The instance ID of the Amazon EC2 instance.

asidAutoScalingGroupName :: Lens' AutoScalingInstanceDetails Text

The name of the Auto Scaling group associated with this instance.

asidAvailabilityZone :: Lens' AutoScalingInstanceDetails Text

The Availability Zone in which this instance resides.

asidLifecycleState :: Lens' AutoScalingInstanceDetails Text

The life cycle state of this instance. for more information, see Instance Lifecycle State in the Auto Scaling Developer Guide.

asidHealthStatus :: Lens' AutoScalingInstanceDetails Text

The health status of this instance. Healthy means that the instance is healthy and should remain in service. Unhealthy means that the instance is unhealthy. Auto Scaling should terminate and replace it.

asidLaunchConfigurationName :: Lens' AutoScalingInstanceDetails Text

The launch configuration associated with this instance.

BlockDeviceMapping

data BlockDeviceMapping

The BlockDeviceMapping data type.

mkBlockDeviceMapping

Arguments

:: Text
bdmDeviceName

-> BlockDeviceMapping

Smart constructor for the minimum required fields to construct a valid BlockDeviceMapping data type to populate a request.

bdmVirtualName :: Lens' BlockDeviceMapping (Maybe Text)

The virtual name associated with the device.

bdmDeviceName :: Lens' BlockDeviceMapping Text

The name of the device within Amazon EC2 (for example, devsdh or xvdh).

bdmEbs :: Lens' BlockDeviceMapping (Maybe Ebs)

The Elastic Block Storage volume information.

bdmNoDevice :: Lens' BlockDeviceMapping (Maybe Bool)

Suppresses the device mapping. If NoDevice is set to true for the root device, the instance might fail the EC2 health check. Auto Scaling launches a replacement instance if the instance fails the health check.

Ebs

data Ebs

The Elastic Block Storage volume information.

mkEbs :: Ebs

Smart constructor for the minimum required fields to construct a valid Ebs data type to populate a request.

eSnapshotId :: Lens' Ebs (Maybe Text)

The snapshot ID.

eVolumeSize :: Lens' Ebs (Maybe Integer)

The volume size, in gigabytes. Valid values: If the volume type is io1, the minimum size of the volume is 10. Default: If you're creating the volume from a snapshot, and you don't specify a volume size, the default is the snapshot size. Required: Required when the volume type is io1.

eVolumeType :: Lens' Ebs (Maybe Text)

The volume type. Valid values: standard | io1 Default: standard.

eDeleteOnTermination :: Lens' Ebs (Maybe Bool)

Indicates whether to delete the volume on instance termination. Default: true.

eIops :: Lens' Ebs (Maybe Integer)

The number of I/O operations per second (IOPS) that the volume supports. The maximum ratio of IOPS to volume size is 30.0 Valid Values: Range is 100 to 4000. Default: None.

EnabledMetric

data EnabledMetric

The EnabledMetric data type.

mkEnabledMetric :: EnabledMetric

Smart constructor for the minimum required fields to construct a valid EnabledMetric data type to populate a request.

emMetric :: Lens' EnabledMetric (Maybe Text)

The name of the enabled metric.

emGranularity :: Lens' EnabledMetric (Maybe Text)

The granularity of the enabled metric.

Filter

data Filter

The Filter data type.

mkFilter

Arguments

:: Text
fName

-> Filter

Smart constructor for the minimum required fields to construct a valid Filter data type to populate a request.

fName :: Lens' Filter Text

The name of the filter. Valid Name values are: "auto-scaling-group", "key", "value", and "propagate-at-launch".

fValues :: Lens' Filter [Text]

The value of the filter.

Instance

data Instance

The Instance data type.

mkInstance

Arguments

:: Text
iInstanceId

-> Text
iAvailabilityZone

-> LifecycleState
iLifecycleState

-> Text
iHealthStatus

-> Text
iLaunchConfigurationName

-> Instance

Smart constructor for the minimum required fields to construct a valid Instance data type to populate a request.

iInstanceId :: Lens' Instance Text

Specifies the ID of the Amazon EC2 instance.

iAvailabilityZone :: Lens' Instance Text

Availability Zones associated with this instance.

iLifecycleState :: Lens' Instance LifecycleState

Contains a description of the current lifecycle state. The Quarantined lifecycle state is currently not used.

iHealthStatus :: Lens' Instance Text

The instance's health status.

iLaunchConfigurationName :: Lens' Instance Text

The launch configuration associated with this instance.

LaunchConfiguration

data LaunchConfiguration

The LaunchConfiguration data type.

mkLaunchConfiguration

Arguments

:: Text
lcLaunchConfigurationName

-> ISO8601
lcCreatedTime

-> Text
lcImageId

-> Text
lcInstanceType

-> LaunchConfiguration

Smart constructor for the minimum required fields to construct a valid LaunchConfiguration data type to populate a request.

lcLaunchConfigurationName :: Lens' LaunchConfiguration Text

Specifies the name of the launch configuration.

lcLaunchConfigurationARN :: Lens' LaunchConfiguration (Maybe Text)

The launch configuration's Amazon Resource Name (ARN).

lcImageId :: Lens' LaunchConfiguration Text

Provides the unique ID of the Amazon Machine Image (AMI) that was assigned during registration.

lcKeyName :: Lens' LaunchConfiguration (Maybe Text)

Provides the name of the Amazon EC2 key pair.

lcSecurityGroups :: Lens' LaunchConfiguration [Text]

A description of the security groups to associate with the Amazon EC2 instances.

lcUserData :: Lens' LaunchConfiguration (Maybe Text)

The user data available to the launched Amazon EC2 instances.

lcInstanceType :: Lens' LaunchConfiguration Text

Specifies the instance type of the Amazon EC2 instance.

lcKernelId :: Lens' LaunchConfiguration (Maybe Text)

Provides the ID of the kernel associated with the Amazon EC2 AMI.

lcRamdiskId :: Lens' LaunchConfiguration (Maybe Text)

Provides ID of the RAM disk associated with the Amazon EC2 AMI.

lcBlockDeviceMappings :: Lens' LaunchConfiguration [BlockDeviceMapping]

Specifies how block devices are exposed to the instance. Each mapping is made up of a virtualName and a deviceName.

lcInstanceMonitoring :: Lens' LaunchConfiguration (Maybe InstanceMonitoring)

Controls whether instances in this group are launched with detailed monitoring or not.

lcSpotPrice :: Lens' LaunchConfiguration (Maybe Text)

Specifies the price to bid when launching Spot Instances.

lcIamInstanceProfile :: Lens' LaunchConfiguration (Maybe Text)

Provides the name or the Amazon Resource Name (ARN) of the instance profile associated with the IAM role for the instance. The instance profile contains the IAM role.

lcCreatedTime :: Lens' LaunchConfiguration ISO8601

Provides the creation date and time for this launch configuration.

lcEbsOptimized :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the instance is optimized for EBS I/O (true) or not (false).

lcAssociatePublicIpAddress :: Lens' LaunchConfiguration (Maybe Bool)

Specifies whether the instance is associated with a public IP address (true) or not (false).

lcPlacementTenancy :: Lens' LaunchConfiguration (Maybe Text)

Specifies the tenancy of the instance. It can be either default or dedicated. An instance with dedicated tenancy runs in an isolated, single-tenant hardware and it can only be launched in a VPC.

LifecycleHook

data LifecycleHook

A lifecycle hook tells Auto Scaling that you want to perform an action when an instance launches or terminates. When you have a lifecycle hook in place, the Auto Scaling group will either: Pause the instance after it launches, but before it is put into service Pause the instance as it terminates, but before it is fully terminated To learn more, see Auto Scaling Pending State and Auto Scaling Terminating State.

mkLifecycleHook :: LifecycleHook

Smart constructor for the minimum required fields to construct a valid LifecycleHook data type to populate a request.

lhLifecycleHookName :: Lens' LifecycleHook (Maybe Text)

The name of the lifecycle action hook.

lhAutoScalingGroupName :: Lens' LifecycleHook (Maybe Text)

The name of the Auto Scaling group to which the lifecycle action belongs.

lhLifecycleTransition :: Lens' LifecycleHook (Maybe Text)

The Amazon EC2 instance state to which you want to attach the lifecycle hook. See DescribeLifecycleHooks for a list of available lifecycle hook types.

lhNotificationTargetARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the notification target that Auto Scaling will use to notify you when an instance is in the transition state for the lifecycle hook. This ARN target can be either an SQS queue or an SNS topic. The notification message sent to the target will include: Lifecycle action token User account ID Name of the Auto Scaling group Lifecycle hook name EC2 instance ID Lifecycle transition Notification metadata.

lhRoleARN :: Lens' LifecycleHook (Maybe Text)

The ARN of the Amazon IAM role that allows the Auto Scaling group to publish to the specified notification target.

lhNotificationMetadata :: Lens' LifecycleHook (Maybe Text)

Contains additional information that you want to include any time Auto Scaling sends a message to the notification target.

lhHeartbeatTimeout :: Lens' LifecycleHook (Maybe Integer)

Defines the amount of time that can elapse before the lifecycle hook times out. When the lifecycle hook times out, Auto Scaling performs the action defined in the DefaultResult parameter. You can prevent the lifecycle hook from timing out by calling RecordLifecycleActionHeartbeat.

lhGlobalTimeout :: Lens' LifecycleHook (Maybe Integer)

The maximum length of time an instance can remain in a Pending:Wait or Terminating:Wait state. Currently, this value is set at 48 hours.

lhDefaultResult :: Lens' LifecycleHook (Maybe Text)

Defines the action the Auto Scaling group should take when the lifecycle hook timeout elapses or if an unexpected failure occurs. The value for this parameter can be either CONTINUE or ABANDON. The default value for this parameter is CONTINUE.

NotificationConfiguration

data NotificationConfiguration

The NotificationConfiguration data type.

mkNotificationConfiguration :: NotificationConfiguration

Smart constructor for the minimum required fields to construct a valid NotificationConfiguration data type to populate a request.

ncAutoScalingGroupName :: Lens' NotificationConfiguration (Maybe Text)

Specifies the Auto Scaling group name.

ncTopicARN :: Lens' NotificationConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon Simple Notification Service (SNS) topic.

ncNotificationType :: Lens' NotificationConfiguration (Maybe Text)

The types of events for an action to start.

ScalingPolicy

data ScalingPolicy

The ScalingPolicy data type.

mkScalingPolicy :: ScalingPolicy

Smart constructor for the minimum required fields to construct a valid ScalingPolicy data type to populate a request.

sprAutoScalingGroupName :: Lens' ScalingPolicy (Maybe Text)

The name of the Auto Scaling group associated with this scaling policy.

sprPolicyName :: Lens' ScalingPolicy (Maybe Text)

The name of the scaling policy.

sprScalingAdjustment :: Lens' ScalingPolicy (Maybe Integer)

The number associated with the specified adjustment type. A positive value adds to the current capacity and a negative value removes from the current capacity.

sprAdjustmentType :: Lens' ScalingPolicy (Maybe Text)

Specifies whether the ScalingAdjustment is an absolute number or a percentage of the current capacity. Valid values are ChangeInCapacity, ExactCapacity, and PercentChangeInCapacity.

sprCooldown :: Lens' ScalingPolicy (Maybe Integer)

The amount of time, in seconds, after a scaling activity completes before any further trigger-related scaling activities can start.

sprPolicyARN :: Lens' ScalingPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the policy.

sprAlarms :: Lens' ScalingPolicy [Alarm]

A list of CloudWatch Alarms related to the policy.

sprMinAdjustmentStep :: Lens' ScalingPolicy (Maybe Integer)

Changes the DesiredCapacity of the Auto Scaling group by at least the specified number of instances.

ScheduledUpdateGroupAction

data ScheduledUpdateGroupAction

This data type stores information about a scheduled update to an Auto Scaling group.

mkScheduledUpdateGroupAction :: ScheduledUpdateGroupAction

Smart constructor for the minimum required fields to construct a valid ScheduledUpdateGroupAction data type to populate a request.

sugaAutoScalingGroupName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of the Auto Scaling group to be updated.

sugaScheduledActionName :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The name of this scheduled action.

sugaScheduledActionARN :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The Amazon Resource Name (ARN) of this scheduled action.

sugaTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

Time is deprecated. The time that the action is scheduled to begin. Time is an alias for StartTime.

sugaStartTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

The time that the action is scheduled to begin. This value can be up to one month in the future. When StartTime and EndTime are specified with Recurrence, they form the boundaries of when the recurring action will start and stop.

sugaEndTime :: Lens' ScheduledUpdateGroupAction (Maybe ISO8601)

The time that the action is scheduled to end. This value can be up to one month in the future.

sugaRecurrence :: Lens' ScheduledUpdateGroupAction (Maybe Text)

The regular schedule that an action occurs.

sugaMinSize :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

The minimum size of the Auto Scaling group.

sugaMaxSize :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

The maximum size of the Auto Scaling group.

sugaDesiredCapacity :: Lens' ScheduledUpdateGroupAction (Maybe Integer)

The number of instances you prefer to maintain in your Auto Scaling group.

SuspendedProcess

data SuspendedProcess

An Auto Scaling process that has been suspended. For more information, see ProcessType.

mkSuspendedProcess :: SuspendedProcess

Smart constructor for the minimum required fields to construct a valid SuspendedProcess data type to populate a request.

spProcessName :: Lens' SuspendedProcess (Maybe Text)

The name of the suspended process.

spSuspensionReason :: Lens' SuspendedProcess (Maybe Text)

The reason that the process was suspended.

Tag

data Tag

The tag applied to an Auto Scaling group.

mkTag

Arguments

:: Text
tResourceId

-> Text
tResourceType

-> Text
tKey

-> Text
tValue

-> Bool
tPropagateAtLaunch

-> Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

tResourceId :: Lens' Tag Text

The name of the Auto Scaling group.

tResourceType :: Lens' Tag Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

tKey :: Lens' Tag Text

The key of the tag.

tValue :: Lens' Tag Text

The value of the tag.

tPropagateAtLaunch :: Lens' Tag Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.

TagDescription

data TagDescription

The tag applied to an Auto Scaling group.

mkTagDescription

Arguments

:: Text
tdResourceId

-> Text
tdResourceType

-> Text
tdKey

-> Text
tdValue

-> Bool
tdPropagateAtLaunch

-> TagDescription

Smart constructor for the minimum required fields to construct a valid TagDescription data type to populate a request.

tdResourceId :: Lens' TagDescription Text

The name of the Auto Scaling group.

tdResourceType :: Lens' TagDescription Text

The kind of resource to which the tag is applied. Currently, Auto Scaling supports the auto-scaling-group resource type.

tdKey :: Lens' TagDescription Text

The key of the tag.

tdValue :: Lens' TagDescription Text

The value of the tag.

tdPropagateAtLaunch :: Lens' TagDescription Bool

Specifies whether the new tag will be applied to instances launched after the tag is created. The same behavior applies to updates: If you change a tag, the changed tag will be applied to all instances launched after you made the change.